% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AbundanceData}
\alias{AbundanceData}
\title{Abundance}
\description{
The Abundance model data holds the abundance of individuals as number per Stratum and Layer.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
