% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{AcousticDensity}
\alias{AcousticDensity}
\title{Calculate number density from NASC in length intervals}
\usage{
AcousticDensity(
  NASCData,
  AssignmentLengthDistributionData,
  AcousticTargetStrength,
  SpeciesLinkTable
)
}
\arguments{
\item{parameterName}{Parameter descrption.}
}
\value{
An object of StoX data type \code{\link{DensityData}}.
}
\description{
This function converts NASC to number density for each species category based on the acoustic target strength as a function of length for each acoustic category.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
