% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AcousticLayer}
\alias{AcousticLayer}
\title{Acoustic Layer}
\description{
The Acoustic Layer process data is a table of the three columns Layer, MinLayerDepth and MaxLayerDepth specifying the depth ranges for the acoustic layers on which the acoustic density should be calculated, where the Layer column gives the names of the layers.
}
\seealso{
This data type is produced by \code{\link{DefineAcousticLayer}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
