% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{addCentroidToProjection}
\alias{addCentroidToProjection}
\title{Add centroid and possibly other fields to an existing proj4String:}
\usage{
addCentroidToProjection(
  proj4String,
  centroid,
  lon_name = "lon",
  lat_name = "lat",
  additionalFields = NULL
)
}
\arguments{
\item{proj4String}{A proj4 string.}

\item{centroid}{Either a vector of longitude and latitude value, or a named list of this information.}

\item{lon_name, lat_name}{Character: The names of the longitude and latitude fields in \code{centroid} if given as a list.}

\item{additionalFields}{Character: Optional extra proj4 fields in a single string.}
}
\description{
Add centroid and possibly other fields to an existing proj4String:
}
