% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{AddHaulDensityToSuperIndividuals}
\alias{AddHaulDensityToSuperIndividuals}
\title{Add haul density to SuperIndividualsData}
\usage{
AddHaulDensityToSuperIndividuals(
  SuperIndividualsData,
  LengthDistributionData,
  SweepWidthMethod = c("Constant", "PreDefined"),
  SweepWidth = double()
)
}
\arguments{
\item{SuperIndividualsData}{The \code{\link{SuperIndividualsData}} data.}

\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{SweepWidthMethod}{The method for calculating the sweep width. Possible options are (1) "Constant", which requires \code{SweepWidth} to be set as the constant sweep width, and (2) "PreDefined", impying that the sweep width is already incorporated in the \code{WeightedNumber} in the \code{MeanLengthDistributionData} using \code{link{GearDependentLengthDistributionCompensation}} or \code{link{LengthDependentLengthDistributionCompensation}}, or in the \code{MeanSpeciesCategoryCatchData} using \code{link{GearDependentSpeciesCategoryCatchCompensation}}.}

\item{SweepWidth}{The constant sweep width in meters.}
}
\description{
This function calculates the swept-area density of each Haul given in the input SuperIndividualsData.
}
\details{
The density is calculated either by multiplying by a constant sweep width, or by using LengthDistributionData which are sweep width compensated.
}
\seealso{
\code{\link{SuperIndividualsData}} and \code{\link{LengthDistribution}} for generating the input to this function.
}
