% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddResolution.R
\name{AddPSUToPreySpeciesCategoryCatch}
\alias{AddPSUToPreySpeciesCategoryCatch}
\title{Add Survey, Stratum and PSU to PreySpeciesCategoryCatchData}
\usage{
AddPSUToPreySpeciesCategoryCatch(
  PreySpeciesCategoryCatchData,
  PSUDefinition = c("FunctionParameter", "FunctionInput"),
  PSUDefinitionMethod = c("StationToPSU"),
  BioticPSU = NULL,
  SurveyDefinition = c("FunctionParameter", "FunctionInput"),
  SurveyDefinitionMethod = c("AllStrata", "Table"),
  SurveyTable = data.table::data.table(),
  Survey = NULL,
  StratumPolygon = NULL
)
}
\arguments{
\item{PreySpeciesCategoryCatchData}{The \code{\link{PreySpeciesCategoryCatchData}} data.}

\item{PSUDefinition}{The method to use for defining the PSUs, one of \code{FunctionParameter} to define the PSUs on the fly in this function, or \code{FunctionInput} to import PSU process data from a previously run process by \code{BioticPSU}.}

\item{PSUDefinitionMethod}{Character: A string naming the automatic PSU definition method to use. Currently only the option "StationToPSU" is supported, which sets each Station as a PSU. See \code{\link{DefineBioticPSU}} for details.}

\item{BioticPSU}{The \code{\link{BioticPSU}} process data.}

\item{SurveyDefinition}{The method to use for defining the Survey, one of \code{FunctionParameter} to define the Survey on the fly in this function, or \code{FunctionInput} to import Survey process data from a previously run process by the input \code{Survey}.}

\item{SurveyDefinitionMethod}{See \code{\link{DefineSurvey}}}

\item{SurveyTable}{A table of the two columns Stratum and Survey.}

\item{Survey}{The \code{\link{Survey}} process data.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}
}
\description{
This function adds biotic PSU definitions (linked to Stratum) and Survey information to PreySpeciesCategoryCatchData.
}
