% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{aggregateBaselineDataOneTable}
\alias{aggregateBaselineDataOneTable}
\title{Function to aggregate baseline data}
\usage{
aggregateBaselineDataOneTable(
  stoxData,
  TargetVariable,
  aggregationFunction = getReportFunctions(),
  subTable = character(),
  GroupingVariables = character(),
  na.rm = FALSE,
  padWithZerosOn = character(),
  WeightingVariable = character()
)
}
\arguments{
\item{stoxData}{Output from any StoX function.}

\item{TargetVariable}{The variable to report.}

\item{aggregationFunction}{The function to apply, see RstoxBase::getRstoxBaseDefinitions("reportFunctions")$functionName.}

\item{subTable}{The name of the sub table to aggregate on, if \code{stoxData} is a list of tables.}

\item{GroupingVariables}{The variables to report by. For most applications \code{GroupingVariables} should include "Survey" or "SpeciesCategory", unless the user needs to sum over all Survey and SpeciesCategory.}

\item{na.rm}{Used in the function specified by \code{aggregationFunction}.}

\item{padWithZerosOn}{Character vector giving the variables for which missing values should be padded with zeros. This is used particularly for bootstrapping, where a fish length missing in a bootstrap run should be considered as samples with zero individuals, and not missing, so that summary statistics end up taking all bootstrap replicates into account (if not a mean would be overestimated). When padWithZerosOn has positive length, padding with zeros is applied to this variable and to the \code{GroupingVariables}.}

\item{WeightingVariable}{The variable to weight by. Only relevant for \code{ReportFunction} "weighted.mean".}
}
\value{
An aggregated version of the input \code{stoxData}.
}
\description{
Function to aggregate baseline data
}
\seealso{
\code{\link{ReportSuperIndividuals}}
}
