% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{allargs}
\alias{allargs}
\title{Get all arguments from inside a function}
\usage{
allargs(orig_values = FALSE)
}
\arguments{
\item{orig_values}{Logical: If TRUE use the original values (as defined in the function).}
}
\description{
Get all arguments from inside a function
}
