% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{allEqual}
\alias{allEqual}
\title{Test whether all values are equal.}
\usage{
allEqual(x, tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{An R object coercable to numeric.}

\item{tol}{The tolerance of the equality}

\item{...}{Arguments passed on to range and mean, particularly na.rm.}
}
\description{
Test whether all values are equal.
}
