% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{AppendNASC}
\alias{AppendNASC}
\title{Append to NASCData}
\usage{
AppendNASC(NASCData, StoxAcousticData)
}
\arguments{
\item{NASCData}{The \code{\link{NASCData}} data.}

\item{StoxAcousticData}{\code{\link[RstoxData]{StoxAcousticData}} (defined in RstoxData).}
}
\description{
This function converts a StoxAcousticData to NASCData and appends to another NASCData.
}
\seealso{
\code{\link{NASC}} and \code{\link{StoxAcousticData}}.
}
