% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{AssignmentLengthDistribution}
\alias{AssignmentLengthDistribution}
\title{Assignment length distribution per acoustic PSU}
\usage{
AssignmentLengthDistribution(LengthDistributionData, BioticAssignment)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}
}
\value{
An object of StoX datatype \code{\link{AssignmentLengthDistributionData}}.
}
\description{
The function \code{AssignmentLengthDistribution} calculates the average length distribution of the hauls assigned to each AcousticPSU defined by \code{\link{DefineBioticAssignment}}, and weighted by \code{\link{BioticAssignmentWeighting}}. The weights are scaled in bootstrapping by the number of times each haul is resampled. See Details.
}
\details{
The length distribution is normalized to percentage per AcousticPSU by dividing by the summed length distribution and scaling by 100.


Bootstrapping of acoustic-trawl models involve resampling the hauls assigned to the AcousticPSUs. For StoX <= 3.6.2 this resampling (with replacement) was from all hauls assigned to at least one AcousticPSU associated to each stratum (using \code{ResampleBioticAssignmentByStratum}). If different hauls were assigned to different AcousticPSUs this resampling included the risk that none of the assigned hauls of a particular AcousticPSU are resampled in a bootstrap replicate. This will result in all zeros in the weights (column \code{WeightingFactor}) of the BioticAssignment input to the \code{AssignmentLengthDistribution} function. When the length distributions are normalized, this involves dividing 0 by sum of 0, which results in NaN. This is intended, as it reflects that there does not exist a length distribution for that AcousticPSU. This will result in missing Abundance for the required species, but for missing \code{IndividualTotalLength}. To avoid this problem it is adviced to use the resampling function \code{ResampleBioticAssignmentByAcousticPSU} instead of \code{ResampleBioticAssignmentByStratum}, which resamples for each AcousticPSU and will never result in missing assignment length distribution.
}
\seealso{
\code{\link{LengthDistribution}} to produce the input LengthDistributionData, and \code{\link{DefineBioticAssignment}} to produce the input BioticAssignment.
}
