% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AssignmentLengthDistributionData}
\alias{AssignmentLengthDistributionData}
\title{Assignment length distribution data}
\description{
The AssignmentLengthDistributionData model data is a table of averaged length distribution for each combination of Stratum (acoustic)PSU and (acoustic)Layer.
}
\seealso{
This data type is produced by \code{\link{AssignmentLengthDistribution}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
