% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\docType{data}
\name{backwardCompatibility}
\alias{backwardCompatibility}
\title{Backward compabitibility actions:}
\format{
An object of class \code{list} of length 1.
}
\usage{
backwardCompatibility
}
\description{
Backward compabitibility actions:
}
\keyword{datasets}
