% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{backwardCompatibility_RstoxBase}
\alias{backwardCompatibility_RstoxBase}
\title{Backward compabitibility actions:}
\usage{
backwardCompatibility_RstoxBase
}
\description{
Backward compabitibility actions:
}
