% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{BioticAssignmentWeightingTempNew}
\alias{BioticAssignmentWeightingTempNew}
\title{Assignnment of biotic hauls to acoustic PSUs}
\usage{
BioticAssignmentWeightingTempNew(
  BioticAssignment,
  WeightingMethod = c("Equal", "NumberOfLengthSamples", "NASC", "NormalizedTotalWeight",
    "NormalizedTotalCount", "SumWeightedCount", "InverseSumWeightedCount"),
  StoxBioticData,
  LengthDistributionData,
  AcousticTargetStrength,
  MaxNumberOfLengthSamples = 100
)
}
\arguments{
\item{BioticAssignment}{The  \code{\link{BioticAssignment}} data.}

\item{WeightingMethod}{Character: A string naming the method to use (dee details).}

\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data.}

\item{LengthDistributionData}{The length distribution data.}

\item{MaxNumberOfLengthSamples}{Numeric: When \code{WeightingMethod} is "NumberOfLengthSamples", \code{MaxNumberOfLengthSamples} set the opper limit for the weights, so that if a haul has more than \code{MaxNumberOfLengthSamples} number of sampled individuals, the weight is set to \code{MaxNumberOfLengthSamples}.}
}
\value{
An object of StoX data type \code{\link{BioticAssignment}}.
}
\description{
This function defines the \code{\link{BioticAssignment}} process data, linking biotic Hauls with acoustic PSUs.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{BioticAssignment}}.
}
