% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{BootstrapData}
\alias{BootstrapData}
\title{Output from Bootstrap}
\description{
The BootstrapData model data is a list of data from one or more baseline processes stacked for all bootstrap runs.
}
\seealso{
This data type is produced by \code{\link{Bootstrap}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
