% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Channel}
\alias{Channel}
\title{Input vertical resolution for acoustic-trawl models}
\description{
The vertical resolution \code{Channel} is the input resolution for acoustic-trawl models. Each \code{\link{EDSU}} can contain several acoustic channels.
}
\seealso{
\code{\link{Resolution}} for other resolutions.
}
