% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{combined.color}
\alias{combined.color}
\title{Combined rainbow and desaturation colour scale}
\usage{
combined.color(
  n,
  nStart = 0,
  nEnd = floor(n/5),
  colStart = "white",
  colEnd = NA,
  s = c(0.3, 1),
  v = c(1, 0.7),
  start = 0.8,
  end = 0,
  flip = FALSE,
  alpha = 1,
  ...
)
}
\arguments{
\item{n}{The number of colors to generate.}

\item{s}{The range of the saturation values applied to the colors, where lower values suppress saturation. The default applies less strong colors at the higher values of x.}

\item{v}{The range of the value applied to the colors, where lower values causes darker colors. The default applies less darker colors at the lover values of x.}

\item{start}{The start color, given as a value in between 0 and 1.}

\item{end}{The end color, given as a value in between 0 and 1.}

\item{flip}{Logical: If TRUE the color scale should be reversed.}

\item{alpha}{The transparency.}

\item{...}{Not used, but allowing for unused arguments.}
}
\description{
Returns a color vector of rainbow colors which are desaturated by 's' and darkened by 'v' to obtain the "combined" color scale.
}
\examples{
\dontrun{
plot(1:10, 1:10, pch = 19, cex = 10, col = combined.color(10))
plot(1:10, 1:10 + 1, pch = 19, cex = 10, col = combined.color(10, flip=TRUE))
}
}
