% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{cv}
\alias{cv}
\title{The coefficient of variation, i.e., the standard deviation divided by the mean.}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric object}

\item{na.rm}{Logical: If TRUE remove the missing values prior to calculation.}
}
\description{
The coefficient of variation, i.e., the standard deviation divided by the mean.
}
