% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{dataTable2sf_LINESTRING}
\alias{dataTable2sf_LINESTRING}
\title{Convert a data table to individual LINESTRINGs}
\usage{
dataTable2sf_LINESTRING(
  x,
  x1y1x2y2 = c("startLongitude", "startLatitude", "endLongitude", "endLatitude"),
  idCol = NULL,
  crs = NULL
)
}
\arguments{
\item{x}{A a data.table or data.frame.}

\item{x1y1x2y2}{Character: A vector of the names of start x, and end x, start y and end y coordinates.}

\item{idCol}{Character: A vector of the names of columns to use as ID columns in the returned sf object.}

\item{crs}{coordinate reference system: integer with the EPSG code, or character with proj4string}
}
\description{
Convert a data table to individual LINESTRINGs
}
