% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{dataTable2sf_POINT}
\alias{dataTable2sf_POINT}
\title{Convert a data table to individual POINTs}
\usage{
dataTable2sf_POINT(
  x,
  coords = c("Longitude", "Latitude"),
  idCol = NULL,
  crs = NULL
)
}
\arguments{
\item{x}{A a data.table or data.frame.}

\item{coords}{Character: A vector of the names of x and y coordinates.}

\item{idCol}{Character: A vector of the names of columns to use as ID columns in the returned sf object.}

\item{crs}{coordinate reference system: integer with the EPSG code, or character with proj4string}
}
\description{
Convert a data table to individual POINTs
}
