% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DataTypes}
\alias{DataTypes}
\title{StoX data types of the RstoxBase package}
\description{
StoX data types are the data types used to transfer data and information between processes in a StoX estimation model.
}
\details{
This RstoxBase package produces the folliwing StoX data types:
\itemize{
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{StratumPolygon}}}
}
}
\seealso{
\code{\link[RstoxData]{RstoxData}} and \code{\link[RstoxFDA]{RstoxFDA}} for a list of all StoX data types produced by the other official StoX function packages.
}
