% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticLayer}
\alias{DefineAcousticLayer}
\title{Define Acoustic Layer}
\usage{
DefineAcousticLayer(
  processData,
  StoxAcousticData,
  DefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  UseProcessData = FALSE
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "WaterColumn", to define one single for the entire water column; "HighestResolution", to use the maximum possible vertical resolution without intersecting hauls; "Resolution", which can be used to set a fixed layer thickness; and "LayerTable" to provide the \code{LayerTable}.}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\value{
An object of StoX data type \code{\link{AcousticLayer}}.
}
\description{
This function defines the \code{\link{AcousticLayer}} process data, which sets the range intervals of the acoustic layers used in acoustic-trawl estimation models in StoX.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{AcousticPSU}}.
}
