% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticPSUByTime}
\alias{DefineAcousticPSUByTime}
\title{Define Acoustic PSU by time}
\usage{
DefineAcousticPSUByTime(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("FunctionInput"),
  AcousticPSU,
  StoxAcousticData
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use. Currently, only "FunctionInput" is available, implying to take previously defined AcousticPSU definition and the StoxAcousticData used in that definition as input and determine start and stop times of unbroken sequences of EDSUs}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}
}
\value{
An object of StoX data type \code{\link{AcousticPSU}}.
}
\description{
This function defines Acoustic PSUs by start and stop times.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{DefineAcousticPSU}} for defining AcouosticPSU in the first place.
}
