% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineBioticAssignment_temp}
\alias{DefineBioticAssignment_temp}
\title{Assignnment of biotic hauls to acoustic PSUs}
\usage{
DefineBioticAssignment_temp(
  processData,
  AcousticPSU,
  AcousticLayer,
  StoxBioticData,
  DefinitionMethod = c("Stratum", "Radius", "EllipsoidalDistance"),
  StratumPolygon,
  AcousticData,
  Radius = double(),
  MinNumStations = integer(),
  RefGCDistance = double(),
  RefTime = "",
  RefBotDepth = double(),
  RefLatitude = double(),
  RefLongitude = double(),
  UseProcessData = FALSE
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{AcousticPSU}{A list of \code{\link{AcousticPSU}} process data.}

\item{AcousticLayer}{A list of \code{\link{AcousticLayer}} process data.}

\item{StoxBioticData}{A list of \code{\link[RstoxData]{StoxBioticData}} data.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Stratum", which assigns all Hauls to every PSU of each stratum; "Radius", which selects Hauls inside a radius around each PSU????; and "EllipsoidalDistance", which selects Hauls within an ellipsoid defined by axes in space, time, bottom depth *************.}

\item{StratumPolygon}{A list of \code{\link{StratumPolygon}} process data.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\value{
An object of StoX data type \code{\link{AcousticPSU}}.
}
\description{
This function defines the \code{\link{BioticAssignment}} process data, linking biotic Hauls with acoustic PSUs.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{AcousticPSU}}.
}
