% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineBioticPSU}
\alias{DefineBioticPSU}
\title{Biotic PSU}
\usage{
DefineBioticPSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxBioticData,
  DefinitionMethod = c("StationToPSU", "None")
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "StationToPSU", which sets each Station as a PSU, and "None" for pure manual actions by the user.}
}
\value{
An object of StoX data type \code{\link{BioticPSU}}.
}
\description{
This function defines the \code{\link{BioticPSU}} process data, linking strata, biotic PSUs and Stations
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
Acousic PSUs are generated using \code{\link{DefineAcousticPSU}}. For the vertical resolution (Layer) see \code{\link{DefineBioticLayer}} and \code{\link{DefineAcousticLayer}}.
}
