% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineLayer}
\alias{DefineLayer}
\title{Define Layers}
\usage{
DefineLayer(
  processData,
  UseProcessData = FALSE,
  StoxData,
  DefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  LayerType = c("Acoustic", "Biotic")
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StoxData}{Either \code{\link[RstoxData]{StoxBioticData}} or \code{\link[RstoxData]{StoxAcousticData}} data.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "WaterColumn", to define one single for the entire water column; "HighestResolution", to use the maximum possible vertical resolution without intersecting hauls; "Resolution", which can be used to set a fixed layer thickness; and "LayerTable" to provide the \code{LayerTable}.}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}

\item{LayerType}{Character: A string naming the type of model, either "Acoustic" or "Biotic".}
}
\value{
An object of StoX data type \code{\link{BioticLayer}}.
}
\description{
This function defines the \code{\link{BioticLayer}} process data, which sets the range intervals of the swetp-area layers used in biotic estimation models in StoX.
}
\seealso{
\code{\link{DefineAcousticLayer}} and \code{\link{DefineBioticLayer}}.
}
