% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineModel}
\alias{DefineModel}
\title{Define a parametric or numeric model}
\usage{
DefineModel(
  modelClass,
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("Table", "ResourceFile"),
  ModelName,
  ParameterTable = data.table::data.table(),
  FileName = character()
)
}
\arguments{
\item{modelClass}{The model class, such as Regression.}

\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Table" to define a table directly (in the GUI), and ResourceFile to read a file.}

\item{ModelName}{The model to use.}

\item{ParameterTable}{A table holding the parameter values.}

\item{FileName}{A file from which to read the \code{ParameterTable}.}
}
\description{
Define a parametric or numeric model
}
