% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineRegression}
\alias{DefineRegression}
\title{Define a regression model and parameters}
\usage{
DefineRegression(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("ResourceFile", "Table"),
  GroupingVariables = character(),
  RegressionModel = c("SimpleLinear", "Power"),
  RegressionTable = data.table::data.table(),
  FileName = character()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Table" to define a table directly (in the GUI), and ResourceFile to read a file.}

\item{GroupingVariables}{An optional vector of strings defining variables seving as grouping variables in the RegressionTable. Setting this adds the its elements as columns in the RegressionTable in the GUI.}

\item{RegressionModel}{Character: A string naming the model to use for the regression. See Details for options.}

\item{RegressionTable}{A table with one row defining the name of the dependent variable (column name \code{DependentVariable}), the name of the independent variable (column name \code{IndependentVariable}), and the \code{Intersect} and \code{Slope} if \code{RegressionModel} = "SimpleLinear" and \code{Factor} and \code{Exponent} if \code{RegressionModel} = "Power".}

\item{FileName}{The path to a CSV file containing the columns \code{DependentVariable}), \code{IndependentVariable} and the \code{RegressionTable}.}
}
\value{
An object of StoX data type \code{\link{Regression}}.
}
\description{
This function defines a regression model with parameters, where the model can be one of a set of pre-defined models (see the argument \code{RegressionModel}). The parameters can either be defined in a table or read from a resource file.
}
\details{
The currently implemented models are listed below:
\itemize{
\item SimpleLinear \deqn{DependentVariable = Intercept + Slope * IndependentVariable}
\item Power \deqn{DependentVariable = Factor * IndependentVariable^{Exponent}}
}
}
\seealso{
\code{\link{EstimateBioticRegression}} for estimating regression parameters from a \code{\link{StoxBioticData}}, \code{\link{IndividualsData}} or \code{\link{SuperIndividualsData}} object, and  \code{\link{ImputeSuperIndividuals}} for applying the regression to \code{\link{SuperIndividualsData}}.
}
