% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineStation}
\alias{DefineStation}
\title{Some title}
\usage{
DefineStation(BioticData, StationKeys = c("missiontype", "startyear",
  "platform", "missionnumber", "serialnumber"), ...)
}
\arguments{
\item{parameterName}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
