% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{DefineStrata}
\alias{DefineStrata}
\title{Define stratum multipolygon}
\usage{
DefineStrata(FileName, processData = NULL, UseProcessData = FALSE)
}
\arguments{
\item{FileName}{The path to a geoJSON file, shape file (folder) or a Stox multipolygon WKT file. Must include file extension.}

\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object.
}
\description{
This function reads a goejson file, shape file or a \code{\link{StratumPolygon}} Stox multipolygon WKT file and returns an object of StoX data type \code{\link{StratumPolygon}} object.
}
\details{
The parameter \code{UseProcessData} is always set to TRUE when running a process, and needs to be explicitely set to FALSE to enable reading a file (It's set to FALSE at the moment).
}
\examples{

}
\seealso{
\code{\link[RstoxBase]{StratumArea}} for calculating the area of the strata.
}
