% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{DefineStratum}
\alias{DefineStratum}
\title{Define stratum multipolygon}
\usage{
DefineStratum(processData, FileName, UseProcessData = FALSE)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{FileName}{The path to a goejson file, shape file (folder) or a Stox multipolygon WKT file.}

\item{UseProcessData}{Logical: If TRUE use the existing funciton output in the process.}
}
\value{
An object of StoX data type \code{\link{StratumPolygon}}.
}
\description{
This function reads a goejson file, shape file or a \code{\link{StratumPolygon}} Stox multipolygon WKT file and returns an object of StoX data type \code{\link{StratumPolygon}} object.
}
\details{
The parameter \code{UseProcessData} is always set to TRUE when running a process, and needs to be explicitely set to FALSE to enable reading a file.
}
\examples{

}
\seealso{
\code{\link[RstoxBase]{StratumArea}} for calculating the area of the strata.
}
