% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineSurvey}
\alias{DefineSurvey}
\title{Biotic Survey}
\usage{
DefineSurvey(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxData,
  DefinitionMethod = c("Identity", "DeleteAllPSUs"),
  PSUType = c("Acoustic", "Biotic")
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "StationToPSU", which sets each Station as a PSU, and "DeleteAllPSUs" to delete all PSUs.}

\item{PSUType}{Character: A string naming the type of model, either "Acoustic" or "Biotic".}
}
\value{
An object of StoX data type \code{\link{BioticPSU}}.
}
\description{
This function defines the Strata associated to different surveys (in the sense that a separate estimate should be made for those strata).
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
Acousic PSUs are generated using \code{\link{DefineAcousticPSU}}. For the vertical resolution (Layer) see \code{\link{DefineBioticLayer}} and \code{\link{DefineAcousticLayer}}.
}
