% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineSweptAreaLayer}
\alias{DefineSweptAreaLayer}
\title{Define Swept-area Layer}
\usage{
DefineSweptAreaLayer(
  processData,
  UseProcessData = FALSE,
  StoxBioticData,
  DefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "WaterColumn", to define one single for the entire water column; "HighestResolution", to use the maximum possible vertical resolution without intersecting hauls; "Resolution", which can be used to set a fixed layer thickness; and "LayerTable" to provide the \code{LayerTable}.}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}
}
\value{
An object of StoX data type \code{\link{SweptAreaLayer}}.
}
\description{
This function defines the \code{\link{SweptAreaLayer}} process data, which sets the range intervals of the swetp-area layers used in swept-area estimation models in StoX.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
Acoustic Layers are generated using \code{\link{DefineAcousticLayer}}. For the horizontal resolution (Stratum/PSU) see \code{\link{DefineSweptAreaPSU}} and \code{\link{DefineAcousticPSU}}.
}
