% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{DefineTransectParameter}
\alias{DefineTransectParameter}
\title{Define transect design}
\usage{
DefineTransectParameter(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("Parameter", "ParameterTable", "ResourceFile"),
  StratumPolygon,
  StratumNames = character(),
  TransectType = c("Parallel", "ZigZagRectangularEnclosure", "ZigZagEqualSpacing"),
  Bearing = c("Along", "Across", "AlongReversed", "AcrossReversed"),
  BearingAngle = numeric(),
  Retour = FALSE,
  SurveyTime = numeric(),
  SurveyDistance = numeric(),
  SurveySpeed = numeric(),
  Seed = numeric(),
  ParameterTable = data.table::data.table(),
  FileName = character()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of the three options "Parameter", for specifying common parameters for all strata given by \code{StratumNames}; "ParameterTable", for specifying individual parameters for each stratum in the table \code{ParameterTable}; and "ResourceFile", for reading the \code{ParameterTable} from a the file given by \code{FileName}.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StratumNames}{Character: The names of the strata to include in the transect design. Defaults to all strata.}

\item{TransectType}{Character: The name of the type of survey design to create, one of "Parallel", "ZigZagEqualSpacing" and "ZigZagRectangularEnclosure" (see details).}

\item{Bearing}{Character: A string indicating the survey bearing (direction) of each . See Details for options.}

\item{BearingAngle}{Numeric: In the case that \code{Bearing = "Angle"}, \code{BearingAngle} gives the angle of the survey bearing (direction) counter clockwise from north in degrees.}

\item{Retour}{Logical: If TRUE the transect design will be doubled by a retour.}

\item{SurveyTime}{Numeric: The time to spend in the stratum including transit between segments, given in hours. Specifying the \code{SurveyTime} requires the \code{SurveySpeed} to be given as well. Note that the resulting accumulated time may not be exactly equal to \code{SurveyTime}.}

\item{SurveyDistance}{Numeric: The distance to travel in the stratum including transit between segments, given in nautical miles. The \code{SurveyDistance} has precedence over \code{SurveyTime}. Note that the resulting accumulated distance may not be exactly equal to \code{SurveyDistance}.}

\item{SurveySpeed}{Numeric: The speed of the vessel, needed if effort is specified by \code{SurveyTime}.}

\item{Seed}{Numeric: The seed to use when drawing the random starting point.}

\item{ParameterTable}{A table specifying the parameters for each stratum to create transect design for. See the Details for a list of required columns.}

\item{FileName}{The path to a resource file from which to read the TransectParameter process data, in the case that \code{DefinitionMethod} is "ResourceFile".}
}
\value{
An object of StoX data type \code{\link{TransectParameter}}.
}
\description{
Defines transects through parameters that can be used to generate a transect design using the function \code{\link{TransectDesign}}.
}
\details{
If the \code{ParameterTable} is given the table must contain the following columns:
\describe{
\item{"StratumName"}{The name of the stratum, which will be matched against the names of the strata in the \code{StratumPolygon}}
\item{"TransectType"}{See the argument \code{TransectType}}
\item{"Bearing"}{See the argument \code{Bearing}}
\item{"BearingAngle"}{If given, this overrides \code{Bearing}. See the argument \code{BearingAngle}}
\item{"Retour"}{See the argument \code{Retour}}
\item{"SurveyTime"}{See the argument \code{SurveyTime}}
\item{"SurveyDistance"}{Overrides the \code{SurveyTime}. See the argument \code{SurveyDistance}}
\item{"SurveySpeed"}{See the argument \code{SurveySpeed}}
\item{"Seed"}{See the argument \code{Seed}}
}

The following types are implemented through \code{TransectType}:
\describe{
\item{"Parallel"}{Parallel transects}
\item{"ZigZagEqualSpacing"}{Equal space zigzag sampler, Strindberg and Buckland (2004). End transects are generated different from Strindberg and Buckland (2004), by mirroring the last transect around the line perpendicular to the survey direction passing through the last intersection point between the stratum border and the parallel lines used to generate the transects.}
\item{"ZigZagRectangularEnclosure"}{Rectangular enclosure zigzag sampler, Harbitz (2019)}
}
}
\examples{

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon <- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

parameterTable <- DefineTransectParameter(
 DefinitionMethod = "Parameter", 
 TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

print(parameterTable)

}
