% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DensityData}
\alias{DensityData}
\title{Density data}
\description{
The DensityData model data gives the density of individuals as number per square nautical mile, by PSU or Stratum, and by Layer.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
