% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{EDSU}
\alias{EDSU}
\title{Horizontal input resolution for acoustic-trawl models}
\description{
The horizontal resolution \code{EDSU} is the input resolution for acoustic data (convensionally named Log).
}
\seealso{
\code{\link{Resolution}} for other resolutions.
}
