% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{EstimateBioticRegression}
\alias{EstimateBioticRegression}
\title{Estimate a regression model and parameters for biotic data}
\usage{
EstimateBioticRegression(
  InputDataType = c("IndividualsData", "SuperIndividualsData"),
  RegressionModel = c("SimpleLinear", "Power"),
  DependentVariable = character(),
  IndependentVariable = character(),
  GroupingVariables = character(),
  IndividualsData,
  SuperIndividualsData
)
}
\arguments{
\item{InputDataType}{The type of biotic data to estimate the regression parameters based on, one of "StoxBioticData", "IndividualsData" and "SuperIndividualsData".}

\item{RegressionModel}{Character: A string naming the model to use for the regression. See Details for options.}

\item{DependentVariable}{The name of the dependent variable (respons variable).}

\item{IndependentVariable}{The name of the independent variable (explanatory variable).}

\item{GroupingVariables}{An optional vector of strings defining variables seving as grouping variables in the RegressionTable. Setting this adds the its elements as columns in the RegressionTable in the GUI.}

\item{IndividualsData}{The \code{\link{IndividualsData}} data.}

\item{SuperIndividualsData}{The \code{\link{SuperIndividualsData}} data.}
}
\value{
An object of StoX data type \code{\link{Regression}}.
}
\description{
This function estimates a regression model with parameters, where the model can be one of a set of pre-defined models.
}
\details{
The \code{RegressionModel} "Power" performs a log-log transformed simple linear regression of the model L ~ a W^b exp(epsilon), where the error term epsilon is assumed to follow the normal distibution with mean 0 (see \href{http://derekogle.com/fishR/examples/oldFishRVignettes/LengthWeight.pdf}{fishR}).
}
\seealso{
\code{\link{DefineRegression}} for defining regression parameters directly of from a file.
}
