% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{filterTable}
\alias{filterTable}
\title{Filter a table}
\usage{
filterTable(table, filter = character())
}
\arguments{
\item{table}{The data.table to filter.}

\item{filter}{A string with an R expression to filter out unwanted rows of the report, e.g. "IndividualAge \%notin\% NA" or "Survey \%notin\% NA & SpeciesCategory \%notin\% NA".}
}
\description{
Filter a table
}
