% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{findInterval_AfterRound}
\alias{findInterval_AfterRound}
\title{A robust version of findInterval where the inputs are rounded off first to avoid floaring point issues}
\usage{
findInterval_AfterRound(x, vec, digits = 10, ...)
}
\arguments{
\item{x}{numeric.}

\item{vec}{numeric, sorted (weakly) increasingly, of length \code{N},
    say.}

\item{digits}{Integer: The number of digits to round \code{x} and \code{vec} to before applying \code{\link{findInterval}}.}

\item{...}{Passed to \code{\link{findInterval}}.}
}
\description{
A robust version of findInterval where the inputs are rounded off first to avoid floaring point issues
}
