% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{formatProcessData}
\alias{formatProcessData}
\title{Format process data}
\usage{
formatProcessData(processData, columnClasses = NULL)
}
\arguments{
\item{processData}{A StoX process data (list of sf or data.table)}

\item{columnClasses}{A list of lists of classes of the columns of process data tables, as read from process data schemas. Used only in RstoxFramework.}
}
\description{
Format process data
}
