% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{fractionOfSum}
\alias{fractionOfSum}
\title{The fraction of occurrence function}
\description{
The \code{ReportFunction} \code{fractionOfSum} represents the fraction of the sum of the \code{TargetVariable} for two different grouping variables, where the \code{GroupingVariables} is used in the numerator and the \code{GroupingVariables} except the \code{FractionOverVariable} is used in the denominator.
}
\details{
E.g., if \code{GroupingVariables} is c("Survey", "SpeciesCategory", "Stratum"), \code{FractionOverVariable} is "Stratum", \code{TargetVariable} is "CatchFractionWeight", the total catch fraction weight for each stratum is divided by the total catch fraction weight of the survey (for each SpeciesCategory).

Warning! This function is used only to construct an expression to be evaluated in a data.table. Do not use this in R as a regular function like the other \code{\link{ReportFunctions}}.
}
