% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{GearDependentLengthDistributionCompensation}
\alias{GearDependentLengthDistributionCompensation}
\title{Apply the sweep of different gear (and cruise)}
\usage{
GearDependentLengthDistributionCompensation(
  LengthDistributionData,
  CompensationMethod = c("Gear", "Cruise", "GearAndCruise"),
  CompensationTable = data.table::data.table()
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{CompensationMethod}{The method to use for the length dependent catch compensation, i.e. specifying which columns to provide the sweep width for.}

\item{CompensationTable}{A table of the sweep width per combination of the variables specified in \code{CompensationMethod}. Note that all combinations present in the data must be given in the table, as the output should be sweep width compensated for all rows with non-missing WeightedCount.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function multiplies the WeightedCount of a LengthDistributionData by the sweep width given by \code{CompensationTable}. The result is a sweep width compensated length distribution (LengthDistributionType starting with "SweepWidthCompensated").
}
