% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_AcousticPSU_plot_arguments}
\alias{general_AcousticPSU_plot_arguments}
\title{General map plot parameters of RstoxBase}
\arguments{
\item{AcousticPSULabelSize}{Numeric: The size of the AcousticPSU label.}

\item{AcousticPSULabelColor}{Character: The color of the AcousticPSU label.}

\item{AcousticPSULabelPosition}{Character: A string determinig where to place the AcousticPSU label, one of "mean", "atMinLongitude", "atMaxLongitude", "atMinLatitude" or "atMaxLatitude".}

\item{AcousticPSULabelHjust, AcousticPSULabelVjust}{Numeric: The hjust and vjust of the AcousticPSU label for more detailed positioning in addition to the \code{AcousticPSULabelPosition}.}

\item{ShowOnlyAcousticPSU}{Logical: If TRUE plot only EDSUs tagged to acoustic PSUs.}

\item{ShowAcousticPSULabel}{Logical: If TRUE show the labels of the acoustic PSUs in the plot.}
}
\description{
General map plot parameters of RstoxBase
}
