% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_arguments}
\alias{general_arguments}
\title{General parameters of RstoxBase}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{LayerType}{Character: A string naming the type of model, either "Acoustic" or "Biotic".}

\item{PSUType}{Character: A string naming the type of model, either "Acoustic" or "Biotic".}
}
\description{
All functions referring to a project, a model, a process or an output table use the same parameters, listed here.
}
