% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{general_arguments2.7}
\alias{general_arguments2.7}
\title{General parameters of treating project.xml file from StoX <= 2.7}
\arguments{
\item{projectPath}{(Optional) The path to the StoX 2.7 project.}

\item{projectXMLFileName}{(Optional) The name of the StoX 2.7 project.xml file, defaulted to "project.xml".}

\item{projectXMLFilePath}{The path to the project.xml file.}

\item{projectXMLList}{A list as read by \code{\link{readProjectXMLToList}}.}

\item{modelName}{The name of the model (possible values are "baseline", "baseline-report", "r" and "r-report").}

\item{processName}{he name of the process.}

\item{parameterName}{The name of the parameter.}

\item{parameterValue}{The value of the parameter.}

\item{newParameterValue}{The new value of the parameter.}

\item{functionName}{The name of the function of the process to add.}

\item{modify}{A named list specifying the modification of a process. Possible elements are remove, add and modify.}
}
\description{
General parameters of treating project.xml file from StoX <= 2.7
}
