% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_file_plot_arguments}
\alias{general_file_plot_arguments}
\title{General file writing parameters for plots in RstoxBase}
\arguments{
\item{Format}{Character: The format of the output file from plotting functions, defaulted to "png". Other options are defined by \code{\link[ggplot2]{ggsave}}, including "tiff", "jpeg"  and "pdf".}

\item{Width}{Numeric: The width of the plot in centimeters, defaulted to 17, as per the instructions to authors for submission to the ICES Journal og Marine Science.}

\item{Height}{Numeric: The height of the plot in centimeters, defaulted to 17.}

\item{DotsPerInch}{Numeric: The number of dots per inch (DPI), defaulted to 1200.}
}
\description{
General file writing parameters for plots in RstoxBase
}
