% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_map_aspect_plot_arguments}
\alias{general_map_aspect_plot_arguments}
\title{General map plot aspect parameters of RstoxBase}
\arguments{
\item{Zoom}{Numeric: The zoom of the plot, defaulted to 1, which uses the range of the data as the plot limits. Higher values zooms out and lower values zooms in.}

\item{LongitudeMin}{Numeric: The minimum longitude limit of the plot, overriding the value derived from \code{Zoom}.}

\item{LongitudeMax}{Numeric: The maximum longitude limit of the plot, overriding the value derived from \code{Zoom}.}

\item{LatitudeMin}{Numeric: The minimum latitude limit of the plot, overriding the value derived from \code{Zoom}.}

\item{LatitudeMax}{Numeric: The maximum latitude limit of the plot, overriding the value derived from \code{Zoom}.}

\item{LongitudeCenter}{Numeric: The longitude of the point around which the plot is zoomed usinng \code{Zoom}.}

\item{LatitudeCenter}{Numeric: The latitude of the point around which the plot is zoomed usinng \code{Zoom}.}
}
\description{
General map plot aspect parameters of RstoxBase
}
