% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_map_plot_arguments}
\alias{general_map_plot_arguments}
\title{General map plot parameters of RstoxBase (mostly colors)}
\arguments{
\item{ShowMap}{Logical: If TRUE include the map in the plot.}

\item{LandColor}{Character: The color to use for the land, defaulted to the color used for land in the StoX GUI map (yellow color, rgb(253, 254, 204, maxColorValue = 255)).}

\item{BorderColor}{Character: The color to use for the borders on land, defaulted to "grey50".}

\item{OceanColor}{Character: The color to use for the (ocean) background, defaulted to "white", as in the StoX GUI map.}

\item{GridColor}{Character: The color to use for the longitude/latitude grid lines, defaulted to the color used for the longitude/latitude grid in the StoX GUI map (blue color, rgb(223, 242, 255, maxColorValue = 255).}
}
\description{
General map plot parameters of RstoxBase (mostly colors)
}
