% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_report_arguments}
\alias{general_report_arguments}
\title{General report parameters of RstoxBase}
\arguments{
\item{TargetVariable}{The variable to report.}

\item{GroupingVariables}{The variables to report by, e.g. "Stratum" or "SpeciesCategory". Can be given as a comma separated list of variables (with no spaces).}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs). The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA.}
}
\description{
All functions referring to a project, a model, a process or an output table use the same parameters, listed here.
}
