% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_station_plot_arguments}
\alias{general_station_plot_arguments}
\title{General station plotting parameters of RstoxBase}
\arguments{
\item{ShowStations}{Logical: If TRUE include stations (along transects) in the plot.}

\item{StationPointColor}{Character: The color to use for the station points, defaulted to "black".}

\item{StationPointSize}{Numeric: The size of the station points.}

\item{StationPointShape}{The shape of the station points, as defined for the argument \code{pch} in \code{\link{points}}.}
}
\description{
General station plotting parameters of RstoxBase
}
