% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_stratum_plot_arguments}
\alias{general_stratum_plot_arguments}
\title{General stratum plotting parameters of RstoxBase}
\arguments{
\item{ShowStratumPolygon}{Logical: If TRUE include the stratumPolygon. in the plot.}

\item{StratumPolygonColor}{Character: The color palette to use for the strata. The default ("hue") is the default HUE color used by ggplot2. See the \code{Palettes} section in \code{\link[ggplot2]{scale_fill_brewer}} for a list of options (both Diverging, Qualitative and Sequential color palettes are possible). A color blind firendly palette is "viridis".}

\item{StratumPolygonBorderColor}{Character: The single color to use for the polygon borders.}

\item{StratumPolygonBorderLineWidth}{Numeric: The line width of the polygon borders.}
}
\description{
General stratum plotting parameters of RstoxBase
}
