% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_track_plot_arguments}
\alias{general_track_plot_arguments}
\title{General track plotting parameters of RstoxBase}
\arguments{
\item{TrackPointColor}{Character: The colors to use when plotting the data points. The colors scale can be given either as vector of colors comprising equally spaced colors of the color scale, or as the name of a color scale function with the first argument being the number of colors. The default is the \code{\link[RstoxBase]{combined.color}}. Other options for color scale function are "rainbow", "hcl.colors", "heat.colors", "terrain.colors", "topo.colors" or "cm.colors".}

\item{MaxTrackPointSize}{Numeric: The maximum size of the points..}

\item{TrackPointSize}{Numeric: The size of the points.}

\item{TrackPointShape}{The shape of the points, as defined for the argument \code{pch} in \code{\link{points}}.}

\item{TrackColor}{Character: The color to use for the vessel track, defaulted to "black".}

\item{TrackLineWidth}{Numeric: The width of the track, defaulted to 1.}
}
\description{
General track plotting parameters of RstoxBase
}
