% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getCommonIntervals}
\alias{getCommonIntervals}
\title{Get the common intervals of possibly overlapping intervals:}
\usage{
getCommonIntervals(
  data,
  varMin = NULL,
  varMax = NULL,
  lowerName = NULL,
  upperName = NULL
)
}
\arguments{
\item{data}{A data.table/data.frame holding the columns representing lower and upper interval.}

\item{varMin}{The name of the lower interval column. By default interpreted from the first column of \code{data}.}

\item{varMax}{The name of the upper interval column. By default interpreted from the second column of \code{data}.}

\item{lowerName}{The name of the lower interval column in the output table. By default set to \code{varMin}.}

\item{upperName}{The name of the upper interval column in the output table. By default set to \code{varMax}.}
}
\description{
Get the common intervals of possibly overlapping intervals:
}
