% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getDataTypeDefinition}
\alias{getDataTypeDefinition}
\title{Get data type definitions}
\usage{
getDataTypeDefinition(
  dataType,
  subTable = NULL,
  elements = NULL,
  unlist = FALSE
)
}
\arguments{
\item{dataType}{The name of the data type to get definitions from, or a logical function of one input DataType.}

\item{subTable}{The sub table to extract, if any. Default is to extract all potential sub tables.}

\item{elements}{A vector of specific elements to extract from the definition.}

\item{unlist}{Logical: If TRUE unlist the list of column names.}
}
\description{
Get data type definitions
}
