% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getFunctionArgumentHierarchyForSpcificationParameters}
\alias{getFunctionArgumentHierarchyForSpcificationParameters}
\title{Return a list of specification parameters with conditions for use in stoxFunctionAttributes}
\usage{
getFunctionArgumentHierarchyForSpcificationParameters(use, functionName)
}
\arguments{
\item{use}{A string vector of uses. Currently supported are "Baseline" and "Bootstrap".}

\item{functionName}{The name of the report function.}
}
\description{
This function is exported for use in stoxFunctionAttributes in RstoxBase, as this is used onLoad, and is also used in stoxFunctionAttributes in RstoxFramework.
}
