% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getProjectXMLFilePath}
\alias{getProjectXMLFilePath}
\title{Get the path to the project.xml file from StoX <= 2.7.}
\usage{
getProjectXMLFilePath(
  projectPath,
  projectXMLFileName = "project.xml",
  projectXMLFilePath = NULL
)
}
\arguments{
\item{projectPath}{(Optional) The path to the StoX 2.7 project.}

\item{projectXMLFileName}{(Optional) The name of the StoX 2.7 project.xml file, defaulted to "project.xml".}

\item{projectXMLFilePath}{The path to the project.xml file.}
}
\description{
Get the path to the project.xml file from StoX <= 2.7.
}
