% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{getReportFunctionPackage}
\alias{getReportFunctionPackage}
\title{List report function packages}
\usage{
getReportFunctionPackage(x)
}
\arguments{
\item{x}{The name of the report function for which to return package name.}
}
\description{
List report function packages
}
