% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getReportFunctions}
\alias{getReportFunctions}
\title{Function returning the report functions defined for reporting in StoX, such as sum, summaryStox, etc.}
\usage{
getReportFunctions(getMultiple = NULL)
}
\arguments{
\item{getMultiple}{If given as FALSE or TRUE, select only function names used for baseline or bootstrap, respectively.}
}
\description{
Function returning the report functions defined for reporting in StoX, such as sum, summaryStox, etc.
}
