% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getSeedVector}
\alias{getSeedVector}
\title{Generate a seed vector consistently}
\usage{
getSeedVector(seed, size = 1)
}
\arguments{
\item{seed}{A single integer setting the seed of the seed vector generation.}

\item{size}{A single integer giving the length of the seed vector.}
}
\description{
This function should be used for all random sampling in all Rstox packages.
}
