% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getSpecificationFunctionsByUse}
\alias{getSpecificationFunctionsByUse}
\title{Function returning report functions with specification parameter}
\usage{
getSpecificationFunctionsByUse(use = c("Baseline", "Bootstrap"))
}
\arguments{
\item{use}{A string vector of uses. Currently supported are "Baseline" and "Bootstrap".}
}
\description{
Function returning report functions with specification parameter
}
