% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{getSpecificationParameterByFunctionName}
\alias{getSpecificationParameterByFunctionName}
\title{List specification parameters}
\usage{
getSpecificationParameterByFunctionName(x)
}
\arguments{
\item{x}{The name of the report function for which to list parameters.}
}
\description{
List specification parameters
}
