% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getSpecificationParameterDisplayNameByUse}
\alias{getSpecificationParameterDisplayNameByUse}
\title{Function returning the specification parameter display name for report functions with specification parameter}
\usage{
getSpecificationParameterDisplayNameByUse(use = c("Baseline", "Bootstrap"))
}
\arguments{
\item{use}{A string vector of uses. Currently supported are "Baseline" and "Bootstrap".}
}
\description{
Function returning the specification parameter display name for report functions with specification parameter
}
