% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{getStratumNames}
\alias{getStratumNames}
\alias{addStratumNames}
\title{Get or add stratum names from a SpatialPolygonsDataFrame}
\usage{
getStratumNames(
  stratumPolygon,
  StratumNameLabel = c("StratumName", "polygonName"),
  check.unique = TRUE,
  accept.wrong.name.if.only.one = FALSE
)

addStratumNames(
  stratumPolygon,
  StratumNameLabel = c("StratumName", "polygonName"),
  check.unique = TRUE,
  accept.wrong.name.if.only.one = FALSE
)
}
\arguments{
\item{stratumPolygon}{A \code{\link{StratumPolygon}}.}

\item{StratumNameLabel}{The name of the attribute representing the stratum names in the GeoJSON file or shapefile. When reading WKT or txt files the StratumNameLabel is ignored, as those formats does not support column names.}

\item{check.unique}{Logical: If TRUE, an error is given if stratum names are not unique.}

\item{accept.wrong.name.if.only.one}{Logical: If TRUE, interpret stratum names if only one column in the \code{StratumPolygon}.}
}
\description{
Get or add stratum names from a SpatialPolygonsDataFrame
}
