% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{getWeightingParameter}
\alias{getWeightingParameter}
\title{List weighting parameters}
\usage{
getWeightingParameter(x)
}
\arguments{
\item{x}{The name of the report function for which to list parameters.}
}
\description{
List weighting parameters
}
