% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Haul}
\alias{Haul}
\title{Input vertical resolution for swept-area models}
\description{
The vertical resolution \code{Haul} is the input resolution for swept-area models. Each \code{\link{Station}} can contain several hauls.
}
\seealso{
\code{\link{Resolution}} for other resolutions.
}
