% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ImputeSuperIndividualsData}
\alias{ImputeSuperIndividualsData}
\title{Imputed super-individuals data}
\description{
The ImputeSuperIndividualsData model data is the \code{\link{SuperIndividualsData}} imputed to fill inn missing data from individual in the same dataset.
}
\seealso{
This data type is produced by \code{\link{ImputeSuperIndividuals}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
