% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{IndividualsData}
\alias{IndividualsData}
\title{Individuals}
\description{
The IndividualsData model data are the individuals from \code{\link{StoxBioticData}} used in the estimate.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
