% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Layer}
\alias{Layer}
\title{Layer vertical resolution}
\description{
The vertical resolution \code{Layer} is the report resolution of a StoX project along with the horizontal report resolution \code{\link{Stratum}}.
}
\seealso{
\code{\link{Resolution}} for other resolutions.
}
