% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{LengthDependentLengthDistributionCompensation}
\alias{LengthDependentLengthDistributionCompensation}
\title{Catchability of trawls by fish length}
\usage{
LengthDependentLengthDistributionCompensation(
  LengthDistributionData,
  CompensationMethod = c("LengthDependentSweepWidth", "LengthDependentSelectivity"),
  LengthDependentSweepWidthParameters = data.table::data.table(),
  LengthDependentSelectivityParameters = data.table::data.table()
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{CompensationMethod}{The method to use for the length dependent catch compensation, one of "LengthDependentSweepWidth" for adjusting the sweep width according to the fish length dependent herding effect and "LengthDependentSelectivity" for compensating for mash size selectivity (not yet implemented).}

\item{LengthDependentSweepWidthParameters}{A table of parameters of the LengthDependentSweepWidth method, containing the columns SpeciesCategory, LMin, LMax, Alpha and Beta (see details).}

\item{LengthDependentSelectivityParameters}{A table of parameters of the LengthDependentSelectivity method, containing the columns SpeciesCategory,  LMin, LMax, Alpha and Beta (see details).}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function compensates for length dependent herding by the trawl doors into the net, or length dependent selectivity by the mesh size.
}
\details{
This function estimates the effective sweep width and divides the length frequencies (WeightedNumber in the \code{\link{LengthDistributionData}}) by this effective sweep width (in nautical miles). The result is a \code{LengthDistributionData} object with LengthDistributionType starting with "SweepWidthCompensated". In the case that the input \code{LengthDistributionData} is normalized (LengthDistributionType = "Normalized") the output will hare LengthDistributionType = "SweepWidthCompensatedNormalized", which is equivalent to area number density.

The IndividualTotalLength is truncated to the interval \eqn{[LMin, LMin]}, i.e., \code{IndividualTotalLength} < \code{LMin} is set to \code{LMin} and \code{IndividualTotalLength} < \code{LMax} is set to \code{LMax}

When \code{CompensationMethod} = "LengthDependentSweepWidth" the sweep width is modeled as
\deqn{Alpha * IndividualTotalLength^{Beta}}

When \code{CompensationMethod} = "LengthDependentSelectivity" the sweep width is modeled as
\deqn{Alpha * exp(Beta * IndividualTotalLength)}.
}
