% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{MeanDensityData}
\alias{MeanDensityData}
\title{Mean density data}
\description{
The MeanDensityData model data holds the density of individuals as number per square nautical mile, by Stratum and Layer.
}
\seealso{
This data type is produced by \code{\link{MeanDensity}} based on \code{\link{DensityData}} See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
