% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{MeanNASC}
\alias{MeanNASC}
\title{(Weighted) Average NASC}
\usage{
MeanNASC(
  NASCData,
  PSUDefinition = c("PreDefined", "FunctionInput"),
  AcousticPSU = NULL,
  TargetResolution = "PSU"
)
}
\arguments{
\item{NASCData}{The \code{\link{NASC}} data.}

\item{PSUDefinition}{A string naming the method to use for defining the PSUs, one of "PreDefined", if the PSU column is already populated in the \code{NASCData}, or "FunctionInput" to provide the PSUs in the input \code{AcousticPSU}}

\item{AcousticPSU}{\code{\link{AcousticPSU}} data.}

\item{TargetResolution}{The vertical resolution of the output.}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
This function averages NASC data horizontally, weighted by the log distance.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{NASC}} and \code{\link{MeanNASC}}.
}
