% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{MeanNASCData}
\alias{MeanNASCData}
\title{Mean NASC data}
\description{
The MeanNASCData model data is a list of two elements; the Data, which are similar to \code{\link{NASCData}} but added the Layer vertical resolution column and the Stratum and PSU horizontal resolution columns, and summed vertically and averaged horizontally; and the Resolution, which is a table of the five columns Stratum, PSU, EDSU, Layer and Channel containing the link between the horizontal and vertical resolution variables before summing and averaging.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
