% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{modifyProj4String}
\alias{modifyProj4String}
\title{Define a proj4 string}
\usage{
modifyProj4String(x, par, list.out = FALSE)
}
\arguments{
\item{x}{A proj4 string or a list of proj4 parameters to modify.}

\item{par}{A proj4 string or a list of proj4 parameters to modify by.}

\item{list.out}{Logical: If TRUE the projection info is returned as a list instead of a concatenate string.}
}
\description{
Define a proj4 string
}
