% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{NASC}
\alias{NASC}
\title{NASC}
\usage{
NASC(StoxAcousticData)
}
\arguments{
\item{StoxAcousticData}{\code{\link[RstoxData]{StoxAcousticData}} (defined in RstoxData).}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
NASC function converts the StoxAcousticData into NASCData format
}
