% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{NASCData}
\alias{NASCData}
\title{Nautical area scattering coefficient (NASC) data}
\description{
The NASCData model data gives the includes NASC for each acoustic channel or Layer, and ESDU, PSU or Stratum.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
