% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{NASCToStoxAcoustic}
\alias{NASCToStoxAcoustic}
\title{Convert NASCData to StoxAcousticData}
\usage{
NASCToStoxAcoustic(NASCData, StoxAcousticData)
}
\arguments{
\item{NASCData}{The \code{\link{NASCData}} data.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
Convert NASCData to StoxAcousticData
}
\seealso{
\code{\link{NASCData}} and \code{\link{StoxAcousticData}}.
}
