% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{number}
\alias{number}
\title{Number of TRUE}
\usage{
number(x, condition = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{An R object}

\item{condition}{A logical expression (string) which when evaluated to TRUE counts as numeric 1}

\item{na.rm}{All functions used as \code{aggregationFunction} in \code{\link{aggregateBaselineDataOneTable}} must have an na.rm agument. This is however not used in this function.}
}
\description{
Number of TRUE
}
