% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{percentile_5_95}
\alias{percentile_5_95}
\title{The 5 and 95 percentile.}
\usage{
percentile_5_95(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric object}

\item{na.rm}{Logical: If TRUE remove the missing values prior to calculation.}
}
\description{
The 5 and 95 percentile.
}
