% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_lon_lat}
\alias{plot_lon_lat}
\title{Plot a map, an optional (multi)polygon, and points and a track given by the first argument.}
\usage{
plot_lon_lat(
  x,
  polygon,
  showMap = TRUE,
  trackLinesBy = NULL,
  lon_name = "lon",
  lat_name = "lat",
  lon_name_end = NULL,
  lat_name_end = NULL,
  type = c("pl", "lp", "ps", "sp", "p", "s", "l"),
  linetype.track = 1,
  linewidth.track = 1,
  linewidth.polygon.border = 0.5,
  size = 1,
  size.max = 10,
  strokeToPointFactor = 0.1,
  color.track.point = 1,
  color.track = 1,
  color.scale = combined.color(2),
  color.land = "grey50",
  color.border = "grey10",
  color.ocean = "grey90",
  color.grid = "white",
  color.polygon = "hue",
  color.polygon.border = "blue",
  shape = 16,
  alpha.point = 1,
  alpha.track = 1,
  alpha.polygon = 1,
  zoom = 1,
  xlim = NA,
  ylim = NA,
  zoomCenter = c(0.5, 0.5),
  main = NULL,
  text = NULL,
  size.text = 4,
  color.text = "black",
  hjust.text = 0.5,
  vjust.text = 0.5,
  x.text = NULL,
  y.text = NULL,
  ...
)
}
\arguments{
\item{x}{A data.table with data, including the longitude and latitude variables specified by the arguments \code{lon_name} and \code{lat_name}.}

\item{polygon}{An sf object with multipolygons.}

\item{showMap}{Logical: If TRUE show the map.}

\item{trackLinesBy}{Character: A grouping variable for the track lines. Typically this could be Stratum for a plot of a SurveyPlan so that the transport between strata is not shown.}

\item{lon_name, lat_name}{Character: The name of the longitude and latitude variable in the data \code{x}.}

\item{lon_name_end, lat_name_end}{Character: The name of the end point of the longitude and latitude variable in the data \code{x}, used in the case that \code{type} contains "s" (segments).}

\item{type}{The type of plot. See Details.}

\item{linetype.track}{The line type as described in \code{\link[ggplot2]{aes_linetype_size_shape}}.}

\item{linewidth.track}{Numeric: The width of the track lines.}

\item{linewidth.polygon.border}{Numeric: The width of the polygon borders.}

\item{size}{Either a numeric value or the name of the variable to set the size by.}

\item{size.max}{Numeric: The maximum size of the points (presumably ignored if \code{size} is numeric).}

\item{strokeToPointFactor}{Numeric: A value between 0 and 1 setting the width of the point symbols relative to their size.}

\item{color.track.point}{Character: The color of the track points}

\item{color.track}{Character: The color of the track lines.}

\item{color.scale}{Character: The name of the function defining the color scale in case \code{color} names a variable in the data \code{x}.}

\item{color.land, color.border, color.ocean, color.grid}{Character: The colors to use for the land, the borders between countries, the ocean and the grid in the map.}

\item{color.polygon}{The color to use for the polygons, given either as a single color equal for all strata or a color palette. The default, "hue", is the default color ggplot palette. See the \code{Palettes} section in \code{\link[ggplot2]{scale_fill_brewer}} for a list of options (both Diverging, Qualitative and Sequential color palettes are possible).}

\item{color.polygon.border}{Character: The color of the borders between poygons.}

\item{shape}{Numeric: The shape of the points as specified in \code{\link[graphics]{points}}. Could possibly also be a character.}

\item{alpha.point, alpha.track, alpha.polygon}{Numeric: The alpha (transparency) values between 0 and 1 for the points, track lines and polygons.}

\item{zoom}{Numeric: The zoom of the plot, where 1 is no zoom and 2 zooms out to double extent of the plot, etc.}

\item{xlim, ylim}{The optional x and y limit of the plot. These are vectors of length 2. NA can be used to leave on of the values unset. The full extent of the data are used when these are not specified.}

\item{zoomCenter}{Numeric: The center of the plot relative to the limits of the plot before zooming.}

\item{main}{Character: The main title of the plot.}

\item{text}{Character: Text to be added to the plot using \code{\link[ggplot2]{annotate}} at the positions specified by \code{x.text}, \code{y.text}, \code{hjust.text} and \code{vjust.text}.}

\item{size.text}{Character: The size of the text.}

\item{color.text}{Character: The color of the text.}

\item{hjust.text, vjust.text}{Adjustment of the text position. See \code{\link[ggplot2]{annotate}}.}

\item{x.text, y.text}{Numeric: The positions of the text.}

\item{...}{Optional arguments, specifically axis.title.size.x, axis.title.size.y, axis.text.size.x,axis.text.size.y, legend.text.size and legend.title.size.}
}
\description{
Plot a map, an optional (multi)polygon, and points and a track given by the first argument.
}
\details{
The \code{type} can have the following values:
 \describe{
   \item{pl}{Track lines on top of points}
   \item{lp}{Points on top of track lines}
   \item{ps}{Segments on top of points}
   \item{sp}{Points on top of segments}
   \item{p}{Only points}
   \item{s}{Only segments}
   \item{l}{Olny lines}
 }
}
