% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{PlotSumNASC}
\alias{PlotSumNASC}
\title{Plot NASCData}
\usage{
PlotSumNASC(
  SumNASCData,
  UseDefaultPlotSettings = TRUE,
  Zoom = numeric(),
  LongitudeLimits = numeric(),
  LatitudeLimits = numeric(),
  Title = character(),
  AxisTitleSize = numeric(),
  AxisTickSize = numeric(),
  LegendTitleSize = numeric(),
  LegendTextSize = numeric(),
  ColourVariable = character(),
  PointColour = character(),
  PointColourScale = c("combined.color", "rainbow", "hcl.colors", "heat.colors",
    "terrain.colors", "topo.colors", "cm.colors"),
  LandColour = character(),
  BorderColour = character(),
  OceanColour = character(),
  GridColour = character(),
  MaxPointSize = numeric(),
  MinPointSize = numeric(),
  TrackColour = character(),
  TrackSize = numeric(),
  Format = c("png", "tiff", "jpeg", "pdf"),
  Width = numeric(),
  Height = numeric(),
  DotsPerInch = numeric()
)
}
\arguments{
\item{SumNASCData}{The \code{\link{SumNASCData}} data.}

\item{UseDefaultPlotSettings}{Logical: If TRUE (default) use the default settings of the plotting function. Setting this to FALSE will show all options of the plotting function in the StoX GUI.}

\item{Zoom}{The zoom of the plot, defaulted to 1, which uses the range of the data as the plot limits. Higher values zooms out and lower values zooms in.}

\item{LongitudeLimits}{The longitude limits of the map plot, defaulted to the range of the longitude of the data.}

\item{LatitudeLimits}{The latitude limits of the map plot, defaulted to the range of the latitude of the data.}

\item{Title}{The title of the plot, defaulted to no title.}

\item{AxisTitleSize}{The size of the axis titles, defaulted to 20.}

\item{AxisTickSize}{The size of the axis tick marks, defaulted to 20.}

\item{LegendTitleSize}{The size of the legend title, defaulted to 20.}

\item{LegendTextSize}{The size of the legend text, defaulted to 20.}

\item{ColourVariable}{The name of the variable determining the colours of the NASC data points, defaulted to "NASC" (both size AND colour reflecting the NASC values). If the variable is a categorical variable (character or integer), the discrete colours can be set by the \code{PointColour} argument. If the variable is a continuous variable (numeric), the colour scale can be set by the \code{PointColourScale} argument.}

\item{PointColour}{The discrete colours to use when plotting the data points, defaulted to the default ggplot2 colour palette (see the scales package, and specifically the function \code{\link[scales]{hue_pal}} for how to generate these colours).}

\item{PointColourScale}{The continuous colour scale to use when plotting the data points, given as the name of a color scale function with the first argument being the number of colours. The default is the \code{\link[RstoxBase]{combined.color}}. Other options are "rainbow", "hcl.colors", "heat.colors", "terrain.colors", "topo.colors" or "cm.colors".}

\item{LandColour}{The colour to use for the land, defaulted to the colour used for land in the StoX GUI map (yellow colour, rgb(253, 254, 204, maxColorValue = 255)).}

\item{BorderColour}{The colour to use for the borders on land, defaulted to "grey50".}

\item{OceanColour}{The colour to use for the (ocean) background, defaulted to "white", as in the StoX GUI map.}

\item{GridColour}{The colour to use for the longitude/latitude grid lines, defaulted to the colour used for the longitude/latitude grid in the StoX GUI map (blue colour, rgb(223, 242, 255, maxColorValue = 255).}

\item{MaxPointSize}{The maximum size of the points, defaulted to 10.}

\item{MinPointSize}{The minimum size of the points (the size of points of data with value 0), defaulted to 0.5.}

\item{TrackColour}{The colour to use for the vessel track, defaulted to "black".}

\item{TrackSize}{The size of the track points, defaulted to 1.}

\item{Format}{The format of the output file from plotting functions, defaulted to "png". Other options are defined by \code{\link[ggplot2]{ggsave}}, including "tiff", "jpeg"  and "pdf".}

\item{Width}{The width of the plot in centimeters, defaulted to 17, as per the \href{https://academic.oup.com/icesjms/pages/General_Instructions}{ICES Journal og Marine Science author recommendations}.}

\item{Height}{The height of the plot in centimeters, defaulted to 17.}

\item{DotsPerInch}{The number of dots per inch (DPI), defaulted to 1200.}
}
\value{
A \code{\link{PlotSumNASCData}} object.
}
\description{
Plots a map with cruise line and points with size and (optionally) color representing NASC.
}
