% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{PlotSurveyPlan}
\alias{PlotSurveyPlan}
\title{Plot a survey plan.}
\usage{
PlotSurveyPlan(SurveyPlan, StratumPolygon)
}
\arguments{
\item{SurveyPlan}{The \code{\link{SurveyPlan}} process data.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}
}
\value{
An object of StoX data type \code{\link{WriteSurveyPlanData}}.
}
\description{
Plot a survey plan.
}
\examples{

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon<- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
surveyPlanZZ_Along <- DefineSurveyPlan(
	DefinitionMethod = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

WriteSurveyPlan(surveyPlanZZ_Along)

}
