% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreySpeciesCategoryCatch.R
\name{PreySpeciesCategoryCatch}
\alias{PreySpeciesCategoryCatch}
\title{Table of total catch weight and number per Haul and SpeciesCategory}
\usage{
PreySpeciesCategoryCatch(StoxBioticData)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} (defined in RstoxData).}
}
\description{
This function sums the catch fraction weight and number per Haul and SpeciesCategory.
}
