% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{PreySpeciesCategoryCatchData}
\alias{PreySpeciesCategoryCatchData}
\title{Total prey catch per PreySpeciesCategory and Individual}
\description{
The PreySpeciesCategoryCatchData model data is a table with one row per Individual and PreySpeciesCategory holding the TotalPreyCatchWeigth which is summed PreyCatchWeight over all PreySamples of the Individual for each PreySpeciesCategory. All variables of the input StoxBioticData are kept except keys (variables with name ending with "Key").
}
\seealso{
This data type is produced by \code{\link{PreySpeciesCategoryCatch}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
