% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ProcessData}
\alias{ProcessData}
\title{Process data used in estimation models in StoX}
\arguments{
\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{AcousticLayer}{The \code{\link{AcousticLayer}} process data.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}

\item{SweptAreaLayer}{The \code{\link{SweptAreaLayer}} process data.}

\item{SweptAreaPSU}{The \code{\link{SweptAreaPSU}} process data.}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}

\item{AcousticTargetStrength}{The \code{\link{AcousticTargetStrength}} process data.}
}
\description{
The process data of a StoX model are data that are saved to the project description (project.json file), typically manual settings (tagging of \code{\link{EDSU}}s to \code{\link{PSU}}s) or data read from resource files other than acoustic, biotic and landing input data files stored in the input folder.
}
