% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{QuantityData}
\alias{QuantityData}
\title{Quantity data}
\description{
The Quantity model data holds the abundance and biomass of individuals as number and weight (g), respectively, per Stratum and Layer. The data type contains the tables Data and Resolution.
}
\seealso{
This data type is produced by \code{\link{Quantity}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
