% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{readAcousticPSUFrom2.7}
\alias{readAcousticPSUFrom2.7}
\title{Read a the AcousticPSU from a project.xml file from StoX <= 2.7}
\usage{
readAcousticPSUFrom2.7(projectXMLFilePath)
}
\arguments{
\item{projectXMLFilePath}{The path to the project.xml file.}
}
\description{
Read a the AcousticPSU from a project.xml file from StoX <= 2.7
}
