% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReadBiotic}
\alias{ReadBiotic}
\title{Read biotic XML files}
\usage{
ReadBiotic(FileNames)
}
\arguments{
\item{FileNames}{The paths of the biotic files.}
}
\value{
An object of StoX data type BioticData: A list of a list of data.tables of the different levels of the input biotic files.
}
\description{
This function reads multiple biotic file to a list with a list of tables for each file.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[RstoxData]{readXmlFile}}.
}
