% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{readBioticAssignmentFrom2.7}
\alias{readBioticAssignmentFrom2.7}
\title{Read a the BioticAssignment from a project.xml file from StoX <= 2.7}
\usage{
readBioticAssignmentFrom2.7(projectXMLFilePath)
}
\arguments{
\item{projectXMLFilePath}{The path to the project.xml file.}
}
\description{
Read a the BioticAssignment from a project.xml file from StoX <= 2.7
}
