% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{readBioticPSUFrom2.7}
\alias{readBioticPSUFrom2.7}
\title{Read a the AcousticPSU from a project.xml file from StoX <= 2.7}
\usage{
readBioticPSUFrom2.7(projectXMLFilePath, MergedStoxDataHaulLevel)
}
\arguments{
\item{projectXMLFilePath}{The path to the project.xml file.}

\item{MergedStoxDataHaulLevel}{Merged StoxBioticData down to Haul.}
}
\description{
Read a the AcousticPSU from a project.xml file from StoX <= 2.7
}
