% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{readProjectXMLToList}
\alias{readProjectXMLToList}
\title{Read a project.xml file from StoX <= 2.7}
\usage{
readProjectXMLToList(projectXMLFilePath)
}
\arguments{
\item{projectXMLFilePath}{The path to the project.xml file.}
}
\description{
Read a project.xml file from StoX <= 2.7
}
