% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{readStratumPolygonFrom2.7}
\alias{readStratumPolygonFrom2.7}
\alias{writeStratumPolygonFrom2.7}
\title{Read a the StratumPolygon from a project.xml file from StoX <= 2.7}
\usage{
readStratumPolygonFrom2.7(
  projectXMLFilePath,
  remove_includeintotal = TRUE,
  StratumNameLabel = "StratumName"
)

writeStratumPolygonFrom2.7(
  projectXMLFilePath,
  stratumPolygonFilePath,
  remove_includeintotal = TRUE
)
}
\arguments{
\item{projectXMLFilePath}{The path to the project.xml file.}

\item{remove_includeintotal}{Logical: If TRUE, remove the column includeintotal column.}

\item{StratumNameLabel}{The label of the column giving the stratum names.}

\item{stratumPolygonFilePath}{The path to the file to write the StratumPolygon to (geojson file).}
}
\description{
Read a the StratumPolygon from a project.xml file from StoX <= 2.7
}
