% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{reasdGeoJSON}
\alias{reasdGeoJSON}
\title{Read geojson file}
\usage{
reasdGeoJSON(FileName)
}
\arguments{
\item{FileName}{The path to a \href{https://geojson.org/}{GeoJSON} file.}
}
\description{
Read geojson file
}
