% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{RegroupLengthDistribution}
\alias{RegroupLengthDistribution}
\title{Regroup length distribution to common intervals}
\usage{
RegroupLengthDistribution(LengthDistributionData, LengthInterval = numeric())
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{LengthInterval}{Specifies the new length intervals, given as a single numeric value representing the constant length interval widths, (starting from 0).}
}
\value{
An object of StoX data type \code{\link{LengthDistributionData}}.
}
\description{
The RegroupLengthDistribution function is used to set a common length group resolution for one or all SpeciesCategories throughout the output length distribution dataset. The function aggregates the \code{WeightedNumber} of the LengthDistributionData
}
\details{
An intevral vector starting from 0 with increments of \code{Resolution} will be created, and the current length intervals (IndividualTotalLength, IndividualTotalLength + LengthResolution) will be located into the interval vector. Present intervals that intersect with more than one interval in the interval vector result in an error (e.g. IndividualTotalLength = 2.4 with LengthResolution = 0.3 will intersect with two intervals when \code{Resolution} = 0.5, and IndividualTotalLength = 2.4 with LengthResolution = 0.1 will intersect with two intervals when \code{Resolution} = 0.05).
}
