% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{RelativeLengthDistribution}
\alias{RelativeLengthDistribution}
\title{Relative length distribution}
\usage{
RelativeLengthDistribution(LengthDistributionData)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function converts a length distribution to a relative length distribution as percent within each SpeciesCategory for the present horizontal and verticacl resolution.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
