% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{renameSSULabelInPSUProcessData}
\alias{renameSSULabelInPSUProcessData}
\title{Rename SSU to either EDSU or Station}
\usage{
renameSSULabelInPSUProcessData(
  PSUProcessData,
  PSUType = c("Acoustic", "Biotic"),
  reverse = FALSE
)
}
\arguments{
\item{PSUProcessData}{A table to rename in.}

\item{PSUType}{Character: A string naming the type of model, either "Acoustic" or "Biotic"}

\item{reverse}{Logical: If TRUE rename to SSU.}
}
\description{
Rename SSU to either EDSU or Station
}
