% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportAbundanceData}
\alias{ReportAbundanceData}
\title{Reported Abundance data}
\description{
The ReportAbundanceData model data is a report of the \code{\link{AbundanceData}}.
}
\seealso{
This data type is produced by \code{\link{ReportAbundance}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}
}
