% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportBootstrap}
\alias{ReportBootstrap}
\title{Report Bootstrap}
\usage{
ReportBootstrap(
  BootstrapData,
  BaselineProcess,
  TargetVariable,
  AggregationFunction = getRstoxBaseDefinitions("reportFunctions")[multiple == FALSE,
    functionName],
  BootstrapReportFunction = getRstoxBaseDefinitions("reportFunctions")$functionName,
  GroupingVariables = character(),
  RemoveMissingValues = FALSE,
  AggregationWeightingVariable = character(),
  BootstrapReportWeightingVariable = character()
)
}
\arguments{
\item{BaselineProcess}{A vector of character strings naming the baseline processes to report from the boostrap output.}

\item{TargetVariable}{The variable to report.}

\item{AggregationFunction}{The function to apply to each bootstrap run. This must be a function returning a single value.}

\item{BootstrapReportFunction}{The function to apply across bootstrap run, such as "cv" or "stoxSummary".}

\item{GroupingVariables}{The variables to report by, e.g. "Stratum" or "SpeciesCategory". Can be given as a comma separated list of variables (with no spaces).}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs). The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA.}

\item{AggregationWeightingVariable}{The variable to weight by in the \code{AggregationFunction}.}

\item{BootstrapReportWeightingVariable}{The variable to weight by in the \code{BootstrapReportFunction}.}
}
\value{
A \code{\link{ReportBootstrapData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{BootstrapData}}.
}
\details{
This function is useful to, e.g, sum Biomass for each SpeciesCategory and IndividualTotalLenght, or average IndividualTotalLenght for each IndiivdualAge and Stratum.
}
\examples{

}
\seealso{

}
