% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportBootstrapData}
\alias{ReportBootstrapData}
\title{Reported Bootstrap data}
\description{
The ReportBootstrapData model data is a report of the \code{\link{BootstrapData}} such as cv of Biomass per Stratum and SpeciesCategory.
}
\seealso{
This data type is produced by \code{\link{ReportBootstrap}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
