% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportDensityData}
\alias{ReportDensityData}
\title{Reported Density data}
\description{
The ReportDensityData model data is a report of the \code{\link{DensityData}}.
}
\seealso{
This data type is produced by \code{\link{ReportDensity}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}.
}
