% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportFunctions}
\alias{ReportFunctions}
\title{Report functions for baseline process output}
\description{
The following functions are used by StoX to report from a Baseline process output:
 \itemize{
   \item{\code{\link[base]{sum}}}
   \item{\code{\link[base]{mean}}}
   \item{\code{\link[stats]{weighted.mean}}}
   \item{\code{\link[stats]{median}}}
   \item{\code{\link[base]{min}}}
   \item{\code{\link[base]{max}}}
   \item{\code{\link[stats]{sd}}}
   \item{\code{\link[stats]{var}}}
   \item{\code{\link{cv}}}
 }
}
