% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportImputeSuperIndividualsData}
\alias{ReportImputeSuperIndividualsData}
\title{Reported imputed super-individuals data}
\description{
The ReportImputeSuperIndividualsData model data is a report of the \code{\link{ImputeSuperIndividualsData}} such as summed Biomass per Stratum and SpeciesCategory.
}
\seealso{
This data type is produced by \code{\link{ReportImputeSuperIndividuals}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
