% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportPreySpeciesCategoryCatchData}
\alias{ReportPreySpeciesCategoryCatchData}
\title{Reported PreySpeciesCategoryCatch data}
\description{
The ReportPreySpeciesCategoryCatchData model data is a report of the \code{\link{PreySpeciesCategoryCatchData}}.
}
\seealso{
This data type is produced by \code{\link{ReportPreySpeciesCategoryCatch}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
