% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportQuantity}
\alias{ReportQuantity}
\title{Report QuantityData}
\usage{
ReportQuantity(
  QuantityData,
  TargetVariable = c("Abundance", "Biomass"),
  TargetVariableUnit = character(),
  ReportFunction = getReportFunctions(getMultiple = FALSE),
  GroupingVariables = character(),
  InformationVariables = character(),
  RemoveMissingValues = FALSE,
  WeightingVariable = character()
)
}
\arguments{
\item{QuantityData}{The \code{\link{QuantityData}} data.}

\item{TargetVariable}{The variable to report.}

\item{TargetVariableUnit}{The unit to use for the \code{TargetVariable}. For possible units, see subset(RstoxData::StoxUnits, quantity == "cardinality") for TargetVariable = "Abundance" and subset(RstoxData::StoxUnits, quantity == "mass") for TargetVariable = "Biomass".}

\item{ReportFunction}{The function to apply, see RstoxBase::getRstoxBaseDefinitions("reportFunctions")$functionName.}

\item{GroupingVariables}{The variables to report by. For most applications \code{GroupingVariables} should include "Survey" or "SpeciesCategory", unless the user needs to sum over all Survey and SpeciesCategory.}

\item{InformationVariables}{Variables to include as columns to the end of the report table. These cannot have more unique combinations than the \code{GroupingVariables}.}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs) from the \code{TargetVariable}. The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA. Use \code{RemoveMissingValues} = TRUE with extreme caution, as it may lead to under-estimation. E.g., if \code{RemoveMissingValues} = TRUE and a super-individual lacks \code{IndividualRoundWeight}, \code{Biomass} will be NA, and the portion of \code{Abundance} distributed to that super-individual will be excluded when summing \code{Biomass} (but included when summing \code{Abundance}). It is advised to always run with \code{RemoveMissingValues} = FALSE first, and make a thorough investigation to identify the source of any missing values. The function \code{link{ImputeSuperIndividuals}} can be used to impute the missing information from other super-individuals.}

\item{WeightingVariable}{The variable to weight by. Only relevant for \code{ReportFunction} "weighted.mean".}
}
\value{
A \code{\link{ReportQuantityData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{QuantityData}}.
}
