% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportQuantityData}
\alias{ReportQuantityData}
\title{Reported Quantity data}
\description{
The ReportQuantityData model data is a report of the \code{\link{QuantityData}}.
}
\seealso{
This data type is produced by \code{\link{ReportQuantity}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}
}
