% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportSpeciesCategoryCatchData}
\alias{ReportSpeciesCategoryCatchData}
\title{Reported SpeciesCategoryCatch data}
\description{
The ReportSpeciesCategoryCatchData model data is a report of the \code{\link{SpeciesCategoryCatchData}} which combines Haul info and a table of the catch of each SpeciesCategory.
}
\seealso{
This data type is produced by \code{\link{ReportSpeciesCategoryCatch}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
