% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{ReportStationsAlongTransectDesign}
\alias{ReportStationsAlongTransectDesign}
\title{Report summary of stations along transect design}
\usage{
ReportStationsAlongTransectDesign(StationsAlongTransectDesignData)
}
\arguments{
\item{StationsAlongTransectDesignData}{The \code{\link{StationsAlongTransectDesignData}} data.}
}
\value{
An object of StoX data type \code{\link{ReportStationsAlongTransectDesignData}}.
}
\description{
Report summary of stations along transect design
}
\examples{

library(ggplot2)

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon <- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
transectDesignZZ_Along <- TransectDesign(
 TransectParameterDefinition = "FunctionParameter", 
 TransectParameterDefinitionMethod = "Parameter", 
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

stations <- StationsAlongTransectDesign(
  transectDesignZZ_Along, 
  Distance = 30, 
  Seed = 1
)

# Report the TransectDesign:
ReportStationsAlongTransectDesign(stations)

}
