% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportStationsAlongTransectDesignData}
\alias{ReportStationsAlongTransectDesignData}
\title{Report stations along transect design}
\description{
The ReportStationsAlongTransectDesignData model data is a table with the columns Stratum, giving the Stratum name, and NumberOfStationssegments, giving the number of stations in each Stratum, contained in the \code{\link{StationsAlongTransectDesignData}}.
}
\seealso{
This data type is produced by \code{\link{ReportStationsAlongTransectDesign}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
