% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportSuperIndividualsData}
\alias{ReportSuperIndividualsData}
\title{Reported super-individuals data}
\description{
The ReportSuperIndividualsData model data is a report of the \code{\link{SuperIndividualsData}}.
}
\seealso{
This data type is produced by \code{\link{ReportSuperIndividuals}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}.
}
