% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{ReportSurveyPlan}
\alias{ReportSurveyPlan}
\title{Report a survey plan with additional information}
\usage{
ReportSurveyPlan(SurveyPlan, StratumAreaData)
}
\arguments{
\item{SurveyPlan}{The \code{\link{SurveyPlan}} process data.}

\item{StratumAreaData}{The \code{\link{StratumAreaData}} data.}
}
\value{
An object of StoX data type \code{\link{ReportSurveyPlanData}}.
}
\description{
Report a survey plan with additional information
}
\examples{

# Read the stratum system:
stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon<- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)
# Get the area of the strata for use in ReportSurveyPlan():
stratumAreaData <- StratumArea(stratumPolygon)

# Harbitz zigzag survey design along each stratum:
surveyPlanZZ_Along <- DefineSurveyPlan(
	DefinitionMethod = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

# Report the SurveyPlan:
ReportSurveyPlan(surveyPlanZZ_Along, stratumAreaData)

}
