% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportSurveyPlanData}
\alias{ReportSurveyPlanData}
\title{Report survey plan}
\description{
The ReportSurveyPlanData model data is a list of three tables Segment, SegmentWithTransport and Stratum.
}
\details{
The table Segment stores the segments from the input SurveyPlan process data and contains the columns Stratum (character), Segment (character), LongitudeStart (decimal degrees), LatitudeStart (decimal degrees), LongitudeEnd (decimal degrees), LatitudeEnd (decimal degrees), Speed (knots) and Distance (nautical mile), 

The table SegmentWithTransport contains the segments from the table Segment but adds the transport between each segment. This table contains the additional column SegmentType with values "Segment" and "Transport".

The table Stratum stores one row per stratum with the columns Stratum (character), DistanceSegment (nautical mile), DistanceTransport (nautical mile), Distance (nautical mile), Speed (knots), Area (square nautical mile) and Coverage (DistanceSegment / square root of Area).
}
\seealso{
This data type is produced by \code{\link{ReportSurveyPlan}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
