% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{ReportTransectDesign}
\alias{ReportTransectDesign}
\title{Report a transect design with additional information}
\usage{
ReportTransectDesign(TransectDesignData, StratumAreaData)
}
\arguments{
\item{TransectDesignData}{The \code{\link{TransectDesignData}} data.}

\item{StratumAreaData}{The \code{\link{StratumAreaData}} data.}
}
\value{
An object of StoX data type \code{\link{ReportTransectDesignData}}.
}
\description{
Report a transect design with additional information
}
\examples{

# Read the stratum system:
stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon<- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)
# Get the area of the strata for use in ReportTransectDesign():
stratumAreaData <- StratumArea(stratumPolygon)

# Harbitz zigzag survey design along each stratum:
transectDesignZZ_Along <- TransectDesign(
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

# Report the TransectDesign:
ReportTransectDesign(transectDesignZZ_Along, stratumAreaData)

}
