% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportTransectDesignData}
\alias{ReportTransectDesignData}
\title{Report transect design}
\description{
The ReportTransectDesignData model data is a list of three tables Segment, SegmentWithTransit and Stratum.
}
\details{
The table Transect stores the transect segments from the input TransectDesignData and contains the columns Stratum (character), Transect(character), Segment (character), LongitudeStart (decimal degrees), LatitudeStart (decimal degrees), LongitudeEnd (decimal degrees), LatitudeEnd (decimal degrees), Speed (knots), Direction (character: "Tour"/"Retour"), SegmentType (character: "Transect"/"Transit"), Distance (nautical mile), and Duration (hours).

The table Stratum stores one row per stratum and direction with the columns Stratum (character), Direction (character: "Tour"/"Retour"), DistanceSegment (nautical mile), DistanceTransit (nautical mile), Distance (nautical mile), Speed (knots), Area (square nautical mile), Coverage (DistanceSegment / square root of Area) and Duration (hours).
}
\seealso{
This data type is produced by \code{\link{ReportTransectDesign}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
