% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Resolution}
\alias{Resolution}
\title{Resolution}
\description{
Horizontal and vertical resolution in StoX
}
\details{
Stox defines specific horizontal and vertical resolutions which are applied at different stages of a StoX model. The resolutions are grouped into the input resolution, the processing resolution and the report resolution.

The input resolution is \code{\link{Station}}/\code{\link{Haul}} (horizontal/vertical) for swept-area models and \code{\link{EDSU}}/\code{\link{Channel}} for acoustic-trawl models, and is the resolution of the input data.  \cr \cr

The process resolution is \code{\link{PSU}}/\code{\link{Layer}} (horizontal/vertical) for both swept-area and acoustic-trawl models, and is the resolution at which density is calculated. All data used as input when calculating density must be in this resolution.  \cr \cr

The report resolution is \code{\link{Stratum}}/\code{\link{Layer}} (horizontal/vertical) for both swept-area and acoustic-trawl models,, and is the resolution of the abundance estimates.
}
