% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RstoxBase}
\alias{RstoxBase}
\alias{RstoxBase-package}
\title{Title of the package}
\description{
Description of the package
}
\details{
Details of the package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxBase}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxBase/issues}
}

}
\author{
\strong{Maintainer}: Given Name Family Name \email{arnejh@hi.no}

}
