% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setBaseUnit}
\alias{setBaseUnit}
\title{Set base unit to a variable of the StoX dataType}
\usage{
setBaseUnit(x, dataType, variableName)
}
\arguments{
\item{dataType}{The name of a StoX \code{\link{DataTypes}}.}

\item{variableName}{The variable name to get base units from.}
}
\description{
Set base unit to a variable of the StoX dataType
}
