% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setDefaultsInStoxFunction}
\alias{setDefaultsInStoxFunction}
\title{Set default general options}
\usage{
setDefaultsInStoxFunction(x, StoxFunctionName)
}
\arguments{
\item{x}{A list of arguments to the StoxFunction.}

\item{StoxFunctionName}{The name of the StoxFunction.}
}
\description{
Set default general options
}
