% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{setPlotAttributes}
\alias{setPlotAttributes}
\title{Set plot attributes to a plot object}
\usage{
setPlotAttributes(plotObject, plotArguments)
}
\arguments{
\item{plotObject}{The plot object of class ggplot.}

\item{plotArguments}{A list of the arguments to the plot function.}
}
\description{
Sets the following attributes to the plot object (class ggplot): Format, Width, Height, DotsPerInch.
}
