% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{setPrecisionToSF}
\alias{setPrecisionToSF}
\title{Change precision of an sf object}
\usage{
setPrecisionToSF(x, digits)
}
\arguments{
\item{x}{An sf object.}

\item{digits}{The number of digits of the output.}
}
\description{
Change precision of an sf object
}
