% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{SpeciesCategoryCatch}
\alias{SpeciesCategoryCatch}
\title{Table of biotic stations as rows with station info and catch of all species categories as columns}
\usage{
SpeciesCategoryCatch(
  StoxBioticData,
  CatchVariable = c("Count", "WeightKilogram")
)
}
\arguments{
\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{CatchVariable}{Specifies whether to output catch or weight (kilogram).}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function organizes the catch of all species in the input \code{StoxBioticData} into a table with biotic stations as rows, and the catch in count or weight (kilogram)   of each species category in columns.
}
\examples{
x <- 1

}
