% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SpeciesCategoryCatchData}
\alias{SpeciesCategoryCatchData}
\title{Total catch per SpeciesCategory and Haul}
\description{
The SpeciesCategoryCatchData contains the columns Station, Haul, SpeciesCategory, TotalCatchWeigth, TotalCatchNumber, MinHaulDepth, MaxHaulDepth, MinLayerDepth, MaxLayerDepth, LengthDistributionWeight, Cruise, EffectiveTowDistance, VerticalNetOpening, HorizontalNetOpening, TrawlDoorSpread and SpeciesCategoryCatchType.
}
\seealso{
This data type is produced by \code{\link{SpeciesCategoryCatch}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
