% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{SpeciesCategoryDensity}
\alias{SpeciesCategoryDensity}
\title{Table of biotic stations as rows with station info and density of all species categories as columns}
\usage{
SpeciesCategoryDensity(DensityData, StoxBioticData, SweptAreaPSU)
}
\arguments{
\item{DensityData}{An object of \code{\link{DensityData}} data.}

\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data.}

\item{SweptAreaPSU}{The  \code{\link{SweptAreaPSU}} data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function organizes density of all species in the input Density into a table with biotic stations as rows, and the density of each species category in columns. The function requires each swept-area PSUs to contain only one station.
}
\examples{
x <- 1

}
