% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SplitMeanNASC}
\alias{SplitMeanNASC}
\title{Split MeanNASCData to NASCData (deprecated)}
\usage{
SplitMeanNASC(
  MeanNASCData,
  AssignmentLengthDistributionData,
  AcousticTargetStrength,
  SpeciesLink,
  AcousticCategoryLink
)
}
\arguments{
\item{MeanNASCData}{The \code{\link{MeanNASCData}} data.}

\item{AssignmentLengthDistributionData}{The \code{\link{AssignmentLengthDistributionData}} data.}

\item{AcousticTargetStrength}{The \code{\link{AcousticTargetStrength}} process data.}

\item{SpeciesLink}{A table of the two columns AcousticCategory and SpeciesCategory.}

\item{AcousticCategoryLink}{A table linking the acoustic categories to split and those to split into.}
}
\description{
This function splits \code{\link{NASCData}} of specific acoustic categories into other categories based on the acoustic target strength of these categories and the length distribution of corresponding species categories.
}
\seealso{
This functions uses similar methods as \code{\link{AcousticDensity}}. Convert back to \code{\link[RstoxData]{StoxAcousticData}} with \code{\link{NASCToStoxAcoustic}}.
}
