% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SplitNASC}
\alias{SplitNASC}
\title{Split NASCData}
\usage{
SplitNASC(
  NASCData,
  AcousticPSU,
  AssignmentLengthDistributionData,
  AcousticTargetStrength,
  AcousticCategoryLink,
  SpeciesLink
)
}
\arguments{
\item{NASCData}{The \code{\link{NASCData}} data.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}

\item{AssignmentLengthDistributionData}{The \code{\link{AssignmentLengthDistributionData}} data.}

\item{AcousticTargetStrength}{The \code{\link{AcousticTargetStrength}} process data.}

\item{AcousticCategoryLink}{A table linking the acoustic categories to split and those to split into.}

\item{SpeciesLink}{A table of the two columns AcousticCategory and SpeciesCategory.}
}
\description{
This function splits NASCData of specific acoustic categories into other categories based on the acoustic target strength of these categories and the length distribution of corresponding species categories.
}
\seealso{
This functions uses similar methods as \code{\link{AcousticDensity}}. Convert back to \code{\link{StoxAcousticData}} with \code{\link{NASCToStoxAcoustic}}.
}
