% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Station}
\alias{Station}
\title{Horizontal input resolution for swept-area models}
\description{
The horizontal resolution \code{Station} is the input resolution for biotic data. One \code{Station} can contain several \code{\link{Haul}}, which represents the vertical resolution of a \code{Station}.
}
\seealso{
\code{\link{Resolution}} for other resolutions.
}
