% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{StationsAlongTransectDesign}
\alias{StationsAlongTransectDesign}
\title{Generate stations along a transect design}
\usage{
StationsAlongTransectDesign(
  TransectDesignData,
  Distance = numeric(),
  Seed = numeric()
)
}
\arguments{
\item{TransectDesignData}{The \code{\link{TransectDesignData}} data.}

\item{Distance}{Numeric: The distance in nautical miles between the stations along the transect design.}

\item{Seed}{Numeric: The seed to use when drawing the random starting point.}
}
\value{
An object of StoX data type \code{\link{TransectDesignData}}.
}
\description{
Generate stations along a transect design
}
\details{
The \code{Seed} is used to generate one seed per Stratum which are then used to generate the random starting point of each Stratum. From the starting point the function \code{\link[sf]{st_line_sample}} is used to sample points along the transect design in each Stratum separated by \code{Distance} nautical miles (ignoring transit between transects).
}
\examples{
library(ggplot2)

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon <- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
transectDesignZZ_Along <- TransectDesign(
 TransectParameterDefinition = "FunctionParameter", 
 TransectParameterDefinitionMethod = "Parameter", 
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

stations <- StationsAlongTransectDesign(
  transectDesignZZ_Along, 
  Distance = 30, 
  Seed = 1
)

# Plot the stratumPolygon with the segments
ggplot() +
  geom_sf(data = stratumPolygon, aes(fill = StratumName), color = 'blue') +
  geom_segment(
    data = transectDesignZZ_Along, 
    aes(x = LongitudeStart, y = LatitudeStart, xend = LongitudeEnd, yend = LatitudeEnd)
  )


}
