% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{StationsAlongTransectDesignData}
\alias{StationsAlongTransectDesignData}
\title{Stations along transect design}
\description{
The StationsAlongTransectDesignData model data is a table of the columns Stratum (character), Station (character), Longitude (decimal degrees) and Latitude (decimal degrees), where each row gives the position of a station along the transect design along which the stations were positioned and separated by a constant distance.
}
\seealso{
This data type is produced by \code{\link{StationsAlongTransectDesign}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
