% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{StoX_multipolygon_WKT}
\alias{StoX_multipolygon_WKT}
\title{StoX multipolygon WKT}
\description{
This data format is used in \eqn{\le} \href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13250}{StoX 2.7} to define strata polygons.
}
\details{
A data frame used in early versions of \href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13250}{StoX} to define strata polygons, which combine one variable with strata names and one variable consisting of \href{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}{multipolygon.wkt} strings. Well-known text (WKT) is a text markup language for representing vector geometry objects.
}
\references{
Additonal information on WKT multipolygon specification and StoX can be found here:

\href{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}

Johnsen, E., Totland, A., Skålevik, Å., Holmin, A. J., Dingsør, G. E., Fuglebakk, E., & Handegard, N. O. (2019). StoX: An open source software for marine survey analyses. Methods in Ecology and Evolution, 10(9), 1523-1528. \doi{10.1111/2041-210X.13250}
}
