% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{StoxDataStartMiddleStopDateTime}
\alias{StoxDataStartMiddleStopDateTime}
\title{Generate Start, Middle and Stop DateTime variables}
\usage{
StoxDataStartMiddleStopDateTime(
  StoxDataStationLevel,
  type = c("Acoustic", "Biotic")
)
}
\arguments{
\item{StoxDataStationLevel}{Either \code{\link[RstoxData]{StoxAcousticData}} or \code{\link[RstoxData]{StoxBioticData}}, depending on \code{type}.}

\item{type}{A string naming the type of StoX data, one of "Acoustic" and "Biotic".}
}
\value{
An object of StoX data type \code{\link[RstoxData]{MergeStoxAcousticData}}.
}
\description{
Generate Start, Middle and Stop DateTime variables
}
