% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\docType{data}
\name{stoxFunctionAttributes}
\alias{stoxFunctionAttributes}
\title{A list of the attributes of the exported StoX functions:
The format describes the actual content, such as catchabilityTable, filePath, filter, etc. These are used by StoX to choose action on these parameters.
The primitive type (one of integer, double, logical, character) will be interpreted in the process property functions from the type of the function input or parameter.}
\format{
An object of class \code{list} of length 32.
}
\usage{
stoxFunctionAttributes
}
\description{
A list of the attributes of the exported StoX functions:
The format describes the actual content, such as catchabilityTable, filePath, filter, etc. These are used by StoX to choose action on these parameters.
The primitive type (one of integer, double, logical, character) will be interpreted in the process property functions from the type of the function input or parameter.
}
\keyword{datasets}
