% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{StoxModelData}
\alias{StoxModelData}
\title{Model data used in estimation models in StoX}
\description{
The LengthDistributionData model data contains the columns Stratum, PSU, Station, Layer, Haul, SpeciesCategory, IndividualTotalLengthCentimeter, LengthResolutionCentimeter, WeightedCount, MinHaulDepth, MaxHaulDepth, MinLayerDepth, MaxLayerDepth, LengthDistributionWeight, Cruise, EffectiveTowedDistance, VerticalNetOpening, HorizontalNetOpening, TrawlDoorSpread and LengthDistributionType. [List all types here in a nice table]
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
