% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Stratum}
\alias{Stratum}
\title{Stratum horizontal resolution}
\description{
The horizontal resolution \code{Stratum} is the lowest horizontal resolution in StoX.
}
\seealso{
\code{\link{Resolution}} for other resolutions.
}
