% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{StratumAreaData}
\alias{StratumAreaData}
\title{Stratum area}
\description{
The StratumAreaData model data is a table of the area in square nautical miles of each stratum.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
