% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{StratumPolygon}
\alias{StratumPolygon}
\title{StoX data type StratumPolygon}
\description{
The StratumPolygon data type contains the polygons defining the strata of a survey, stored as an sf data.frame with at least two columns: one column named "geometry" giving the multipolygons, and one column naming the multipolygons.
}
\details{
The polygons are stored as gsojson in the project.json file and in output text files.
}
\seealso{
This data type is produced by \code{\link{DefineStratumPolygon}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
