% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{SumLengthDistribution}
\alias{SumLengthDistribution}
\title{Sum length distribution vertically}
\usage{
SumLengthDistribution(
  LengthDistributionData,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution",
    "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  BioticLayer = NULL
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{LayerDefinition}{A string naming the method to use for defining the Layers, one of "FunctionParameter", requiring \code{LayerDefinitionMethod} or \code{PSUDefinitionMethod} and the conditionally \code{Resolution}, \code{LayerTableLayerTable} or \code{StratumPolygon} to be set, or "FunctionInput", requiring the inputs \code{BioticLayer} or \code{BioticPSU}.}

\item{LayerDefinitionMethod}{See \code{DefinitionMethod} of \code{\link{DefineBioticLayer}}.}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}

\item{BioticLayer}{The \code{\link{BioticLayer}} process data.}
}
\value{
An \code{\link{SumLengthDistributionData}} object.
}
\description{
This function summes LengthDistributionData data vertically.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
