% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SumLengthDistributionData}
\alias{SumLengthDistributionData}
\title{Summed length distribution data}
\description{
The SumLengthDistributionData model data is a list of two elements; the Data, which are similar to \code{\link{LengthDistributionData}} but added the Layer resolution column and summed vertically; and the Resolution, which is a table of the three columns Station, Layer and Haul containing the link between the vertical resolution variables before summing.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
