% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SumNASC}
\alias{SumNASC}
\title{Sum NASC}
\usage{
SumNASC(
  NASCData,
  LayerDefinition = c("PreDefined", "FunctionInput"),
  AcousticLayer = NULL,
  TargetResolution = "Layer"
)
}
\arguments{
\item{NASCData}{The \code{\link{NASC}} data.}

\item{LayerDefinition}{A string naming the method to use for defining the Layers, one of "PreDefined", if the Layer column is already populated in the \code{NASCData}, or "FunctionInput" to provide the Layers in the input \code{AcousticLayer}}

\item{AcousticLayer}{\code{\link{AcousticLayer}} data.}

\item{TargetResolution}{The vertical resolution of the output.}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
This function sums NASC data vertically into layer resolution
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{NASC}} and \code{\link{SumNASC}}.
}
