% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SumNASC_old}
\alias{SumNASC_old}
\title{Sum NASC}
\usage{
SumNASC_old(NASCData = NULL, TargetResolution = "Layer", AcousticLayer = NULL)
}
\arguments{
\item{parameterName}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function sums all the vertical channels into layer per edsu
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
