% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCategoryCatch.R
\name{SumSpeciesCategoryCatch}
\alias{SumSpeciesCategoryCatch}
\title{Sum species category catch over Hauls for each Station}
\usage{
SumSpeciesCategoryCatch(
  SpeciesCategoryCatchData,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  BioticLayer = NULL
)
}
\arguments{
\item{SpeciesCategoryCatchData}{The \code{\link{SpeciesCategoryCatchData}} data.}

\item{LayerDefinition}{The method to use for defining the Layers, one of \code{FunctionParameter} to define the Layers on the fly in this function, or \code{FunctionInput} to import Layer process data from a previously run process by \code{BioticLayer}.}

\item{LayerDefinitionMethod}{See \code{\link{DefineBioticLayer}}}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}

\item{BioticLayer}{The \code{\link{BioticLayer}} process data.}
}
\value{
An \code{\link{SumSpeciesCategoryCatchData}} object.
}
\description{
This function summes \code{link{SpeciesCategoryCatchData}} data vertically.
}
