% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{SuperIndividualsErrorInHaulDensity}
\alias{SuperIndividualsErrorInHaulDensity}
\title{Super-indivduals with abundance}
\usage{
SuperIndividualsErrorInHaulDensity(
  IndividualsData,
  AbundanceData,
  DistributionMethod = c("Equal", "HaulDensity"),
  LengthDistributionData
)
}
\arguments{
\item{IndividualsData}{The \code{\link{IndividualsData}} data.}

\item{AbundanceData}{The \code{\link{AbundanceData}} data.}

\item{DistributionMethod}{The method used for distributing the abundance, one of "Equal" for equal abundance to all individuals of each Stratum, Layer, SpeciesCategory and length group, and "HaulDensity" to weight by the haul density. For \code{DistributionMethod} = "HaulDensity" the \code{LengthDistributionData} must be given. It is recommended to use the same \code{LengthDistributionData} that was used to produce the \code{\link{AbundanceData}} (via \code{link{DensityData}}).}

\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}
}
\description{
This function distributes Abundance to the individuals defined by \code{\link{Individuals}}.
}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
