% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SurveyPlan}
\alias{SurveyPlan}
\title{Survey plan}
\description{
The SurveyPlan process data is a table of the columns Stratum (character), Segment (character), LongitudeStart (decimal degrees), LatitudeStart (decimal degrees), LongitudeEnd (decimal degrees), LatitudeEnd (decimal degrees) and Speed (knots), where each row represents one segment of a survey plan (e.g. a transect). The segments do not need to be connected, so there may be transport between the segments. Use \code{\link{ReportSurveyPlan}} to add rows with transport between segments.
}
\seealso{
This data type is produced by \code{\link{DefineSurveyPlan}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
