% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{SweptAreaDensity}
\alias{SweptAreaDensity}
\title{Swept-area density}
\usage{
SweptAreaDensity(
  MeanLengthDistributionData,
  SweepWidthMethod = c("Constant", "PreDefined"),
  SweepWidth = double()
)
}
\arguments{
\item{MeanLengthDistributionData}{The \code{\link{MeanLengthDistributionData}} data.}

\item{SweepWidthMethod}{The method for calculating the sweep width to multiply the \code{WeightedCount} in the \code{LengthDistributionData} with. Possible options are (1) "Constant", which requires \code{SweepWidth} to be set as the constant sweep width, and (2) "PreDefined", impying that the sweep width is already incorporated in the \code{WeightedCount} in the \code{LengthDistributionData}, e.g. using \code{link{GearDependentCatchCompensation}} or \code{link{LengthDependentCatchCompensation}}.}

\item{SweepWidth}{The constant sweep width in meters.}
}
\description{
This function calculates the area density of fish as number of individuals per square nautical mile.
}
\seealso{
See \code{\link{AcousticDensity}} for acoustic density.
}
