% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SweptAreaLayer}
\alias{SweptAreaLayer}
\title{Swept-area Layer}
\description{
The Swept-area Layer process data is a table of the three columns Layer, MinLayerDepth and MaxLayerDepth specifying the depth ranges for the swept-area layers on which the swept-area density should be calculated, where the Layer column gives the names of the layers.

The SweptAreaLayer process data is a table of the three columns Layer, MinLayerDepth and MaxLayerDepth specifying the depth ranges for the swept-area layers on which the swept-area density should be calculated, where the Layer column gives the names of the layers.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}

\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
