% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SweptAreaPSU}
\alias{SweptAreaPSU}
\title{Swept-area PSU}
\description{
The SweptAreaPSU process data contains two tables Stratum_PSU and Station_PSU, where Stratum_PSU has the two column Stratum and PSU giving the swept-area PSUs defined in each stratum (possibly none), and Station_PSU has the two column Station listing every Station in the data and PSU which tags each Station to a PSU. Stations can be un-tagged, which is indicated with PSU = "".
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
