% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{toJSON_Rstox}
\alias{toJSON_Rstox}
\title{Convert to JSON}
\usage{
toJSON_Rstox(x, ...)
}
\arguments{
\item{x}{An object to convert to JSON.}

\item{...}{Parameters overriding the defaults digits = NA, auto_unbox = TRUE, na = "null", null = "null".}
}
\description{
This function takes care of the defaults preferred by the Rstox packages
}
