% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{TransectDesign}
\alias{TransectDesign}
\title{Define transect design}
\usage{
TransectDesign(
  StratumPolygon,
  TransectParameterDefinition = c("FunctionParameter", "FunctionInput"),
  TransectParameterDefinitionMethod = c("Parameter", "ParameterTable"),
  StratumNames = character(),
  TransectType = c("Parallel", "ZigZagRectangularEnclosure", "ZigZagEqualSpacing"),
  Bearing = c("Along", "Across", "AlongReversed", "AcrossReversed"),
  BearingAngle = numeric(),
  Retour = FALSE,
  SurveyTime = numeric(),
  SurveyDistance = numeric(),
  SurveySpeed = numeric(),
  SurveyCoverage = numeric(),
  Seed = numeric(),
  ParameterTable = data.table::data.table(),
  TransectParameter,
  OrderAllToursFirst = FALSE,
  Margin = 0.1
)
}
\arguments{
\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{TransectParameterDefinition}{Character: A string naming the method to use for defining the TransectParameter, one of \code{FunctionParameter} to define the TransectParameter on the fly in this function, or \code{FunctionInput} to import TransectParameter process data generated using the function \code{DefineTransectParameter}.}

\item{TransectParameterDefinitionMethod}{Character: A string naming the TransectParameter definition method to use, one of the two options "Parameter", for specifying common parameter for all strata given by \code{StratumNames}, or "ParameterTable", for specifying individual parameters for each stratum in the table \code{ParameterTable}. See \code{\link{DefineTransectParameter}} for details.}

\item{StratumNames}{Character: The names of the strata to include in the transect design. Defaults to all strata.}

\item{TransectType}{Character: The name of the type of survey design to create, one of "Parallel", "ZigZagEqualSpacing" and "ZigZagRectangularEnclosure" (see details).}

\item{Bearing}{Character: A string indicating the survey bearing (direction) of each . See Details for options.}

\item{BearingAngle}{Numeric: In the case that \code{Bearing = "Angle"}, \code{BearingAngle} gives the angle of the survey bearing (direction) counter clockwise from north in degrees.}

\item{Retour}{Logical: If TRUE the transect design will be doubled by a retour.}

\item{SurveyTime}{Numeric: The time to spend in the stratum including transit between segments, given in hours. Specifying the \code{SurveyTime} requires the \code{SurveySpeed} to be given as well. Note that the resulting accumulated time may not be exactly equal to \code{SurveyTime}.}

\item{SurveyDistance}{Numeric: The distance to travel in the stratum including transit between segments, given in nautical miles. The \code{SurveyDistance} has precedence over \code{SurveyTime}. Note that the resulting accumulated distance may not be exactly equal to \code{SurveyDistance}.}

\item{SurveySpeed}{Numeric: The speed of the vessel, needed if effort is specified by \code{SurveyTime}.}

\item{SurveyCoverage}{Numeric: The survey coverage, which when given and at least one of \code{SurveyTime} and \code{SurveySpeed} is not given will be used to calculate the \code{SurveyDistance} using the definition SurveyCoverage = SurveyDistance / sqrt(area of the stratum).}

\item{Seed}{Numeric: The seed to use when drawing the random starting point.}

\item{ParameterTable}{A table specifying the parameters for each stratum to create transect design for. See the Details for a list of required columns.}

\item{TransectParameter}{The \code{\link{TransectParameter}} process data.}

\item{OrderAllToursFirst}{Logical: If TRUE order all tours first and all retours last, which can be useful for multiple Strata in the same survey direction (e.g. a row of strata along a coast line).}

\item{Margin}{Numeric: The margin to use when iterating to fit the transect design to the desired survey distance (including transit between segments). The function iterates until the survey distance is within \code{Margin} of the desired survey distance, and jumps out of the iteration after 100 tries or if not converging.}
}
\value{
An object of StoX data type \code{\link{TransectDesignData}}.
}
\description{
Define transect design
}
\details{
The \code{TransectDesign} function generates a transect design in a Cartesian coordinate system, and transforms the positions to the geographical coordinate system (longitude, latitude) using the azimuthal equal distance projection, which ensures that distances are preserved in the transformation.

If the \code{ParameterTable} is given the table must contain the following columns:
\describe{
\item{"StratumName"}{The name of the stratum, which will be matched against the names of the strata in the \code{StratumPolygon}}
\item{"TransectType"}{See the argument \code{TransectType}}
\item{"Bearing"}{See the argument \code{Bearing}}
\item{"BearingAngle"}{If given, this overrides \code{Bearing}. See the argument \code{BearingAngle}}
\item{"Retour"}{See the argument \code{Retour}}
\item{"SurveyTime"}{See the argument \code{SurveyTime}}
\item{"SurveyDistance"}{Overrides the \code{SurveyTime}. See the argument \code{SurveyDistance}}
\item{"SurveySpeed"}{See the argument \code{SurveySpeed}}
\item{"SurveyCoverage"}{See the argument \code{SurveyCoverage}}
\item{"Seed"}{See the argument \code{Seed}}
}

The following types are implemented through \code{TransectType}:
\describe{
\item{"Parallel"}{Parallel transects}
\item{"ZigZagEqualSpacing"}{Equal space zigzag sampler, Strindberg and Buckland (2004). End transects are generated different from Strindberg and Buckland (2004), by mirroring the last transect around the line perpendicular to the survey direction passing through the last intersection point between the stratum border and the parallel lines used to generate the transects.}
\item{"ZigZagRectangularEnclosure"}{Rectangular enclosure zigzag sampler, Harbitz (2019)}
}
}
\examples{
library(ggplot2)

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon <- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
transectDesignZZ_Along <- TransectDesign(
 TransectParameterDefinition = "FunctionParameter", 
 TransectParameterDefinitionMethod = "Parameter", 
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

# Plot the stratumPolygon with the segments
ggplot() +
  geom_sf(data = stratumPolygon, aes(fill = StratumName), color = 'blue') +
  geom_segment(
    data = transectDesignZZ_Along, 
    aes(x = LongitudeStart, y = LatitudeStart, xend = LongitudeEnd, yend = LatitudeEnd)
  )

# Harbitz zigzag survey design across each stratum:
transectDesignZZ_Across <- TransectDesign(
 TransectParameterDefinition = "FunctionParameter", 
 TransectParameterDefinitionMethod = "Parameter", 
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Across"
)
# Plot the stratumPolygon with the segments
ggplot() +
  geom_sf(data = stratumPolygon, aes(fill = StratumName), color = 'blue') +
  geom_segment(
    data = transectDesignZZ_Across, 
    aes(x = LongitudeStart, y = LatitudeStart, xend = LongitudeEnd, yend = LatitudeEnd)
  )

# Parallel survey design across each stratum:
transectDesignParallel_Across <- TransectDesign(
 TransectParameterDefinition = "FunctionParameter", 
 TransectParameterDefinitionMethod = "Parameter", 
	TransectType = "Parallel", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)
# Plot the stratumPolygon with the segments
ggplot() +
  geom_sf(data = stratumPolygon, aes(fill = StratumName), color = 'blue') +
  geom_segment(
    data = transectDesignParallel_Across, 
    aes(x = LongitudeStart, y = LatitudeStart, xend = LongitudeEnd, yend = LatitudeEnd)
  )

}
\references{
Strindberg, S., & Buckland, S. T. (2004). Zigzag survey designs in line transect sampling. Journal of Agricultural, Biological, and Environmental Statistics, 9, 443-461.

Harbitz, A. (2019). A zigzag survey design for continuous transect sampling with guaranteed equal coverage probability. Fisheries Research, 213, 151-159.
}
