% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{TransectDesignData}
\alias{TransectDesignData}
\title{Transect design}
\description{
The TransectDesignData model data is a table of the columns Stratum (character), Transect (character), Segment (character), LongitudeStart (decimal degrees), LatitudeStart (decimal degrees), LongitudeEnd (decimal degrees), LatitudeEnd (decimal degrees), Distance (nautical mile), Direction (character) and Speed (knots), where each row represents one segment of a transect design. The segments do not need to be connected, so there may be transit between the segments. Use \code{\link{ReportTransectDesign}} to add rows with transit between segments.
}
\seealso{
This data type is produced by \code{\link{TransectDesign}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
