% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{TransectParameter}
\alias{TransectParameter}
\title{Transect parameter table}
\description{
The TransectParameter process data type contains a table of the columns "StratumName", "TransectType", "Bearing", "BearingAngle", "Retour", "SurveyTime", "SurveyDistance", "SurveySpeed", "Seed", where either "Bearing" or "BearingAngle" must be present and either SurveyTime" and "SurveySpeed" or only "SurveyDistance" must be present.
}
\details{
The polygons are stored as gsojson in the project.json file and in output text files.
}
\seealso{
This data type is produced by \code{\link{DefineTransectParameter}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
