% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{writeProjectXMLToList}
\alias{writeProjectXMLToList}
\title{Write a project.xml file from StoX <= 2.7 from a list}
\usage{
writeProjectXMLToList(
  projectXMLList,
  projectXMLFilePath,
  projectPath = NULL,
  projectXMLFileName = "project.xml"
)
}
\arguments{
\item{projectXMLList}{A list as read by \code{\link{readProjectXMLToList}}.}

\item{projectXMLFilePath}{The path to the project.xml file.}

\item{projectPath}{(Optional) The path to the StoX 2.7 project.}

\item{projectXMLFileName}{(Optional) The name of the StoX 2.7 project.xml file, defaulted to "project.xml".}
}
\description{
Write a project.xml file from StoX <= 2.7 from a list
}
