% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{WriteStationsAlongTransectDesign}
\alias{WriteStationsAlongTransectDesign}
\title{Write a transect design to GPX file(s).}
\usage{
WriteStationsAlongTransectDesign(
  StationsAlongTransectDesignData,
  Format = c("GPX")
)
}
\arguments{
\item{StationsAlongTransectDesignData}{The \code{\link{StationsAlongTransectDesignData}} data.}

\item{Format}{The format of the output files. Currently the only option is "GPX", which prodices files that can be read by the Olex software.}
}
\value{
An object of StoX data type \code{\link{WriteStationsAlongTransectDesignData}}.
}
\description{
Write a transect design to GPX file(s).
}
\examples{

library(ggplot2)

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon <- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
transectDesignZZ_Along <- TransectDesign(
 TransectParameterDefinition = "FunctionParameter", 
 TransectParameterDefinitionMethod = "Parameter", 
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

stations <- StationsAlongTransectDesign(
  transectDesignZZ_Along, 
  Distance = 30, 
  Seed = 1
)

# Convert the transect design to an sf object and write this as a gpx file 
# (this is done automatically by RstoxFramework in StoX):
gpxData <- WriteStationsAlongTransectDesign(stations)

}
