% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{WriteStationsAlongTransectDesignData}
\alias{WriteStationsAlongTransectDesignData}
\title{Stations along transect design as GPX}
\description{
The WriteStationsAlongTransectDesignData model data is an sf object which can be written as GPX waypoints.
}
\seealso{
This data type is produced by \code{\link{WriteStationsAlongTransectDesign}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
