% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{WriteStratumPolygon}
\alias{WriteStratumPolygon}
\title{Write StratumPolygon}
\usage{
WriteStratumPolygon(
  StratumPolygon,
  FileFormat = c("GeoJSON", "StoX_multipolygon_WKT", "StoX_shapefile")
)
}
\arguments{
\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{FileFormat}{The format of the files to write the StratumPolygon to. \code{\link{StoX_multipolygon_WKT}} is the table of stratum name and WKT multipolygon used by StoX.}
}
\value{
A \code{\link{WriteStratumPolygonData}} object.
}
\description{
Writes a StratumPolygon to GeoJSON, StoX_WKT or shapefile.
}
\details{
The actual writing takes place in RstoxFramework. This function only converts the data to appropriate classes interpreted by RstoxFramework.
}
