% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{WriteStratumPolygonData}
\alias{WriteStratumPolygonData}
\title{Write StratumPolygon data}
\description{
The WriteStratumPolygonData model data is used to write \code{\link{StratumPolygon}}.
}
\seealso{
This data type is produced by \code{\link{WriteStratumPolygon}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}.
}
