% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{WriteSurveyPlan}
\alias{WriteSurveyPlan}
\title{Write a survey plan to a GPX file.}
\usage{
WriteSurveyPlan(SurveyPlan)
}
\arguments{
\item{SurveyPlan}{The \code{\link{SurveyPlan}} process data.}
}
\value{
An object of StoX data type \code{\link{WriteSurveyPlanData}}.
}
\description{
Write a survey plan to a GPX file.
}
\examples{

library(sf)

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon<- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
surveyPlanZZ_Along <- DefineSurveyPlan(
	DefinitionMethod = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

# Convert the survey plan to an sf object and write this as a gpx file 
# (this is done automatically by RstoxFramework in StoX):
gpxData <- WriteSurveyPlan(surveyPlanZZ_Along)
filePath <- tempfile(fileext = "gpx")
st_write(
     gpxData,
     dsn = filePath,
     layer = "track_points",
     driver = "GPX"
 )
 
 # Read the data back in:
 gpxData_backin <- st_read(filePath)
 identical(gpxData, gpxData_backin)

}
