% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{WriteSurveyPlanData}
\alias{WriteSurveyPlanData}
\title{Survey plan as GPX}
\description{
The WriteSurveyPlanData model data is an sf object which can be written as GPX tracks (one segment with waypoints.
}
\seealso{
This data type is produced by \code{\link{WriteSurveyPlan}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
