% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPlan.R
\name{WriteTransectDesign}
\alias{WriteTransectDesign}
\title{Write a transect design to GPX file(s).}
\usage{
WriteTransectDesign(
  TransectDesignData,
  Format = c("GPX"),
  TrackGroupingVariables = character(),
  FileGroupingVariables = character()
)
}
\arguments{
\item{TransectDesignData}{The \code{\link{TransectDesignData}} data.}

\item{Format}{The format of the output files. Currently the only option is "GPX", which prodices files that can be read by the Olex software.}

\item{TrackGroupingVariables}{Character: A vector of one or more variables to group the tracks by in the file(s). Possible values are "Stratum", "Direction" and "Transect".}

\item{FileGroupingVariables}{Character: A vector of one or more variables to group the files by. Possible values are "Stratum" and "Direction". One file is written for each combination of the \code{FileGroupingVariables}.}
}
\value{
An object of StoX data type \code{\link{WriteTransectDesignData}}.
}
\description{
Write a transect design to GPX file(s).
}
\examples{

library(sf)

stratumFile <- system.file(
  "testresources", 
  "strata_sandeel_2020_firstCoverage.wkt", package = "RstoxBase"
 )
stratumPolygon<- DefineStratumPolygon(
  DefinitionMethod = "ResourceFile", 
  FileName = stratumFile
)

# Harbitz zigzag survey design along each stratum:
transectDesignZZ_Along <- TransectDesign(
	TransectType = "ZigZagRectangularEnclosure", 
	StratumPolygon = stratumPolygon, 
	SurveyTime = 200, 
	SurveySpeed = 10, 
	Seed = 1, 
	Bearing = "Along"
)

# Convert the transect design to an sf object and write this as a gpx file 
# (this is done automatically by RstoxFramework in StoX):
gpxData <- WriteTransectDesign(transectDesignZZ_Along)
filePath <- tempfile(fileext = ".gpx")
layers <- c("geometry", "track_name", "track_fid", "track_seg_id", "track_seg_point_id")
st_write(
     subset(gpxData, select = layers),
     dsn = filePath,
     layer = "track_points",
     driver = "GPX"
 )
 
 # Read the data back in:
 gpxData_backin <- st_read(filePath)
 identical(gpxData, gpxData_backin)

}
