% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Abundance}
\alias{Abundance}
\title{Some title}
\usage{
Abundance(DensityData, StratumArea)
}
\arguments{
\item{DensityData}{The \code{\link{DensityData}} data.}

\item{StratumArea}{The \code{\link{StratumArea}} data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
