% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Abundance}
\alias{Abundance}
\title{Calculate abundance of a stratum}
\usage{
Abundance(MeanDensityData, StratumAreaData)
}
\arguments{
\item{StratumAreaData}{The \code{\link{StratumAreaData}} data.}
}
\value{
An object of StoX data type \code{\link{AbundanceData}}.
}
\description{
This function calcualtes abundance as the product of density and area of each stratum.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to individuals.
}
