% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AbundanceData}
\alias{AbundanceData}
\title{Abundance}
\description{
The AbundanceData model data is a table holding the abundance for each \code{\link{Stratum}}/\code{\link{Layer}}.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
