% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{AcousticDensity}
\alias{AcousticDensity}
\title{Calculate number density from NASC in length intervals}
\usage{
AcousticDensity(NASCData, m = 20, a = -70, d = 1)
}
\arguments{
\item{parameterName}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function takes NASC and
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
