% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AcousticTargetStrength}
\alias{AcousticTargetStrength}
\title{Acoustic target strength}
\description{
The AcousticTargetStrength process data is a list of two tables, the AcousticTargetStrengthModel, holding only the string naming the model to use, and the AcousticTargetStrengthTable, holding the table of the parameters/values of the model. See \code{\link{DefineAcousticTargetStrength}} for details of the different models.
}
\seealso{
This data type is produced by \code{\link{DefineAcousticTargetStrength}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
