% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{AcousticTargetStrength}
\alias{AcousticTargetStrength}
\title{Acoustic target strength}
\description{
The AcousticTargetStrength process data is a list of two tables, the TargetStrengthMethod, holding only the string naming the method to use, and the TargetStrengthTable, holding the parameters/values of the method. See \code{\link{DefineAcousticTargetStrength}} for details on the different methods.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
