% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{aggregateBaselineDataOneTable}
\alias{aggregateBaselineDataOneTable}
\title{Function to aggregate baseline data}
\usage{
aggregateBaselineDataOneTable(
  stoxData,
  TargetVariable,
  aggregationFunction = getRstoxBaseDefinitions("reportFunctions")$functionName,
  subTable = character(),
  GroupingVariables = character(),
  na.rm = FALSE,
  WeightingVariable = character()
)
}
\arguments{
\item{stoxData}{Output from any StoX function.}

\item{TargetVariable}{The variable to report.}

\item{aggregationFunction}{The function to apply, one of "summaryStox", "sum", "mean", "weighted.mean", "median", "min", "max", "sd", "var", "cv", "summary", "quantile", "percentile_5_95".}

\item{subTable}{The name of the sub table to aggregate on, if \code{stoxData} is a list of tables.}

\item{GroupingVariables}{The variables to report by, e.g. "Stratum" or "SpeciesCategory". Can be given as a comma separated list of variables (with no spaces).}

\item{na.rm}{Used in the function specified by \code{aggregationFunction}.}

\item{WeightingVariable}{The variable to weight by. Only relevant for \code{ReportFunction} "weighted.mean".}
}
\value{
An aggregated version of the input \code{stoxData}.
}
\description{
Function to aggregate baseline data
}
\seealso{
\code{\link{ReportSuperIndividuals}}
}
