% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{aggregateBaselineDataOneTable}
\alias{aggregateBaselineDataOneTable}
\title{Function to aggregate baseline data}
\usage{
aggregateBaselineDataOneTable(
  stoxData,
  subTable = character(),
  TargetVariable,
  ReportFunction = getReportFunctions(),
  GroupingVariables = character(),
  InformationVariables = character(),
  na.rm = FALSE,
  padWithZerosOn = character(),
  Specification = list(),
  uniqueGroupingVariablesToKeep = NULL
)
}
\arguments{
\item{stoxData}{Output from any StoX function.}

\item{subTable}{The name of the sub table to aggregate on, if \code{stoxData} is a list of tables.}

\item{TargetVariable}{The variable to report.}

\item{ReportFunction}{The name of a function to report the Baseline process output by. This must be a function returning a single value. See \code{\link{ReportFunctions}} for implemented funcitons.}

\item{GroupingVariables}{The variables to report by. For most applications \code{GroupingVariables} should include "Survey" and "SpeciesCategory", unless the user needs to sum over all Survey or SpeciesCategory.}

\item{InformationVariables}{Variables to include as columns to the end of the report table. These cannot have more unique combinations than the \code{GroupingVariables}.}

\item{na.rm}{Used in the function specified by \code{ReportFunction}.}

\item{padWithZerosOn}{Character vector giving the variables for which missing values should be padded with zeros. This is used particularly for bootstrapping, where a fish length missing in a bootstrap run should be considered as samples with zero individuals, and not missing, so that summary statistics end up taking all bootstrap replicates into account (if not a mean would be overestimated). When padWithZerosOn has positive length, padding with zeros is applied to this variable and to the \code{GroupingVariables}.}

\item{Specification}{A named list of specification parameters used in the \code{ReportFunction}.}

\item{uniqueGroupingVariablesToKeep}{A data.table holding unique combinations to extract from the data, used by RstoxFramework::ReportBootstrap() to discard combinations of the grouping variables what do not exist in the data. Such combinations are introduced with the CJ operation when padWithZerosOn is given.}
}
\value{
An aggregated version of the input \code{stoxData}.
}
\description{
Function to aggregate baseline data
}
\seealso{
\code{\link{ReportSuperIndividuals}}
}
