% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R, R/LengthDistribution.R
\name{AssignmentLengthDistribution}
\alias{AssignmentLengthDistribution}
\title{Assignment length distribution data}
\usage{
AssignmentLengthDistribution(LengthDistributionData, BioticAssignment)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
The AssignmentLengthDistribution model data is a table averaged length distribution for each combination of Stratum (acoustic)PSU and (acoustic)Layer.

This funciton calculates weighted average of the length distribution of hauls assigned to each acoustic PSU and Layer. The weights are set by \code{\link{BioticAssignmentWeighting}}.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}

\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
