% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{AssignmentLengthDistribution}
\alias{AssignmentLengthDistribution}
\title{Length distribution assigned to each cell of acoustic PSU and Layer}
\usage{
AssignmentLengthDistribution(LengthDistributionData, BioticAssignment)
}
\arguments{
\item{LengthDistributionData}{A list of \code{\link{LengthDistributionData}} data.}

\item{BioticAssignment}{A list of \code{\link{BioticAssignment}} data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This funciton calculates weighted average of the length distribution of hauls assigned to each acoustic PSU and Layer. The weights are set by \code{\link{BioticAssignmentWeighting}}.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
