% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\docType{data}
\name{backwardCompatibility_RstoxBase}
\alias{backwardCompatibility_RstoxBase}
\title{Backward compabitibility actions:}
\format{
An object of class \code{list} of length 7.
}
\usage{
backwardCompatibility_RstoxBase
}
\description{
Backward compabitibility actions:
}
\keyword{datasets}
