% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{BioticAssignment}
\alias{BioticAssignment}
\title{Biotic assignment}
\description{
The BioticAssignment process data is a table containing the columns Stratum, PSU, Layer, Haul and WeightingFactor, where Haul is a list of all Hauls assigned to each acoustic PSU of each acoustic layer.
}
\seealso{
This data type is produced by \code{\link{DefineBioticAssignment}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
