% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{BioticAssignmentWeighting}
\alias{BioticAssignmentWeighting}
\title{Assignnment of biotic hauls to acoustic PSUs}
\usage{
BioticAssignmentWeighting(
  BioticAssignment,
  WeightingMethod = c("Equal", "NumberOfLengthSamples", "NASC", "NormalizedTotalWeight",
    "NormalizedTotalCount", "SumWeightedCount", "InverseSumWeightedCount"),
  StoxBioticData,
  LengthDistributionData,
  MaxNumberOfLengthSamples = 100,
  StoxAcousticData,
  Radius,
  LengthExponentTable
)
}
\arguments{
\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data.}

\item{Radius}{Numeric: The radius inside which to assign biotic stations to each acoustic PSU.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Stratum", assign all stations of each stratum to all acoustic PSUs; "Radius", to assign all stations within the radius given in \code{Radius} to each acoustic PSU; and "EllipsoidalDistance" to provide the \code{EllipsoidalDistanceTable} specifying the axes of an ellipsoid inside which to assign stations to acoustic PSUs.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}

\item{AcousticLayer}{The \code{\link{AcousticLayer}} process data.}

\item{EllipsoidalDistanceTable}{Not yet implemented.}
}
\value{
An object of StoX data type \code{\link{BioticAssignment}}.
}
\description{
This function defines the \code{\link{BioticAssignment}} process data, linking biotic Hauls with acoustic PSUs.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{BioticAssignment}}.
}
