% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{BioticAssignmentWeighting}
\alias{BioticAssignmentWeighting}
\title{Weighting of biotic hauls in biotic assignment}
\usage{
BioticAssignmentWeighting(
  BioticAssignment,
  WeightingMethod = c("Equal", "NumberOfLengthSamples", "NASC", "NormalizedTotalWeight",
    "NormalizedTotalCount", "SumWeightedCount", "InverseSumWeightedCount"),
  StoxBioticData,
  LengthDistributionData,
  MaxNumberOfLengthSamples = 100,
  StoxAcousticData,
  Radius = double(),
  LengthExponent = double()
)
}
\arguments{
\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}

\item{WeightingMethod}{Character: A string naming the method to use, one of "Equal", giving weight 1 to all Hauls; "NumberOfLengthSamples", weighting hauls by the number of length samples; "NASC", weighting by the surrounding NASC converted by the haul length distribution to a density equivalent; "NormalizedTotalWeight", weighting hauls by the total weight of the catch, normalized by dividing by towed distance; "NormalizedTotalCount", the same as "NormalizedTotalWeight" but for total count, "SumWeightedCount", weighting by the summed WeightedCount of the input LengthDistributionData; and "InverseSumWeightedCount", weighting by the inverse of the summed WeightedCount.}

\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{MaxNumberOfLengthSamples}{For \code{WeightingMethod} = "NumberOfLengthSamples": Values of the number of length samples that exceed \code{MaxNumberOfLengthSamples} are set to \code{MaxNumberOfLengthSamples}. This avoids giving too high weight to e.g. experimental hauls with particularly large length samples.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}

\item{Radius}{For \code{WeightingMethod} = "NASC": The radius inside which the average NASC is calculated.}

\item{LengthExponent}{For \code{WeightingMethod} = "NASC": A table linking AcousticCategory with the LengthExponent used to convert from NASC to density.}
}
\value{
An object of StoX data type \code{\link{BioticAssignment}}.
}
\description{
This function puts weights to the hauls assigned to acoustic PSUs in \code{\link{BioticAssignment}} process data.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{DefineBioticAssignment}} for generating BioticAssignment.
}
