% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{BootstrapData}
\alias{BootstrapData}
\title{Bootstrap data}
\description{
The BootstrapData model data is a list of outputs from the baseline generated by re-running baseline with random sampling of the PSUs and assignments (in the case of acoustic-trawl estimates).

The BootstrapData model data is a list of data from one or more baseline processes stacked for all bootstrap runs.
}
\seealso{
This data type is produced by the Bootstrap function of RstoxFramework. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxBase}}

See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
