% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticLayer}
\alias{DefineAcousticLayer}
\title{Acoustic Layer}
\usage{
DefineAcousticLayer(
  processData,
  StoxAcousticData,
  DefinitionMethod = c("WaterColumn", "HighestResolution", "UserDefined"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  UseProcessData = FALSE
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{StoxAcousticData}{A list of \code{\link[RstoxDatas]{StoxAcousticData}} data.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "EDSUToPSU", which sets each EDSU as a PSU, and "None" for pure manual actions by the user.}

\item{UseProcessData}{Logical: If TRUE use the existing funciton output in the process.}
}
\value{
An object of StoX data type \code{\link{AcousticLayer}}.
}
\description{
This function defines the \code{\link{AcousticLayer}} process data, which sets the range intervals of the acoustic layers used in acoustic-trawl estimation models in StoX.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{AcousticPSU}}.
}
