% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticPSU}
\alias{DefineAcousticPSU}
\title{Acoustic PSU}
\usage{
DefineAcousticPSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxAcousticData,
  DefinitionMethod = c("EDSUToPSU", "DeleteAllPSUs")
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "EDSUToPSU", which sets each EDSU as a PSU, and "DeleteAllPSUs" to delete all PSUs.}
}
\value{
An object of StoX data type \code{\link{AcousticPSU}}.
}
\description{
This function defines the \code{\link{AcousticPSU}} process data, linking strata, acoustic PSUs and EDSUs.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
Biotic PSUs are generated using \code{\link{DefineBioticPSU}}. For the vertical resolution (Layer) see \code{\link{DefineBioticLayer}} and \code{\link{DefineAcousticLayer}}.
}
