% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticPSU}
\alias{DefineAcousticPSU}
\title{Acoustic PSU}
\usage{
DefineAcousticPSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxAcousticData,
  DefinitionMethod = c("Manual", "EDSUToPSU", "DeleteAllPSUs", "PreDefined",
    "ResourceFile"),
  FileName = character(),
  AcousticPSU
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StoxAcousticData}{\code{\link[RstoxData]{StoxAcousticData}} (defined in RstoxData).}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Manual" for manual tagging of EDUSs to PSUs; "EDSUToPSU", which sets each EDSU as a PSU; "DeleteAllPSUs" to delete all PSUs; "PreDefined" to read from a previous process; and "ResourceFile" to read from a project.xml file from StoX <= 2.7. Note that in the latter case it is assumed that the contents of the project.xml file is actually used in the StoX <= 2.7 project, i.e. that UseProcessData = TRUE in DefineAcousticPSU().}

\item{FileName}{The path to a resource file from which to read PSUs, in the case that \code{DefinitionMethod} is "ResourceFile". Currently, only a project.xml file from StoX 2.7 can be read. Must include file extension.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}
}
\value{
An object of StoX data type \code{\link{AcousticPSU}}.
}
\description{
Defines the \code{\link{AcousticPSU}} process data, linking strata, acoustic PSUs and EDSUs.
}
\seealso{
Biotic PSUs are generated using \code{\link{DefineBioticPSU}}. For the vertical resolution (Layer) see \code{\link{DefineBioticLayer}} and \code{\link{DefineAcousticLayer}}.
}
