% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticPSU}
\alias{DefineAcousticPSU}
\title{Acoustic PSU}
\usage{
DefineAcousticPSU(processData, StratumPolygon, StoxAcousticData,
  DefinitionMethod = c("EDSUToPSU", "None"), UseProcessData = FALSE)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{StratumPolygon}{A list of \code{\link{StratumPolygon}} process data.}

\item{StoxAcousticData}{A list of \code{\link{StoxAcousticData}} data.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "EDSUToPSU", which sets each EDSU as a PSU, and "None" for pure manual actions by the user.}

\item{UseProcessData}{Logical: If TRUE use the existing funciton output in the process.}
}
\value{
An object of StoX data type \code{\link{AcousticPSU}}.
}
\description{
This function defines the \code{\link{AcousticPSU}} process data, linking strata, acoustic PSUs and EDSUs.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{AcousticPSU}}.
}
