% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticPSUByTime}
\alias{DefineAcousticPSUByTime}
\title{Re-define Acoustic PSU}
\usage{
DefineAcousticPSUByTime(
  processData,
  UseProcessData = FALSE,
  StoxAcousticData,
  AcousticPSU
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}
}
\value{
An object of StoX data type \code{\link{AcousticPSU}}.
}
\description{
Re- defines the \code{\link{AcousticPSU}} process data, linking strata, acoustic PSUs and EDSUs, based on the start and end time of preivously defined acoustic PSUs.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{DefineAcousticPSU}} for defining AcouosticPSU in the first place.
}
