% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticTargetStrength}
\alias{DefineAcousticTargetStrength}
\title{Acoustic target strength definition}
\usage{
DefineAcousticTargetStrength(
  processData,
  DefinitionMethod = c("Table", "ResourceFile"),
  ParameterTable = data.table::data.table(),
  FileName,
  UseProcessData = FALSE
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Table", for providing the acoustic target strength parameters in the table \code{ParameterTable}; and "ResourceFile" for reading the acoustic tfarget strength table from the text file \code{FileName}.}

\item{ParameterTable}{A table of the columns AcousticCategory, Frequency, m, a and d.}

\item{FileName}{A file from which to read the \code{ParameterTable}.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
This function returns a table of parameters specifying the acoustic target strength as a function of length for different values of user selected variables in the NASC data.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
