% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticTargetStrength}
\alias{DefineAcousticTargetStrength}
\title{Acoustic target strength definition}
\usage{
DefineAcousticTargetStrength(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("ResourceFile", "TargetStrengthTable"),
  TargetStrengthMethod = c("LengthDependent", "LengthAndDepthDependent",
    "TargetStrengthByLength"),
  TargetStrengthTable = data.table::data.table(),
  FileName = character()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "TargetStrengthTable", for providing the acoustic target strength parameters in the table \code{TargetStrengthTable}; and "ResourceFile" for reading the acoustic tfarget strength table from the text file \code{FileName}.}

\item{TargetStrengthMethod}{Character: The target strength methdo/function to use. Currently implemented are "LengthDependent", "LengthAndDepthDependent", "LengthExponent" and "TargetStrengthByLength". See Details.}

\item{TargetStrengthTable}{A table holding the specification of the target strength function/table. The first two columns are AcocusticCategory and Frequenccy. See details for other columns.}

\item{FileName}{A file from which to read the \code{TargetStrengthTable}.}
}
\value{
An \code{\link{AcousticTargetStrength}} object.
}
\description{
This function returns a table of parameters specifying the acoustic target strength as a function of length for different values of user selected variables in the NASC data.
}
\details{
The \code{TargetStrengthMethod} has the following possible values: 
\enumerate{
  \item LengthDependent, applying the logarithmic function TargetStrength = Targetstrength0 + LengthExponent * log10(Length). Required columns: Targetstrength0 and LengthExponent.
  \item LengthAndDepthDependent, applying the logarithmic function TargetStrength = Targetstrength0 + LengthExponent * log10(Length) + DepthExponent * log10(1 + Depth/10). Required columns: Targetstrength0, LengthExponent and DepthExponent.
  \item TargetStrengthByLength, applying a table of TargetStrength and TotalLength. Required columns: TargetStrength and TotalLength.
  \item LengthExponent, applying the logarithmic function TargetStrength = LengthExponent * log10(Length). Required columns: LengthExponent.
}
The parameters/values can be given by tables with the first columns being AcousticCategory and Frequency, or as a csv file.
}
\seealso{
\code{\link{AcousticDensity}} for applying the AcousticTargetStrength.
}
