% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineAcousticTargetStrength}
\alias{DefineAcousticTargetStrength}
\title{Acoustic target strength definition}
\usage{
DefineAcousticTargetStrength(
  processData,
  DefinitionMethod = c("Table", "ResourceFile"),
  ParameterTable = data.table::data.table(),
  FileName,
  UseProcessData = FALSE
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{ParameterTable}{A table of target strength to length equation parameters, including the column \code{EquationType} and \code{m} and \code{a} for \code{EquationType = "Standard"} and \code{m}, \code{a} and \code{d} for \code{EquationType = "DepthDependent"}. Also allowed are columns representing variables in the \code{\link{NASCData}}, such as "SpeciesCategory" and "Frequency".}

\item{FileName}{The path to a geoJSON file, shape file (folder) or a Stox multipolygon WKT file. Must include file extension.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
This function returns a table of parameters specifying the acoustic target strength as a function of length for different values of user selected variables in the NASC data.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
