% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineBioticAssignment}
\alias{DefineBioticAssignment}
\title{Assignnment of biotic hauls to acoustic PSUs}
\usage{
DefineBioticAssignment(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("Stratum", "Radius", "EllipsoidalDistance",
    "DeleteAllAssignments"),
  StoxBioticData,
  StratumPolygon,
  AcousticPSU,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution",
    "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL,
  StoxAcousticData,
  Radius = double(),
  MinNumberOfHauls = double(),
  Distance = double(),
  TimeDifference = double(),
  BottomDepthDifference = double(),
  LongitudeDifference = double(),
  LatitudeDifference = double()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "Stratum", assign all stations of each stratum to all acoustic PSUs; "Radius", to assign all stations within the radius given in \code{Radius} to each acoustic PSU; and "EllipsoidalDistance" to provide \code{MinNumberOfHauls}, \code{Distance}, \code{TimeDifference}, \code{BottomDepthDifference}, \code{LongitudeDifference} and \code{LatitudeDifference}, specifying the axes of an ellipsoid inside which to assign stations to acoustic PSUs.}

\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}

\item{LayerDefinition}{The method to use for defining the Layers, one of \code{FunctionParameter} to define the Layers on the fly in this function, or \code{FunctionInput} to import Layer process data from a previously run process by \code{AcousticLayer}.}

\item{LayerDefinitionMethod}{See \code{\link{DefineAcousticLayer}}}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}

\item{AcousticLayer}{The \code{\link{AcousticLayer}} process data.}

\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}

\item{Radius}{Numeric: The radius inside which to assign biotic stations to each acoustic PSU.}

\item{MinNumberOfHauls}{For DefinitionMethod "EllipsoidalDistance": Integer minimum number of hauls selected inside the ellipsoid. If the number of hauls inside the ellispoid is lower than the \code{MinNumberOfHauls}, the \code{MinNumberOfHauls} closest Hauls will be used (in ellipsoidal distance).}

\item{Distance}{For DefinitionMethod "EllipsoidalDistance": The semi axis of the ellipsoid representing distance in nautical miles.}

\item{TimeDifference}{For DefinitionMethod "EllipsoidalDistance": The semi axis of the ellipsoid representing time difference in hours.}

\item{BottomDepthDifference}{For DefinitionMethod "EllipsoidalDistance": The semi axis of the ellipsoid representing difference in bottom depth in meters.}

\item{LongitudeDifference}{For DefinitionMethod "EllipsoidalDistance": The semi axis of the ellipsoid representing difference in longitude in degrees.}

\item{LatitudeDifference}{For DefinitionMethod "EllipsoidalDistance": The semi axis of the ellipsoid representing difference in latitude in degrees.}
}
\value{
An object of StoX data type \code{\link{BioticAssignment}}.
}
\description{
This function defines the \code{\link{BioticAssignment}} process data, linking biotic Hauls with acoustic PSUs.
}
\details{
See Equation 8 in "Factors affecting the diel variation in commercial CPUE of Namibian hake - Can new information improve standard survey estimates?".
}
\examples{
x <- 1

}
\seealso{
\code{\link{BioticAssignmentWeighting}} for weighting BioticAssignment.
}
