% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineBioticPSU}
\alias{DefineBioticPSU}
\title{Biotic PSU}
\usage{
DefineBioticPSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxBioticData,
  DefinitionMethod = c("Manual", "StationToPSU", "DeleteAllPSUs", "ResourceFile"),
  FileName = character()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} (defined in RstoxData).}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "StationToPSU", which sets each Station as a PSU, and "DeleteAllPSUs" to delete all PSUs.}

\item{FileName}{The path to a resource file from which to read PSUs, in the case that \code{DefinitionMethod} is "ResourceFile". Currently, only a project.xml file from StoX 2.7 can be read.}
}
\value{
An object of StoX data type \code{\link{BioticPSU}}.
}
\description{
This function defines the \code{\link{BioticPSU}} process data, linking strata, biotic PSUs and Stations
}
\seealso{
Acousic PSUs are generated using \code{\link{DefineAcousticPSU}}. For the vertical resolution (Layer) see \code{\link{DefineBioticLayer}} and \code{\link{DefineAcousticLayer}}.
}
