% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineLayer}
\alias{DefineLayer}
\title{Define Layers}
\usage{
DefineLayer(
  processData,
  StoxData,
  DefinitionMethod = c("WaterColumn", "HighestResolution", "UserDefined"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  UseProcessData = FALSE,
  modelType = c("Acoustic", "SweptArea")
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "EDSUToPSU", which sets each EDSU as a PSU, and "None" for pure manual actions by the user.}

\item{UseProcessData}{Logical: If TRUE use the existing funciton output in the process.}

\item{StoxBioticData}{A list of \code{\link[RstoxDatas]{StoxBioticData}} data.}
}
\value{
An object of StoX data type \code{\link{SweptAreaLayer}}.
}
\description{
This function defines the \code{\link{SweptAreaLayer}} process data, which sets the range intervals of the swetp-area layers used in swept-area estimation models in StoX.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{SweptAreaPSU}}.
}
