% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefinePSU}
\alias{DefinePSU}
\title{Define PSU}
\usage{
DefinePSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxData = NULL,
  MergedStoxDataStationLevel = NULL,
  DefinitionMethod = c("Manual", "Identity", "DeleteAllPSUs", "PreDefined",
    "ResourceFile"),
  FileName = character(),
  PSUType = c("Acoustic", "Biotic"),
  SavePSUByTime = FALSE,
  PSUProcessData
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StoxData}{Either \code{\link[RstoxData]{StoxBioticData}} or \code{\link[RstoxData]{StoxAcousticData}} data.}

\item{MergedStoxDataStationLevel}{The merged StoxData at the station level. Used in \code{meanRawResolutionData}.}

\item{DefinitionMethod}{A string naming the method to use, see \code{\link{DefineBioticPSU}} and \code{\link{DefineAcousticPSU}}.}

\item{FileName}{The path to a resource file from which to read PSUs, in the case that \code{DefinitionMethod} is "ResourceFile". Currently, only a project.xml file from StoX 2.7 can be read. Must include file extension.}

\item{PSUType}{Character: A string naming the type of model, either "Acoustic" or "Biotic"}

\item{SavePSUByTime}{Logical: If TRUE save the start and end times of sequences of EDSUs or Stations for each PSU.}

\item{PSUProcessData}{Previously generated PSU process data, one of \code{\link{AcousticPSU}} or \code{\link{BioticPSU}}.}
}
\value{
An list of two objects, Stratum_PSU and SSU_PSU.
}
\description{
Underlying function for \code{\link{DefineBioticPSU}} and \code{\link{DefineAcousticPSU}}.
}
\seealso{
\code{\link{DefineAcousticPSU}} and \code{\link{DefineBioticPSU}}..
}
