% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefinePSU}
\alias{DefinePSU}
\title{Definne PSU}
\usage{
DefinePSU(
  processData,
  UseProcessData = FALSE,
  StratumPolygon,
  StoxData,
  DefinitionMethod = c("Identity", "None"),
  PSUType = c("Acoustic", "Biotic")
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{StoxData}{Either \code{\link[RstoxData]{StoxBioticData}} or \code{\link[RstoxData]{StoxAcousticData}} data.}

\item{DefinitionMethod}{Character: A string naming the method to use, see \code{\link{DefineBioticPSU}} and \code{\link{DefineAcousticPSU}}.}

\item{PSUType}{Character: A string naming the type of model, either "Acoustic" or "Biotic".}
}
\value{
An list of two objects, Stratum_PSU and SSU_PSU.
}
\description{
Underlying function for \code{\link{DefineBioticPSU}} and \code{\link{DefineAcousticPSU}}.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{DefineAcousticPSU}} and \code{\link{DefineBioticPSU}}..
}
