% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{DefineStratumPolygon}
\alias{DefineStratumPolygon}
\title{Define stratum multipolygon}
\usage{
DefineStratumPolygon(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("ResourceFile", "Manual"),
  FileName
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{A string naming the method to use, one of "ResourceFile", to read the file \code{FileName} holding the stratum multipolygon, or "Manual" to start off with no strata and create the strata manually in the map of the StoX GUI. Strata can be added, modified and removed in the StoX GUI.}

\item{FileName}{The path to a \href{https://geojson.org/}{GeoJSON} file, \href{https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm}{shapefile} (may be a folder) or a \code{\link{StoX_multipolygon_WKT}} file. Must include file extension.}
}
\value{
An object of StoX data type \code{\link{StratumPolygon}}.
}
\description{
This function reads a \href{https://geojson.org/}{GeoJSON} file, \href{https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm}{shapefile} or a \code{\link{StoX_multipolygon_WKT}} file and returns an object of StoX data type \code{\link{StratumPolygon}} file.
}
\details{
The parameter \code{UseProcessData} is always set to TRUE when running a process, and needs to be explicitely set to FALSE to enable reading a file (It's set to FALSE at the moment).

Polygons are expected to have the the following projection: getRstoxBaseDefinitions("proj4string")
}
\references{
Additonal information on GeoJSON and shapefiles specification can be found here:
\url{https://geojson.org/}

\url{https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm}
}
\seealso{
\code{\link{StratumArea}} for calculating the area of the strata.
}
