% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{DefineStratumPolygon}
\alias{DefineStratumPolygon}
\title{Define stratum multipolygon}
\usage{
DefineStratumPolygon(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("ResourceFile", "Manual"),
  FileName = character(),
  StratumNameLabel = character(),
  SimplifyStratumPolygon = FALSE,
  SimplificationFactor = 1
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{A string naming the method to use, one of "ResourceFile", to read the file \code{FileName} holding the stratum multipolygon, or "Manual" to start off with no strata and create the strata manually in the map of the StoX GUI. Strata can be added, modified and removed in the StoX GUI.}

\item{FileName}{The path to a \href{https://geojson.org/}{GeoJSON} file, \href{https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm}{shapefile} (may be a folder) or a \code{\link{StoX_multipolygon_WKT}} file. Must include file extension.}

\item{StratumNameLabel}{The name of the attribute representing the stratum names in the GeoJSON file or shapefile.}

\item{SimplifyStratumPolygon}{Logical: If TRUE a simplification algorithm is applied to the stratum polygons (sf::st_simplify(), which "uses the GEOS implementation of the Douglas-Peucker algorithm to reduce the vertex count" as per quote from \href{https://geocompr.robinlovelace.net/geometric-operations.html#simplification}{Geocomputation with R}). This option is useful for very complex stratum polygon files to reduce time for opening, running and saving a project.}

\item{SimplificationFactor}{A value between 0 and 1 specifying the desired object size of the stratum polygon after simplification. An iterative method is used to derive the to the desired object size. Strata that end up as empty strata after simplification are left un-simplified, and will contribute to a larger final object size than the desired object size. Note that larger values require more  time than smaller values. It is advisable to start with a low value (e.g. 0.01) and check the resolution of the strata borders in the map, and gradually increase \code{SimplificationFactor} until an acceptable resolution is obtained. Note that simplification of stratum polygons can lead to inaccuracies such as gaps between strata, which can lead to hauls falling outside of any stratum, causing under-estimation of survey indices, or failure to complete the model in the case of SplitNASC models which require all hauls to be located in a stratum. If such problems occur, the strata can be modified manually after simplification to include all hauls. Another problem is the change in stratum areas, which is reported to the console when running the simplification. A very rough rule of thumb is that \code{SimplificationFactor} = 0.1 seems to be a compromise between increased speed and reduction of stratum border resolution.}
}
\value{
An object of StoX data type \code{\link{StratumPolygon}}.
}
\description{
This function reads a \href{https://geojson.org/}{GeoJSON} file, \href{https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm}{shapefile} or a \code{\link{StoX_multipolygon_WKT}} file and returns an object of StoX data type \code{\link{StratumPolygon}} file.
}
\details{
The parameter \code{UseProcessData} is always set to TRUE when running a process, and needs to be explicitely set to FALSE to enable reading a file (It's set to FALSE at the moment).

Polygons are expected to have the the following projection: getRstoxBaseDefinitions("proj4string")
}
\references{
Additonal information on GeoJSON and shapefiles specification can be found here:
\url{https://geojson.org/}

\url{https://doc.arcgis.com/en/arcgis-online/reference/shapefiles.htm}
}
\seealso{
\code{\link{StratumArea}} for calculating the area of the strata.
}
