% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.R
\name{DefineSurvey}
\alias{DefineSurvey}
\title{Biotic Survey}
\usage{
DefineSurvey(
  processData,
  UseProcessData = FALSE,
  DefinitionMethod = c("AllStrata", "SurveyTable"),
  StratumPolygon,
  SurveyTable = data.table::data.table()
)
}
\arguments{
\item{processData}{The current data produced by a previous instance of the function.}

\item{UseProcessData}{Logical: If TRUE use the existing function output in the process.}

\item{DefinitionMethod}{Character: A string naming the method to use, one of "AllStrata", which defines all strata as the same survey named "Survey"; and "SurveyTable", which requires the \code{SurveyTable} to be given.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{SurveyTable}{A table of the two columns Stratum and Survey.}
}
\value{
An object of StoX data type \code{\link{BioticPSU}}.
}
\description{
This function defines the Strata associated to different surveys (in the sense that a separate estimate should be made for those strata).
}
\seealso{
Acousic PSUs are generated using \code{\link{DefineAcousticPSU}}. For the vertical resolution (Layer) see \code{\link{DefineBioticLayer}} and \code{\link{DefineAcousticLayer}}.
}
