% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{formatOutput}
\alias{formatOutput}
\title{Function to format the output of a function returning StoX data (ModelData or ProcessData).}
\usage{
formatOutput(
  data,
  dataType,
  keep.all = TRUE,
  allow.missing = FALSE,
  primaryColumnOrder = NULL,
  secondaryColumnOrder = NULL,
  secondaryRowOrder = NULL,
  removeStoXKeys = FALSE
)
}
\arguments{
\item{data}{A table or list of tables to return from the StoX function.}

\item{dataType}{The data type to format against.}

\item{keep.all}{Logical: If TRUE keep all columns, and if FALSE delete undefined columns.}

\item{allow.missing}{Logical: If TRUE allow for unrelevant column names defined in \code{secondaryColumnOrder}.}

\item{primaryColumnOrder}{Character: A vector of names of columns that should be placed first regardless of the definition of columns specified in the dataTypeDefinition.}

\item{secondaryColumnOrder, secondaryRowOrder}{A vector of column names specifying order of column not defined by \code{\link{getDataTypeDefinition}} used to prioritize when ordering columns and rows, respectively.}

\item{removeStoXKeys}{Logical: If TRUE remove the key columns, which are those ending with "Key".}
}
\description{
The function removes duplicated columns, orders the columns as per the order defined by \code{\link{getDataTypeDefinition}}, optionally removed undefined columns, orders the rows, and finally deletes any data.table keys from the output tables.
}
