% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_plot_arguments}
\alias{general_plot_arguments}
\title{General ploting parameters of RstoxBase}
\arguments{
\item{UseDefaultMapSettings}{Logical: If TRUE (default) use the default settings of the map, specifically LandColor, BorderColor, OceanColor and GridColor. Setting this to FALSE will show all map options of the plotting function in the StoX GUI.}

\item{UseDefaultPointSettings}{Logical: If TRUE (default) use the default point settings of the plot, specifically PointColor, MaxPointSize and MinPointSize. Setting this to FALSE will show all point size options of the plotting function in the StoX GUI.}

\item{UseDefaultTrackSettings}{Logical: If TRUE (default) use the default point size settings of the plotting function. Setting this to FALSE will show all point size options of the plotting function in the StoX GUI.}

\item{UseDefaultAspectSettings}{Logical: If TRUE (default) use the default aspect settings of the plot, specifically Zoom, LongitudeMin, LongitudeMax, LatitudeMin, LatitudeMax, LongitudeCenter and LatitudeCenter. Setting this to FALSE will show all aspect options of the plotting function in the StoX GUI.}

\item{UseDefaultLabelSettings}{Logical: If TRUE (default) use the default text settings of the plot, specifically Title, AxisTitleSize, AxisTickSize, LegendTitleSize and LegendTextSize. Setting this to FALSE will show all text options of the plotting function in the StoX GUI.}

\item{UseDefaultFileSettings}{Logical: If TRUE (default) use the default file settings of the output file of the plot, specifically Format, Width, Height and DotsPerInch. Setting this to FALSE will show all file options of the plotting function in the StoX GUI.}

\item{UseDefaultAcousticPSULabelSettings}{Logical: If TRUE (default) use the default settings for acoustic PSU labels in the plots (text position, color and size). Setting this to FALSE will show all acoustic PSU options of the plotting function in the StoX GUI.}

\item{UseDefaultStratumPolygonSettings}{Logical: If TRUE (default) use the default settings for plotting the StratumPolygon. Setting this to FALSE will show all Stratum plotting options of the plotting function in the StoX GUI.}

\item{Title}{Character: The title of the plot, defaulted to no title.}

\item{AxisTitleSize}{Numeric: The size of the axis titles, defaulted to 20.}

\item{AxisTickSize}{Numeric: The size of the axis tick marks, defaulted to 20.}

\item{LegendTitleSize}{Numeric: The size of the legend title, defaulted to 20.}

\item{LegendTextSize}{Numeric: The size of the legend text, defaulted to 20.}
}
\description{
General ploting parameters of RstoxBase
}
