% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_plot_arguments}
\alias{general_plot_arguments}
\title{General ploting parameters of RstoxBase}
\arguments{
\item{UseDefaultPlotSettings}{Logical: If TRUE (default) use the default settings of the plotting function. Setting this to FALSE will show all options of the plotting function in the StoX GUI.}

\item{Zoom}{The zoom of the plot, defaulted to 1, which uses the range of the data as the plot limits. Higher values zooms out and lower values zooms in.}

\item{Title}{The title of the plot, defaulted to no title.}

\item{AxisTitleSize}{The size of the axis titles, defaulted to 20.}

\item{AxisTickSize}{The size of the axis tick marks, defaulted to 20.}

\item{LegendTitleSize}{The size of the legend title, defaulted to 20.}

\item{LegendTextSize}{The size of the legend text, defaulted to 20.}

\item{Format}{The format of the output file from plotting functions, defaulted to "png". Other options are defined by \code{\link[ggplot2]{ggsave}}, including "tiff", "jpeg"  and "pdf".}

\item{Width}{The width of the plot in centimeters, defaulted to 17, as per the \href{https://academic.oup.com/icesjms/pages/General_Instructions}{ICES Journal og Marine Science author recommendations}.}

\item{Height}{The height of the plot in centimeters, defaulted to 17.}

\item{DotsPerInch}{The number of dots per inch (DPI), defaulted to 1200.}
}
\description{
General ploting parameters of RstoxBase
}
