% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getBaseUnit}
\alias{getBaseUnit}
\title{Get base unit of a variable of a StoX dataType}
\usage{
getBaseUnit(
  dataType,
  variableName,
  element = c("unit", "quantity"),
  list.out = FALSE
)
}
\arguments{
\item{dataType}{The name of a StoX \code{\link{DataTypes}}.}

\item{variableName}{The variable name to get base units from.}

\item{element}{The element to get, one of "unit" and "quantity".}

\item{list.out}{Logical: If TRUE return both elements in a list.}
}
\description{
Get base unit of a variable of a StoX dataType
}
