% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{getReportFunctionVariableName}
\alias{getReportFunctionVariableName}
\title{Get the name of the target variable after aggregating.}
\usage{
getReportFunctionVariableName(functionName, TargetVariable)
}
\arguments{
\item{functionName}{The aggregation function name.}

\item{TargetVariable}{The variable to report.}
}
\description{
Get the name of the target variable after aggregating.
}
