% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{getStratumNames}
\alias{getStratumNames}
\title{Extract stratum names from a SpatialPolygonsDataFrame}
\usage{
getStratumNames(stratum)
}
\arguments{
\item{stratum}{A \code{\link[sp]{SpatialPolygonsDataFrame}} with a column polygonName of the data of the \code{\link[sp]{SpatialPolygonsDataFrame}} \code{stratum}.}
}
\description{
The stratum names must be stored as the column polygonName of the data of the \code{\link[sp]{SpatialPolygonsDataFrame}} \code{stratum}.
}
