% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{getStratumPolygonList}
\alias{getStratumPolygonList}
\title{Convert StratumPolygon to list of polygon tables}
\usage{
getStratumPolygonList(StratumPolygon)
}
\arguments{
\item{StratumPolygon}{An object of StoX class StratumPolygon, which is s SpatialPolygons object (sp package).}
}
\value{
A table of stratum name and area.
}
\description{
This function extracts the polygons as a named list of tables of class \code{\link[data.table]{data.table}}
}
\examples{

}
\seealso{
\code{\link[RstoxBase]{DefineStrata}} for the \code{StratumPolygon} input to the function.
}
