% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{hasBaseUnit}
\alias{hasBaseUnit}
\title{Does the variable of the StoX dataType have a base unit?}
\usage{
hasBaseUnit(dataType, variableName)
}
\arguments{
\item{dataType}{The name of a StoX \code{\link{DataTypes}}.}

\item{variableName}{The variable name to get base units from.}
}
\description{
Does the variable of the StoX dataType have a base unit?
}
