% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{ImputeSuperIndividuals}
\alias{ImputeSuperIndividuals}
\title{Indivduals to distribute abundance to to create super-individuals}
\usage{
ImputeSuperIndividuals(SuperIndividualsData, Seed = 1)
}
\arguments{
\item{SuperIndividualsData}{The \code{\link{SuperIndividualsData}} data.}

\item{Seed}{An integer giving the seed to use for the random sampling used to obtain the imputed data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function defines and returns the individuals used in the estimation model to which to distribute the abundance to create super-individuals.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to the Individuals.
}
