% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{ImputeSuperIndividuals}
\alias{ImputeSuperIndividuals}
\title{Impute missing information from indivduals with present age for each each IndividualTotalLength}
\usage{
ImputeSuperIndividuals(
  SuperIndividualsData,
  ImputationMethod = c("RandomSampling"),
  ImputeAtMissing = character(),
  ImputeByEqual = character(),
  ToImpute = character(),
  Seed = 1
)
}
\arguments{
\item{SuperIndividualsData}{The \code{\link{SuperIndividualsData}} data.}

\item{ImputationMethod}{The method to use for the imputation. Currently, only "RandomSampling" is implemented, but may be accompanied "Regression" in a coming release.}

\item{ImputeAtMissing}{The name of the variable identifying which individuals to impute data to. In StoX 3.0.0 and older, this was hard coded to IndividualAge.}

\item{ImputeByEqual}{The name of the variables identifying which individuals to impute data from. In StoX 3.0.0 and older, this was hard coded to IndividualTotalLength.}

\item{ToImpute}{The name of the variable(s) to impute. In StoX 3.0.0 and older, this was hard coded to all available variables of the BioticData contained in the \code{\link{SuperIndividualsData}}. The variable specified by \code{ImputeAtMissing} is always imputed, and need not be included in \code{ToImpute}.}

\item{Seed}{An integer giving the seed to use for the random sampling used to obtain the imputed data.}
}
\value{
An object of StoX data type \code{\link{SuperIndividualsData}}.
}
\description{
This function identifies individuals with missing value in the column gievn by \code{ImputeAtMissing} (defaulted to "IndividualAge"), and imputes the variables specified by \code{ToImpute} by the method specified by \code{ImputationMethod} (defaulted to "RandomSampling", which currently is the only option) from individuals with values identical to the target individual in all of the columns specified by \code{ImputeByEqual} (defaulted to c("IndividualTotalLength", "SpeciesCategory")). In the imputation, individual to impute from are randomly selected from individuals at the same Haul, then the same Stratum if no individuals with identical values in \code{ImputeByEqual} are found, and finally the same Survey. If no individuals are found, the missing values are not imputed, but kept as missing values (currently with no warning).
}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to the Individuals.
}
