% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{ImputeSuperIndividuals}
\alias{ImputeSuperIndividuals}
\title{Impute missing super-individual data}
\usage{
ImputeSuperIndividuals(
  SuperIndividualsData,
  ImputationMethod = "RandomSampling",
  ImputeAtMissing = character(),
  ImputeByEqual = character(),
  ToImpute = character(),
  Seed = 1
)
}
\arguments{
\item{SuperIndividualsData}{The \code{\link{SuperIndividualsData}} data.}

\item{ImputationMethod}{The method to use for the imputation. Currently, only "RandomSampling" is implemented, but may be accompanied "Regression" in a coming release.}

\item{ImputeAtMissing}{A single string naming the variable which when missing identifies target individuals to input data \bold{to}. I.e., if \code{ImputeAtMissing} is missing for an individual, perform the imputation. In StoX 3.0.0 and older, \code{ImputeAtMissing} was hard coded to IndividualAge.}

\item{ImputeByEqual}{A vector of strings naming the variable(s) which, when identical to the target individual, identifies the source individuals to impute data \bold{from}. The source individuals need also to have non-missing \code{ImputeAtMissing}. In StoX 3.0.0 and older, \code{ImputeByEqual} was hard coded to c("SpeciesCategory","IndividualTotalLength").}

\item{ToImpute}{A vector of strings naming the variable(s) to impute (copy to the target individual). Only variables that are non-missing in the source individual and missing in the target individual are copied to the target individual. In StoX 3.0.0 and older, \code{ToImpute} was hard coded to all available variables of the BioticData contained in the \code{\link{SuperIndividualsData}}.}

\item{Seed}{An integer giving the seed to use for the random sampling used to obtain the imputed data.}
}
\value{
An object of StoX data type \code{\link{SuperIndividualsData}}.
}
\description{
For each (target) individual with missing value in \code{ImputeAtMissing}, identify all (source) individuals in the haul for which \code{ImputeAtMissing} is non-missing and for which the values in \code{ImputeByEqual} are identical to the target individual. Then sample one of these source individuals, and copy values of \code{ToImpute} to the target individual. Only values that are non-missing are copied from the sampled individual, and  only missing values in the target individual are replaced. If no source individuals are found in the haul, expand the search to the stratum, and finally to the survey. If no source individuals are found in the survey, leave the target individual unchanged.
}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to the Individuals.
}
