% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Individuals}
\alias{Individuals}
\title{Indivduals to distribute abundance to to create super-individuals}
\usage{
Individuals(
  StoxBioticData,
  AbundanceType = c("Acoustic", "SweptArea"),
  BioticAssignment,
  MeanLengthDistributionData
)
}
\arguments{
\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{AbundanceType}{The type of abundance, one of "Acoustic" and "SweptArea".}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}

\item{MeanLengthDistributionData}{The \code{\link{MeanLengthDistributionData}} data.}
}
\value{
An object of StoX datatype \code{\link{IndividualsData}}.
}
\description{
This function defines and returns the individuals used in the estimation model, to which to distribute the abundance to create super-individuals.
}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to the Individuals.
}
