% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Individuals}
\alias{Individuals}
\title{Some title}
\usage{
Individuals(
  StoxBioticData,
  AbundanceType = c("Acoustic", "SweptArea"),
  BioticAssignment,
  SweptAreaPSU,
  SweptAreaLayer
)
}
\arguments{
\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data.}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}

\item{DensityType}{The type of density, one of "Acoustic" and "SweptArea".}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
