% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Individuals}
\alias{Individuals}
\title{Indivduals to distribute abundance to to create super-individuals}
\usage{
Individuals(
  StoxBioticData,
  QuantityType = c("Acoustic", "SweptArea"),
  BioticAssignment,
  MeanLengthDistributionData
)
}
\arguments{
\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} (defined in RstoxData).}

\item{QuantityType}{The type of abundance, one of "Acoustic" and "SweptArea".}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}

\item{MeanLengthDistributionData}{The \code{\link{MeanLengthDistributionData}} data.}
}
\value{
An object of StoX datatype \code{\link{IndividualsData}}.
}
\description{
This function defines and returns the individuals used in the estimation model, to which to distribute the abundance to create super-individuals.
}
\details{
Individuals are retrieved from only Hauls that contribute to the Abundance. For QuantityType = "Acoustic" these are the Hauls that have positive WeightingFactor in the BioticAssignment input, and for QuantityType = "SweptArea" the Hauls that have positive WeightedNumber.

The \code{\link{IndividualsData}} contains variables from \code{\link[RstoxData]{StoxBioticData}} in addition to the columns Stratum and Layer. The Stratum column is not necessarily the actual stratum containing the Haul in which an individual was sampled, but rather the stratum linked to the haul via \code{\link{DefineBioticAssignment}} for acoustic-trawl models and \code{\link{MeanLengthDistribution}} for swept-area models. In detail:
\itemize{
\item DefineAcousticPSU(): Defines acoustic PSUs, which are assigned to a stratum, and that containns EDSUs, possibly from other strata than the assigned stratum.
\item DefineBioticPSU(): Defines biotic PSUs, which are assigned to a stratum, and that conntainns Stations, possibly from other strata than the assigned stratum.
\item DefineBioticAssignment(): Assignes hauls to acoustic PSUs, possibly from other strata than the stratum assigned to the acoustic PSU. In this datatype it is possible that EDSUs of an acoustic PSU are located in different strata, or even that all the EDSUs of the acoustic PSU are located in another stratum than the assigned stratum; and that the hauls assigned to the acoustic PSU are located in yet another stratum.
\item In Individuals() the StoxBioticData are merged with BioticAssignment in the case of acoustic-trawl models and with MeanLengthDistributionData in the case of swept-area models, by the Haul identifier stored in the StoxBioticData, the BioticAssignment, and in the Resolution table of the MeanLengthDistributionData. As the hauls may be linked to a different statum than the one containing the haul, as per the description of DefineBioticAssignment() above, the Stratum column of the IndividualsData may not correspond to the actual stratum of the haul.
}
}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to the Individuals.
}
