% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Individuals}
\alias{Individuals}
\title{Some title}
\usage{
Individuals(
  StoxBioticData,
  DensityType = c("Acoustic", "SweptArea"),
  BioticAssignment,
  LengthDistributionData
)
}
\arguments{
\item{parameterName}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
