% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{LengthDependentLengthDistributionCompensation}
\alias{LengthDependentLengthDistributionCompensation}
\title{Catchability of trawls by fish length}
\usage{
LengthDependentLengthDistributionCompensation(
  LengthDistributionData,
  CompensationMethod = c("LengthDependentSweepWidth", "LengthDependentSelectivity"),
  LengthDependentSweepWidthParameters = data.table::data.table(),
  LengthDependentSelectivityParameters = data.table::data.table()
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{CompensationMethod}{The method to use for the length dependent catch compensation, one of "LengthDependentSweepWidth" for adjusting the sweep width according to the fish length dependent herding effect and "LengthDependentSelectivity" for compensating for mash size selectivity (not yet implemented).}

\item{LengthDependentSweepWidthParameters}{A table of parameters of the LengthDependentSweepWidth method, containing the columns SpeciesCategory, LMin, LMax, Alpha and Beta (see details).}

\item{LengthDependentSelectivityParameters}{A table of parameters of the LengthDependentSelectivity method, containing the columns SpeciesCategory, LMax, Alpha and Beta (see details). Currently not supported.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function compensates for length dependent herding by the trawl doors into the net, or length dependent selectivity by the mesh size.
}
\details{
When \code{CompensationMethod} = "LengthDependentSweepWidth" the length frequencies (WeightedNumber in the LengthDistributionData) are divided by the effective sweep width estimated by the following expression

\deqn{Alpha * IndividualTotalLength^{Beta}}

In this expression \code{IndividualTotalLength} is truncated to the range \eqn{[LMin, LMin]}.
}
