% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{LengthDependentLengthDistributionCompensation}
\alias{LengthDependentLengthDistributionCompensation}
\title{Catchability of trawls by fish length}
\usage{
LengthDependentLengthDistributionCompensation(
  LengthDistributionData,
  CompensationMethod = c("LengthDependentSweepWidth", "LengthDependentSelectivity"),
  LengthDependentSweepWidthParameters = data.table::data.table(),
  LengthDependentSelectivityParameters = data.table::data.table()
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{CompensationMethod}{The method to use for the length dependent catch compensation, one of "LengthDependentSweepWidth" for adjusting the sweep width according to the fish length dependent herding effect quantified through the function alpha * length ^ beta; and "LengthDependentSelectivity" for compensating for mash size selectivity through the net using the function Alpha * e ^ (length * Beta).}

\item{LengthDependentSweepWidthParameters}{A table of parameters of the LengthDependentSweepWidth method, containing the columns SpeciesCategory, LMin, LMax, Alpha and Beta (see details).}

\item{LengthDependentSelectivityParameters}{A table of parameters of the LengthDependentSelectivity method, containing the columns SpeciesCategory, LMax, Alpha and Beta (see details).}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function compensates for length dependent herding by the trawl doors into the net, or length dependent selectivity by the mesh size.
}
