% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{LengthDistribution}
\alias{LengthDistribution}
\title{Length distribution}
\usage{
LengthDistribution(
  StoxBioticData,
  LengthDistributionType = c("Normalized", "Standard", "Percent"),
  RaisingFactorPriority = c("Weight", "Count")
)
}
\arguments{
\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{LengthDistributionType}{The type of length distribution to use, one of "LengthDist", "NormLengthDist" and "PercentLengthDist" (see 'Details').}

\item{RaisingFactorPriority}{A character string naming the variable to prioritise when generating raising factors for summing length distributions from different (sub)samples of one SpeciesCategory of a Haul, one of "Weight" and "Count".}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function calculates length frequency distribution per Stratum, biotic PSU, biotic layer, SpeciesCategory and length group defined by the combination of IndividualTotalLength and LengthResolution.
}
\details{
*********
}
\examples{

}
\seealso{

}
