% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{LengthDistribution}
\alias{LengthDistribution}
\title{Length distribution}
\usage{
LengthDistribution(
  StoxBioticData,
  SweptAreaPSU = NULL,
  SweptAreaLayer = NULL,
  LengthDistributionType = c("Normalized", "Standard", "Percent"),
  RaisingFactorPriority = c("Weight", "Count"),
  acceptNA = TRUE
)
}
\arguments{
\item{StoxBioticData}{A list of \code{\link[RstoxData]{StoxBioticData}} data.}

\item{SweptAreaPSU}{A list of \code{\link{SweptAreaPSU}} data.}

\item{SweptAreaLayer}{A list of \code{\link{SweptAreaLayer}} data.}

\item{RaisingFactorPriority}{A character string naming the variable to prioritise when generating raising factors for summing length distributions from different (sub)samples of one SpeciesCategory of a Haul, one of "Weight" and "Count".}

\item{LengthDistType}{The type of length distribution to use, one of "LengthDist", "NormLengthDist" and "PercentLengthDist" (see 'Details').}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function calculates length frequency distribution per Stratum, swept-area PSU, swept-area layer, SpeciesCategory and length group defined by the combination of IndividualTotalLengthCentimeter and LengthResolutionCentimeter.
}
\details{
*********
}
\examples{

}
\seealso{

}
