% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{LengthDistributionData}
\alias{LengthDistributionData}
\title{Length distribution data}
\description{
The LengthDistributionData model data gives the frequency per Haul, SpeciesCategory and length group defined by the two variables IndividualTotalLengthCentimeter and LengthResolutionCentimeter.
}
\seealso{
\code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
