% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{MeanDensity}
\alias{MeanDensity}
\title{Mean density in each stratum}
\usage{
MeanDensity(DensityData)
}
\arguments{
\item{DensityData}{The \code{\link{DensityData}} data.}
}
\description{
This function calculates the weighted average of the density in each Stratum/Layer/SpeciesCategory/Beam. The weights are summed effective log distance for each acoustic PSU for acoustic density, and number of stations per biotic PSU for swept area density.
}
\seealso{
See \code{\link{AcousticDensity}} for acoustic density and  \code{\link{SweptAreaDensity}} for swept-area density.
}
