% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{MeanDensity}
\alias{MeanDensity}
\title{Mean density in each stratum}
\usage{
MeanDensity(DensityData)
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function calculates the weighted avverage of the density in each stratum (or possibly in each PSU, which impies returning the input DensityData unchanged). The weights are effective log distance for acoustic density and number of hauls ??????????????????? per PSU for swept area density.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
See \code{\link{AcousticDensity}} for acoustic density and  \code{\link{SweptAreaDensity}} for swept-area density.
}
