% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{MeanLengthDistribution}
\alias{MeanLengthDistribution}
\title{(Weighted) Average length distribution horizontally}
\usage{
MeanLengthDistribution(LengthDistributionData, TargetResolution = "PSU")
}
\arguments{
\item{LengthDistributionData}{The length distribution data.}

\item{TargetResolution}{The horizontal resolution of the output.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function calculates average length distribution, weighted by the TowedDistance for the case that LengthDistributionType = "NormalizedLengthDistribution".
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
