% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{MeanLengthDistribution}
\alias{MeanLengthDistribution}
\title{Mean length distribution}
\usage{
MeanLengthDistribution(
  LengthDistributionData,
  SumLengthDistributionData,
  LayerDefinition = c("FunctionParameter", "FunctionInput", "PreDefined"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution",
    "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  SweptAreaLayer = NULL,
  PSUDefinition = c("FunctionParameter", "FunctionInput"),
  PSUDefinitionMethod = c("StationToPSU", "None"),
  StratumPolygon = NULL,
  SweptAreaPSU = NULL
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{SumLengthDistributionData}{The \code{\link{SumLengthDistributionData}} data.}

\item{SweptAreaLayer}{The \code{\link{SweptAreaLayer}} process data.}

\item{PSUDefinition}{A string naming the method to use for defining the Layers, one of "FunctionParameter", requiring \code{LayerDefinitionMethod} or \code{PSUDefinitionMethod} and the conditionally \code{Resolution}, \code{LayerTableLayerTable} or \code{StratumPolygon} to be set, or "FunctionInput", requiring the inputs \code{SweptAreaLayer} or \code{SweptAreaPSU}.}

\item{PSUDefinitionMethod}{See \code{DefinitionMethod} of \code{\link{DefineSweptAreaPSU}}.}

\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{SweptAreaPSU}{The \code{\link{SweptAreaPSU}} process data.}
}
\value{
An \code{\link{MeanLengthDistributionData}} object.
}
\description{
This function averages LengthDistributionData data horizontally, weighted by the effective towed distance.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
