% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{MeanLengthDistribution}
\alias{MeanLengthDistribution}
\title{(Weighted) Average length distribution horizontally}
\usage{
MeanLengthDistribution(
  LengthDistributionData,
  PSUDefinition = c("PreDefined", "FunctionInput"),
  SweptAreaPSU = NULL,
  TargetResolution = "PSU"
)
}
\arguments{
\item{LengthDistributionData}{The length distribution data.}

\item{PSUDefinition}{A string naming the method to use for defining the PSUs, one of "PreDefined", if the PSU column is already populated in the \code{LengthDistributionData}, or "FunctionInput" to provide the PSUs in the input \code{SweptAreaPSU}}

\item{SweptAreaPSU}{The  \code{\link{SweptAreaPSU}} data.}

\item{TargetResolution}{The horizontal resolution of the output.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function calculates average length distribution, weighted by the EffectiveTowedDistance for the case that LengthDistributionType = "Normalized".
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
