% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{MeanNASC}
\alias{MeanNASC}
\title{(Weighted) Average NASC}
\usage{
MeanNASC(NASCData, TargetResolution = "PSU")
}
\value{
A \code{\link{NASCData}} object.
}
\description{
This function averages NASC data horizontally, weighted by the log distance.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{NASC}} and \code{\link{MeanNASC}}.
}
