% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ModelData}
\alias{ModelData}
\title{Model data used in estimation models in StoX}
\description{
The model data of a StoX model is the data generated during the model run based on input data and user settings and resources given in the project description (project.json file). Model data are transient and only exists from a process has been run until the project is closed.
}
\details{
This \code{\link{RstoxBase}} package produces the folliwing model data:
\itemize{
\item{\code{\link{LengthDistributionData}}}
\item{\code{\link{NASCData}}}
\item{\code{\link{DensityData}}}
\item{\code{\link{StratumAreaData}}}
\item{\code{\link{AbundanceData}}}
}
}
