% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{modifyParameter_ProjectXMLList}
\alias{modifyParameter_ProjectXMLList}
\title{Modify a parameter of a process of a list representation of a project.xml file from StoX <= 2.7.}
\usage{
modifyParameter_ProjectXMLList(
  projectXMLList,
  modelName,
  processName,
  parameterName,
  newParameterValue,
  functionName = NULL
)
}
\arguments{
\item{projectXMLList}{A list as read by \code{\link{readProjectXMLToList}}.}

\item{modelName}{The name of the parameter of the process of the model to modify.}

\item{processName}{The name of the parameter of the process of the model to modify.}

\item{parameterName}{The name of the parameter of the process of the model to modify.}

\item{newParameterValue}{The new value of the parameter.}

\item{functionName}{The name of the function of the process to add.}
}
\description{
Modify a parameter of a process of a list representation of a project.xml file from StoX <= 2.7.
}
