% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{NASC}
\alias{NASC}
\title{NASC}
\usage{
NASC(StoxAcousticData = NULL, AcousticLayer = NULL, AcousticPSU = NULL)
}
\arguments{
\item{StoxAcousticData}{input in a StoxAcoustcdata format}

\item{AcousticLayer}{The \code{\link{AcousticLayer}} process data.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
NASC function converts the StoxAcousticData into
a NASCData format
}
\details{
This function is awesome and does excellent stuff.
}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
