% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{NASC}
\alias{NASC}
\title{NASC}
\usage{
NASC(StoxAcousticData = NULL)
}
\arguments{
\item{StoxAcousticData}{The \code{\link[RstoxData]{StoxAcousticData}} data (defined in RstoxData).}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
NASC function converts the StoxAcousticData into NASCData format
}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
