% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{NASC}
\alias{NASC}
\title{NASC}
\usage{
NASC(StoxAcousticData = NULL, AcousticLayer = NULL,
  AcousticPSU = NULL)
}
\arguments{
\item{StoxAcousticData}{input in a StoxAcoustcdata format}
}
\value{
A data.frame are returned
}
\description{
NASC function converts the StoxAcousticData into
a NASCData format
}
\details{
This function is awesome and does excellent stuff.
}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
