% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{number}
\alias{number}
\title{Number of TRUE}
\usage{
number(..., na.rm = FALSE)
}
\arguments{
\item{...}{One or more expressions to be pasted with collapse = " ".}

\item{na.rm}{All functions used as \code{ReportFunction} in must have an na.rm argument. This is however not used in this function.}
}
\description{
Warning! This function is used only to construct an expression to be evaluated in a data.table. Do not use this in R as a regular function like the other \code{\link{ReportFunctions}}.
}
