% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{PlotAcousticTrawlSurveyData}
\alias{PlotAcousticTrawlSurveyData}
\title{Acoustic-Trawl Survey plot}
\description{
The PlotAcousticTrawlSurveyData model data is a ggplot2 object which can be plotted at a later stage.
}
\seealso{
This data type is produced by \code{\link{PlotAcousticTrawlSurvey}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
