% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ProcessData}
\alias{ProcessData}
\title{Process data used in estimation models in StoX}
\description{
The process data of a StoX model are data that are saved to the project description (project.json file), typically manual settings (tagging of \code{\link{EDSU}}s to \code{\link{PSU}}s) or data read from resource files other than acoustic, biotic and landing input data files stored in the input folder.
}
\details{
This \code{\link{RstoxBase}} package produces the folliwing process data:
\itemize{
\item{\code{\link{StratumPolygon}}}
\item{\code{\link{AcousticLayer}}}
\item{\code{\link{SweptAreaLayer}}}
\item{\code{\link{AcousticLayer}}}
\item{\code{\link{SweptAreaLayer}}}
\item{\code{\link{AcousticPSU}}}
\item{\code{\link{SweptAreaPSU}}}
\item{\code{\link{BioticAssignment}}}
\item{\code{\link{AcousticTargetStrength}}}
}
}
