% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ProcessData}
\alias{ProcessData}
\title{Process data used by RstoxBase}
\arguments{
\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{Survey}{The \code{\link{Survey}} process data.}

\item{AcousticLayer}{The \code{\link{AcousticLayer}} process data.}

\item{AcousticPSU}{The \code{\link{AcousticPSU}} process data.}

\item{BioticLayer}{The \code{\link{BioticLayer}} process data.}

\item{BioticPSU}{The \code{\link{BioticPSU}} process data.}

\item{BioticAssignment}{The \code{\link{BioticAssignment}} process data.}

\item{AcousticTargetStrength}{The \code{\link{AcousticTargetStrength}} process data.}

\item{Regression}{The \code{\link{Regression}} process data.}
}
\description{
The process data of a StoX model are data that are saved to the project description (project.json file), typically manual settings (tagging of \code{\link{EDSU}}s to \code{\link{PSU}}s) or data read from resource files other than acoustic, biotic and landing input data files stored in the input folder.
}
\seealso{
\code{\link{ModelData}} for model data types and \code{\link{DataTypes}} for all data types produced by \code{\link{RstoxBase}}.
}
