% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{Quantity}
\alias{Quantity}
\title{Calculate abundance in the strata}
\usage{
Quantity(MeanDensityData, StratumAreaData)
}
\arguments{
\item{MeanDensityData}{The \code{\link{MeanDensityData}} data.}

\item{StratumAreaData}{The \code{\link{StratumAreaData}} data.}
}
\value{
An object of StoX data type \code{\link{QuantityData}}.
}
\description{
This function calculates abundance as the product of density and stratum area
}
\details{
The function merge the \code{\link{MeanDensityData}} with the \code{\link{StratumAreaData}} by Stratum and calculates the abundance  as the product of density (number by square nautical miles) and area (square nautical miles). 
For acoustic-trawl surveys the abundance is calculated by Stratum, Layer, Beam, SpeciesCategory and IndividualTotalLength.
In swept-area surveys the abundance is calculated by Stratum, Layer, SpeciesCategory and IndividualTotalLength
}
\seealso{
\code{\link{SuperIndividuals}} for distributing Abundance to individuals.
}
