% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{RegroupLengthDistribution}
\alias{RegroupLengthDistribution}
\title{Regroup length distribution to common intervals}
\usage{
RegroupLengthDistribution(LengthDistributionData, LengthInterval)
}
\arguments{
\item{LengthDistributionData}{The length distribution data.}

\item{LengthInterval}{The new length intervals, either given as a vector of interval breaks, or a single numeric value, in which case a vector of intervavl breaks is created covering the range of the original length interval breaks.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function aggregates the \code{WeightedCount} of the LengthDistributionData
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
