% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{RegroupLengthDistribution}
\alias{RegroupLengthDistribution}
\title{Regroup length distribution to common intervals}
\usage{
RegroupLengthDistribution(
  LengthDistributionData,
  LengthIntervalCentimeter = numeric()
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}

\item{LengthIntervalCentimeter}{Specifies the new length intervals, either given as a single numeric value representing the constant length interval widths, (starting from 0), or a vector of the interval breaks.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function aggregates the \code{WeightedCount} of the LengthDistributionData
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
