% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{replaceNAByReference}
\alias{replaceNAByReference}
\title{Replace all NAs in a data.table by reference}
\usage{
replaceNAByReference(DT, replacement = 0)
}
\arguments{
\item{DT}{A data.table.}

\item{replacement}{the object to replace by.}
}
\description{
Replace all NAs in a data.table by reference
}
