% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportDensity}
\alias{ReportDensity}
\title{Report DensityData}
\usage{
ReportDensity(
  DensityData,
  TargetVariable,
  ReportFunction = getRstoxBaseDefinitions("reportFunctions")$functionName,
  GroupingVariables = character(),
  RemoveMissingValues = FALSE,
  WeightingVariable = character()
)
}
\arguments{
\item{DensityData}{The \code{\link{DensityData}} data.}

\item{TargetVariable}{The variable to report.}

\item{ReportFunction}{The function to apply, see RstoxBase::getRstoxBaseDefinitions("reportFunctions")$functionName.}

\item{GroupingVariables}{The variables to report by, e.g. "Stratum" or "SpeciesCategory". Can be given as a comma separated list of variables (with no spaces).}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs). The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA.}

\item{WeightingVariable}{The variable to weight by. Only relevant for \code{ReportFunction} "weighted.mean".}
}
\value{
A \code{\link{ReportDensityData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{DensityData}}.
}
\examples{

}
\seealso{

}
