% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportSpeciesCategoryCatch}
\alias{ReportSpeciesCategoryCatch}
\title{Report SpeciesCategoryCatch}
\usage{
ReportSpeciesCategoryCatch(
  SpeciesCategoryCatchData,
  ReportVariable = c("TotalCatchNumber", "TotalCatchWeight"),
  ReportVariableUnit = character()
)
}
\arguments{
\item{SpeciesCategoryCatchData}{The \code{\link{SpeciesCategoryCatchData}} data.}

\item{ReportVariable}{The column to report.}

\item{ReportVariableUnit}{The unit to use for the \code{ReportVariable}. See RstoxData::StoxUnits for possible units (look for the appropriate quantity, e.g. "length" for IndividualTotalLength, and use the shortname in the \code{ReportVariableUnit}).}
}
\value{
A \code{\link{ReportSpeciesCategoryCatchData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{SpeciesCategoryCatch}}.
}
\details{
This function is useful to, e.g, sum Biomass for each SpeciesCategory and IndividualTotalLenght, or average IndividualTotalLenght for each IndiivdualAge and Stratum.
}
