% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportSpeciesCategoryCatch}
\alias{ReportSpeciesCategoryCatch}
\title{Report SpeciesCategoryCatch}
\usage{
ReportSpeciesCategoryCatch(SpeciesCategoryCatchData, StoxBioticTranslation)
}
\arguments{
\item{SpeciesCategoryCatchData}{The \code{\link{SpeciesCategoryCatchData}} data.}

\item{TargetVariable}{The variable to report.}

\item{ReportFunction}{The function to apply, one of "summaryStox", "sum", "mean", "weighted.mean", "median", "min", "max", "sd", "var", "cv", "summary", "quantile", "percentile_5_95".}

\item{GroupingVariables}{The variables to report by, e.g. "Stratum" or "SpeciesCategory".}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs). The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA.}

\item{WeightingVariable}{The variable to weight by. Only relevant for \code{ReportFunction} "weighted.mean".}
}
\value{
A \code{\link{ReportSpeciesCategoryCatchData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{SpeciesCategoryCatch}}.
}
\details{
This function is useful to, e.g, sum Biomass for each SpeciesCategory and IndividualTotalLenght, or average IndividualTotalLenght for each IndiivdualAge and Stratum.
}
\examples{

}
\seealso{

}
