% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportSpeciesCategoryCatch}
\alias{ReportSpeciesCategoryCatch}
\title{Report SpeciesCategoryCatch}
\usage{
ReportSpeciesCategoryCatch(SpeciesCategoryCatchData, Translation)
}
\arguments{
\item{SpeciesCategoryCatchData}{The \code{\link{SpeciesCategoryCatchData}} data.}

\item{Translation}{The \code{\link[RstoxData]{Translation}} process data.}
}
\value{
A \code{\link{ReportSpeciesCategoryCatchData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{SpeciesCategoryCatch}}.
}
\details{
This function is useful to, e.g, sum Biomass for each SpeciesCategory and IndividualTotalLenght, or average IndividualTotalLenght for each IndiivdualAge and Stratum.
}
