% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{ReportSuperIndividuals}
\alias{ReportSuperIndividuals}
\title{Report SuperIndividualsData}
\usage{
ReportSuperIndividuals(
  SuperIndividualsData,
  TargetVariable,
  ReportFunction = getReportFunctions(getMultiple = FALSE),
  GroupingVariables = character(),
  RemoveMissingValues = FALSE,
  WeightingVariable = character()
)
}
\arguments{
\item{SuperIndividualsData}{The \code{\link{SuperIndividualsData}} data.}

\item{TargetVariable}{The variable to report.}

\item{ReportFunction}{The function to apply, see RstoxBase::getRstoxBaseDefinitions("reportFunctions")$functionName.}

\item{GroupingVariables}{The variables to report by. For most applications \code{GroupingVariables} should include "Survey" or "SpeciesCategory", unless the user needs to sum over all Survey and SpeciesCategory.}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs). The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA. Use \code{RemoveMissingValues} = TRUE with extreme caution, as it may lead to under-estimation. E.g., if \code{RemoveMissingValues} = TRUE and a super-individual lacks \code{IndividualRoundWeight}, \code{Biomass} will be NA, and the portion of \code{Abundance} distributed to that super-individual will be excluded when summing \code{Biomass} (but included when summing \code{Abundance}). It is advised to always run with \code{RemoveMissingValues} = FALSE first, and make a thorough investigation to identify the source of any missing values. The function \code{link{ImputeSuperIndividuals}} can be used to impute the missing information from other super-individuals.}

\item{WeightingVariable}{The variable to weight by. Only relevant for \code{ReportFunction} "weighted.mean".}
}
\value{
A \code{\link{ReportSuperIndividualsData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{SuperIndividualsData}}.
}
\details{
This function is useful to, e.g, sum Biomass for each SpeciesCategory and IndividualTotalLength, or average IndividualTotalLength for each IndiivdualAge and Stratum.
}
