% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RstoxBase}
\alias{RstoxBase}
\alias{RstoxBase-package}
\title{Base StoX functions}
\description{
This package contains the base StoX functions used for survey estimation.
}
\details{
The StoX functions defined in RstoxBase are those for defining resolution (e.g., PSUs and Layers), assignment, NASC data and StationLengthDistribution data, density, abundance and superindividual abundance
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxBase}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxBase/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

}
