% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{sampleSorted}
\alias{sampleSorted}
\title{Function to sample after sorting:}
\usage{
sampleSorted(x, size, seed, replace = TRUE, sorted = TRUE, index.out = FALSE)
}
\arguments{
\item{x}{The vector to sample from.}

\item{size}{The length of the sampled vector.}

\item{seed}{The seed to use for the sampling.}

\item{replace}{Logical: If TRUE sample with replacement.}

\item{sorted}{Sort the vector before sampling.}

\item{index.out}{Return indices at which to sample rather than the actual samples.}
}
\description{
Function to sample after sorting:
}
