% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setBaseUnit}
\alias{setBaseUnit}
\title{Set base unit to a variable of the StoX dataType}
\usage{
setBaseUnit(x, dataType, variableName, variableType = NULL)
}
\arguments{
\item{x}{The object to set base units to.}

\item{dataType}{The name of a StoX \code{\link{DataTypes}}.}

\item{variableName}{The variable name to get base units from.}

\item{variableType}{For variables with more than one interpretation, such as Density, which is dependent on the DensityType. The type is given in the getRstoxBaseDefinitions("dataTypeUnits").}
}
\description{
Set base unit to a variable of the StoX dataType
}
