% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{setDefaultsInStoxFunction}
\alias{setDefaultsInStoxFunction}
\title{Set default general options}
\usage{
setDefaultsInStoxFunction(x, StoxFunctionName, stoxFunctionAttributes)
}
\arguments{
\item{x}{A list of arguments to the StoxFunction.}

\item{StoxFunctionName}{The name of the StoxFunction.}

\item{stoxFunctionAttributes}{The stoxFunctionAttributes to use.}
}
\description{
Set default general options
}
