% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCategoryCatch.R
\name{SpeciesCategoryCatch}
\alias{SpeciesCategoryCatch}
\title{Table of total catch weight and number per Haul and SpeciesCategory}
\usage{
SpeciesCategoryCatch(
  StoxBioticData,
  SpeciesCategoryCatchType = c("Normalized", "Standard")
)
}
\arguments{
\item{StoxBioticData}{The \code{\link[RstoxData]{StoxBioticData}} data (defined in RstoxData).}

\item{SpeciesCategoryCatchType}{The type of the total catch variables, either "Standard" for the raw total catch and "Normalized" for normalizing by tow distance by dividing by the EffectiveTowDistance in nautical miles.}
}
\description{
This function sums the catch fraction weight and number per Haul and SpeciesCategory.
}
