% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SpeciesCategoryCatchData}
\alias{SpeciesCategoryCatchData}
\title{Total catch of each SpeciesCategory}
\description{
The SpeciesCategoryDensityData model data is a list of two elements; the HaulInfo, containing information about each Haul; and the SpeciesCategoryCatch, holding the total catch of each SpeciesCategory of each Haul, where the Hauls are organized as rows and SpeciesCategories as columns.
}
\seealso{
This data type is produced by \code{\link{SpeciesCategoryCatch}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
