% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SpeciesCategoryCatchData}
\alias{SpeciesCategoryCatchData}
\title{Total catch per SpeciesCategory and Haul}
\description{
The SpeciesCategoryCatchData model data is a table with one row per Haul and SpeciesCategory holding the TotalCatchWeigth and TotalCatchNumber which are summed CatchFractionWeight and CatchFractionNumber over all samples of the Haul for each SpeciesCategory. All variables of the input StoxBioticData are kept except keys (variables with name ending with "Key").
}
\seealso{
This data type is produced by \code{\link{SpeciesCategoryCatch}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
