% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{StoxDataStartMiddleStopDateTime}
\alias{StoxDataStartMiddleStopDateTime}
\title{Generate Start, Middle and Stop DateTime variables}
\usage{
StoxDataStartMiddleStopDateTime(
  StoxDataStationLevel,
  type = c("Acoustic", "Biotic")
)
}
\arguments{
\item{type}{A string naming the type of StoX data, one of "Acoustic" and "Biotic".}

\item{StoxData}{Either \code{\link{StoxAcousticData}} or \code{\link{StoxBioticData}}, depending on \code{type}.}
}
\value{
An object of StoX data type \code{\link{MergeStoxAcousticData}}.
}
\description{
Generate Start, Middle and Stop DateTime variables
}
