% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelData.R, R/Spatial.R
\name{StratumArea}
\alias{StratumArea}
\title{Some title}
\usage{
StratumArea(StratumPolygon, AreaMethod = c("Accurate", "Simple"))

StratumArea(StratumPolygon, AreaMethod = c("Accurate", "Simple"))
}
\arguments{
\item{StratumPolygon}{An object of StoX class StratumPolygon, which is s SpatialPolygons object (sp package).}

\item{AreaMethod}{The method to use for the area calculation, defaulted to "Accurate", which applied a lambert azimuthal equal area projection.}

\item{parameterName}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.

A table of stratum name and area.
}
\description{
Some description

This function calculated the area of each stratum.
}
\details{
This function is awesome and does excellent stuff.

The \code{AreaMethod} "Simple" is deprecated, but kept for backwards campatibility.
}
\examples{
x <- 1


}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.

\code{\link[RstoxBase]{DefineStrata}} for the \code{StratumPolygon} input to the function.
}
