% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{StratumArea}
\alias{StratumArea}
\title{Calculate area of each stratum}
\usage{
StratumArea(StratumPolygon, AreaMethod = c("Accurate", "Simple"))
}
\arguments{
\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{AreaMethod}{The method to use for the area calculation, defaulted to "Accurate", which applied a lambert azimuthal equal area projection.}
}
\value{
A table of stratum name and area.
}
\description{
This function calculated the area of each stratum.
}
\details{
The \code{AreaMethod} "Simple" is deprecated, but kept for backwards campatibility.
}
\examples{

}
\seealso{
\code{\link{DefineStratumPolygon}} for the \code{StratumPolygon} input to the function.
}
