% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{StratumArea}
\alias{StratumArea}
\title{Calculate area of each stratum}
\usage{
StratumArea(StratumPolygon, AreaMethod = c("Accurate", "Simple"))
}
\arguments{
\item{StratumPolygon}{The \code{\link{StratumPolygon}} process data.}

\item{AreaMethod}{The method to use for the area calculation, defaulted to "Accurate", which applied a lambert azimuthal equal area projection.}
}
\value{
A table of stratum name and area.
}
\description{
This function calculated the area of each stratum.
}
\details{
The area output is given in international square nautical miles. 

The \code{AreaMethod} "Accurate" projects the latitude and longitude startum polygons to Lambert Azimuthal Equal Area with origo in wkt center and coordinate reference system (CRS="+proj=longlat +ellps=WGS84") The areas are calcualted using \code{\link{rgeos::gArea}}. ??? hvordan setter jeg inn link til annen pakke?	

The \code{AreaMethod} "Simple" is used in StoX 2.7 and earlier versions and kept for backwards campatibility. The area of each stratum polygon is calculated as simple great circle distance polygon area (REF_to be inserted ???), where any hole in a stratum is substracted. The substraction of holes was not implemented in StoX 2.7 and earlier versions ??? Is this correct ???.
}
\examples{

}
\references{
To be inserted
}
\seealso{
\code{\link{DefineStratumPolygon}} for the \code{StratumPolygon} input to the function.
\code{\link{rgeos::gArea}}
\code{\link{sp::CRS}}
}
