% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{StratumAreaData}
\alias{StratumAreaData}
\title{Stratum area data}
\description{
The StratumAreaData model data is a table of the area in square nautical miles of each stratum.
}
\seealso{
This data type is produced by \code{\link{StratumArea}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}.
}
