% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{StratumPolygon}
\alias{StratumPolygon}
\title{StratumPolygon}
\description{
datatype.... \code{\link[sp]{SpatialPolygons}}
}
\details{
Bla bla.
}
\seealso{
\code{\link[RstoxData]{DataTypes}} for a list of all StoX data types.
}
