% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthDistribution.R
\name{SumLengthDistribution}
\alias{SumLengthDistribution}
\title{Sum length distribution vertically}
\usage{
SumLengthDistribution(LengthDistributionData, TargetResolution = "Layer")
}
\arguments{
\item{LengthDistributionData}{The length distribution data.}

\item{TargetResolution}{The vertical resolution of the output.}
}
\value{
A \code{\link{LengthDistributionData}} object.
}
\description{
This function sums length distribution to swept area layers.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
