% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{summaryStox}
\alias{summaryStox}
\title{The summary function introduced in StoX <= 2.7.}
\usage{
summaryStox(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric object}

\item{na.rm}{Logical: If TRUE remove the missing values prior to calculation.}
}
\description{
The summary function introduced in StoX <= 2.7.
}
