% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SumNASC}
\alias{SumNASC}
\title{Sum NASC}
\usage{
SumNASC(
  NASCData,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution",
    "LayerTable"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL
)
}
\value{
A \code{\link{SumNASCData}} object.
}
\description{
This function summes NASC data vertically.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{NASC}} and \code{\link{MeanNASC}}.
}
