% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SumNASC}
\alias{SumNASC}
\title{Sum NASC}
\usage{
SumNASC(NASCData, TargetResolution = "Layer")
}
\arguments{
\item{NASCData}{The NASC data.}

\item{TargetResolution}{The vertical resolution of the output.}
}
\value{
A \code{\link{NASCData}} object.
}
\description{
This function sums NASC data vertically into layer resolution
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link{NASC}} and \code{\link{SumNASC}}.
}
