% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acoustic.R
\name{SumNASC}
\alias{SumNASC}
\title{Sum NASC}
\usage{
SumNASC(
  NASCData,
  LayerDefinition = c("FunctionParameter", "FunctionInput"),
  LayerDefinitionMethod = c("WaterColumn", "HighestResolution", "Resolution", "Table"),
  Resolution = double(),
  LayerTable = data.table::data.table(),
  AcousticLayer = NULL
)
}
\arguments{
\item{NASCData}{The \code{\link{NASCData}} data.}

\item{LayerDefinition}{The method to use for defining the Layers, one of \code{FunctionParameter} to define the Layers on the fly in this function, or \code{FunctionInput} to import Layer process data from a previously run process by \code{AcousticLayer}.}

\item{LayerDefinitionMethod}{See \code{\link{DefineAcousticLayer}}}

\item{Resolution}{Numeric: A single numeric giving the thickness of the layers.}

\item{LayerTable}{A table of Layer name, MinLayerDepth in meters and MaxLayerDepth in meters, defining the Layers.}

\item{AcousticLayer}{The \code{\link{AcousticLayer}} process data.}
}
\value{
A \code{\link{SumNASCData}} object.
}
\description{
This function summes NASC data vertically.
}
\seealso{
\code{\link{NASC}} and \code{\link{MeanNASC}}.
}
