% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SumNASCData}
\alias{SumNASCData}
\title{Summed NASC data}
\description{
The SumNASCData model data is a list of two elements; the Data, which are similar to \code{\link{NASCData}} but added the Layer resolution column and summed vertically; and the Resolution, which is a table of the three columns EDSU, Layer and Channel containing the link between the vertical resolution variables before summing.
}
\seealso{
This data type is produced by \code{\link{SumNASC}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
