% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{SuperIndividuals}
\alias{SuperIndividuals}
\title{Some title}
\usage{
SuperIndividuals(
  IndividualsData,
  AbundanceData,
  AbundWeightMethod = c("Equal", "HaulDensity"),
  LengthDistributionData
)
}
\arguments{
\item{IndividualsData}{The \code{\link{IndividualsData}} data.}

\item{AbundanceData}{The \code{\link{AbundanceData}} data.}

\item{AbundWeightMethod}{The method used for distributing the abundance, one of "Equal" for equal abundance to all individuals of each Stratum, Layer, SpeciesCategory and length group, and "HaulDensity" to weight by the haul density.}

\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
