% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{SuperIndividuals}
\alias{SuperIndividuals}
\title{Some title}
\usage{
SuperIndividuals(
  IndividualsData,
  AbundanceData,
  AbundWeightMethod = c("Equal", "HaulDensity"),
  LengthDistributionData
)
}
\arguments{
\item{IndividualsData}{The}

\item{AbundanceData}{Parameter descrption.}

\item{AbundWeightMethod}{Parameter descrption.}

\item{LengthDistributionData}{The length distribution data.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
