% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\name{SuperIndividuals}
\alias{SuperIndividuals}
\title{Super-indivduals with abundance}
\usage{
SuperIndividuals(
  IndividualsData,
  QuantityData,
  DistributionMethod = c("Equal", "HaulDensity"),
  LengthDistributionData
)
}
\arguments{
\item{IndividualsData}{The \code{\link{IndividualsData}} data.}

\item{QuantityData}{The \code{\link{QuantityData}} data.}

\item{DistributionMethod}{The method used for distributing the abundance, one of "Equal" for equal abundance to all individuals of each Stratum, Layer, SpeciesCategory and length group, and "HaulDensity" to weight by the haul density. For \code{DistributionMethod} = "HaulDensity" the \code{LengthDistributionData} must be given. It is recommended to use the same \code{LengthDistributionData} that was used to produce the \code{\link{QuantityData}} (via \code{link{DensityData}}).}

\item{LengthDistributionData}{The \code{\link{LengthDistributionData}} data.}
}
\description{
This function distributes Abundance to the individuals defined by \code{\link{Individuals}}.
}
\details{
The \code{\link{SuperIndividualsData}} contains the variables \code{Abundance} and \code{Biomass}. The \code{Biomass} is given in gram, as it is generated from IndividualRoundWeight which is in gram. This is different from the \code{Biomass} column of \code{\link{QuantityData}}, which is in kilogram, as it is originates from \code{CatchFractionWeight} in \code{\link{StoxBioticData}}, which is in kilogram.
}
\seealso{
\code{\link{Individuals}}, \code{\link{Quantity}} and \code{\link{LengthDistribution}} for generating the input to this function.
}
