% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{SuperIndividualsData}
\alias{SuperIndividualsData}
\title{Super-individuals data}
\description{
The SuperIndividualsData model data is the \code{\link{IndivdualsData}} added Abundance and Biomass.
}
\seealso{
This data type is produced by \code{\link{SuperIndividuals}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxBase}}
}
