% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{SweptAreaDensity}
\alias{SweptAreaDensity}
\title{Swept-area density}
\usage{
SweptAreaDensity(
  MeanLengthDistributionData,
  SweepWidthMethod = c("Constant", "PreDefined", "CruiseDependent"),
  SweepWidth = double(),
  SweepWidthTable = data.table::data.table()
)
}
\arguments{
\item{MeanLengthDistributionData}{The \code{\link{MeanLengthDistributionData}} data.}

\item{SweepWidthMethod}{The method for calculating the sweep width to multiply the \code{WeightedCount} in the \code{LengthDistributionData} with. Possible options are (1) "Constant", which requires \code{SweepWidth} to be set as the constant sweep width, (2) "PreDefined", impying that the sweep width is already incorporated in the \code{WeightedCount} in the \code{LengthDistributionData}, by using LengthDistributionType == "SweepWidthCompensatedNormalized" in the function \code{\link{LengthDependentCatchCompensation}}, and (3) "CruiseDependent", which requires the \code{SweepWidthTable} to be given.}

\item{SweepWidth}{The constant sweep width in meters.}

\item{SweepWidthTable}{A table of two columns, \code{Cruise} and \code{SweepWidth}, giving the sweep width for each cruise.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function calculates the area density of fish as number of individuals per square nautical mile.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
See \code{\link{AcousticDensity}} for acoustic density.
}
