% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{SweptAreaDensity}
\alias{SweptAreaDensity}
\title{Swept-area density}
\usage{
SweptAreaDensity(
  LengthDistributionData,
  SweptAreaMethod = c("LengthDependent", "TotalCatch"),
  SweepWidthMethod = c("Constant", "CruiseDependent"),
  SweepWidth = integer(),
  SweepWidthExpr = character(),
  CatchVariable = c("Weight", "Count")
)
}
\arguments{
\item{LengthDistributionData}{The length distribution data.}

\item{SweptAreaMethod}{The method used to calculate the density, one of "LengthDependent" and "TotalCatch" (the latter is not )}

\item{SweepWidthMethod}{Parameter descrption.}

\item{SweepWidthExpr}{Parameter descrption.}

\item{SweepWidthSweepWidth}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function calculates the area density of fish as number of individuals per square nautical mile.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
